CmdStanDiagnose objects

A CmdStanDiagnose object is the object returned by the $diagnose() method of a CmdStanModel object.

Methods

CmdStanDiagnose objects have the following associated methods:

Method	Description

$gradients()	Return gradients from diagnostic mode.

$lp()	Return the total log probability density (target).

$init()	Return user-specified initial values.

$metadata()	Return a list of metadata gathered from the CmdStan CSV files.

$save_output_files()	Save output CSV files to a specified location.

$save_data_file()	Save JSON data file to a specified location.

See also

Examples

# \dontrun{
test <- cmdstanr_example("logistic", method = "diagnose")

# retrieve the gradients
test$gradients()
#>   param_idx     value     model finite_diff        error
#> 1         0  1.710110 -18.30620   -18.30620 -1.81553e-08
#> 2         1 -0.468770  -5.97695    -5.97695  1.10331e-08
#> 3         2 -1.800050  20.42290    20.42290  1.19712e-08
#> 4         3 -0.508314  16.91120    16.91120  3.84095e-10
# }