CmdStanMCMC objects

A CmdStanMCMC object is the fitted model object returned by the $sample() method of a CmdStanModel object. Like CmdStanModel objects, CmdStanMCMC objects are R6 objects.

Methods

CmdStanMCMC objects have the following associated methods, all of which have their own (linked) documentation pages.

Extract contents of fitted model object

Method	Description

$draws()	Return posterior draws using formats from the posterior package.

$sampler_diagnostics()	Return sampler diagnostics as a draws_array.

$lp()	Return the total log probability density (target).

$inv_metric()	Return the inverse metric for each chain.

$init()	Return user-specified initial values.

$metadata()	Return a list of metadata gathered from the CmdStan CSV files.

$num_chains()	Return the number of MCMC chains.

$code()	Return Stan code as a character vector.

Summarize inferences and diagnostics

Method	Description

$print()	Run posterior::summarise_draws().

$summary()	Run posterior::summarise_draws().

$diagnostic_summary()	Get summaries of sampler diagnostics and warning messages.

$cmdstan_summary()	Run and print CmdStan's bin/stansummary.

$cmdstan_diagnose()	Run and print CmdStan's bin/diagnose.

$loo()	Run loo::loo.array() for approximate LOO-CV

Save fitted model object and temporary files

Method	Description

$save_object()	Save fitted model object to a file.

$save_output_files()	Save output CSV files to a specified location.

$save_data_file()	Save JSON data file to a specified location.

$save_latent_dynamics_files()	Save diagnostic CSV files to a specified location.

Report run times, console output, return codes

Method	Description

$output()	Return the stdout and stderr of all chains or pretty print the output for a single chain.

$time()	Report total and chain-specific run times.

$return_codes()	Return the return codes from the CmdStan runs.

Expose Stan functions and additional methods to R

Method	Description

$expose_functions()	Expose Stan functions for use in R.

$init_model_methods()	Expose methods for log-probability, gradients, parameter constraining and unconstraining.

$log_prob()	Calculate log-prob.

$grad_log_prob()	Calculate log-prob and gradient.

$hessian()	Calculate log-prob, gradient, and hessian.

$constrain_variables()	Transform a set of unconstrained parameter values to the constrained scale.

$unconstrain_variables()	Transform a set of parameter values to the unconstrained scale.

$unconstrain_draws()	Transform all parameter draws to the unconstrained scale.

$variable_skeleton()	Helper function to re-structure a vector of constrained parameter values.

See also