Create a draws object from a CmdStanR fitted model object

Create a draws object supported by the posterior package. These methods are just wrappers around CmdStanR's $draws() method provided for convenience.

# S3 method for CmdStanMCMC
as_draws(x, ...)

# S3 method for CmdStanMLE
as_draws(x, ...)

# S3 method for CmdStanVB
as_draws(x, ...)

# S3 method for CmdStanGQ
as_draws(x, ...)

Arguments

x
A CmdStanR fitted model object.

...
Optional arguments passed to the $draws() method (e.g., variables, inc_warmup, etc.).

Details

To subset iterations, chains, or draws, use the posterior::subset_draws() method after creating the draws object.

Examples

# \dontrun{
fit <- cmdstanr_example()
as_draws(fit)

#> # A draws_array: 1000 iterations, 4 chains, and 105 variables
#> , , variable = lp__
#> 
#>          chain
#> iteration   1   2   3   4
#>         1 -68 -69 -66 -65
#>         2 -66 -66 -70 -65
#>         3 -66 -65 -67 -64
#>         4 -65 -67 -66 -65
#>         5 -65 -68 -67 -66
#> 
#> , , variable = alpha
#> 
#>          chain
#> iteration    1    2     3    4
#>         1 0.59 0.65 0.084 0.51
#>         2 0.28 0.27 0.653 0.33
#>         3 0.49 0.29 0.731 0.37
#>         4 0.31 0.18 0.506 0.33
#>         5 0.41 0.75 0.225 0.45
#> 
#> , , variable = beta[1]
#> 
#>          chain
#> iteration      1     2     3     4
#>         1 -0.047 -0.92 -0.71 -0.57
#>         2 -1.181 -0.68 -0.33 -0.72
#>         3 -0.310 -0.91 -0.56 -0.49
#>         4 -0.513 -1.03 -0.41 -0.88
#>         5 -0.953 -1.10 -0.82 -0.48
#> 
#> , , variable = beta[2]
#> 
#>          chain
#> iteration      1     2     3       4
#>         1 -0.360  0.14 -0.48 -0.4900
#>         2 -0.269 -0.39  0.25 -0.0093
#>         3 -0.082 -0.24 -0.49 -0.4064
#>         4  0.032 -0.67 -0.69 -0.0911
#>         5 -0.445 -0.47  0.20  0.0281
#> 
#> # ... with 995 more iterations, and 101 more variables

# posterior's as_draws_*() methods will also work
posterior::as_draws_rvars(fit)

#> # A draws_rvars: 1000 iterations, 4 chains, and 4 variables
#> $lp__: rvar<1000,4>[1] mean ± sd:
#> [1] -66 ± 1.5 
#> 
#> $alpha: rvar<1000,4>[1] mean ± sd:
#> [1] 0.38 ± 0.22 
#> 
#> $beta: rvar<1000,4>[3] mean ± sd:
#> [1] -0.67 ± 0.26  -0.28 ± 0.23   0.69 ± 0.27 
#> 
#> $log_lik: rvar<1000,4>[100] mean ± sd:
#>   [1] -0.516 ± 0.100  -0.400 ± 0.148  -0.498 ± 0.219  -0.447 ± 0.152 
#>   [5] -1.190 ± 0.284  -0.588 ± 0.197  -0.641 ± 0.127  -0.276 ± 0.134 
#>   [9] -0.691 ± 0.173  -0.741 ± 0.236  -0.277 ± 0.127  -0.492 ± 0.246 
#>  [13] -0.662 ± 0.217  -0.362 ± 0.178  -0.276 ± 0.105  -0.273 ± 0.087 
#>  [17] -1.600 ± 0.296  -0.478 ± 0.109  -0.231 ± 0.077  -0.112 ± 0.079 
#>  [21] -0.209 ± 0.088  -0.566 ± 0.152  -0.328 ± 0.139  -0.135 ± 0.067 
#>  [25] -0.454 ± 0.122  -1.524 ± 0.355  -0.303 ± 0.123  -0.444 ± 0.084 
#>  [29] -0.725 ± 0.226  -0.694 ± 0.193  -0.485 ± 0.167  -0.421 ± 0.108 
#>  [33] -0.409 ± 0.128  -0.063 ± 0.051  -0.588 ± 0.187  -0.324 ± 0.132 
#>  [37] -0.698 ± 0.226  -0.310 ± 0.147  -0.178 ± 0.109  -0.680 ± 0.130 
#>  [41] -1.139 ± 0.263  -0.932 ± 0.201  -0.417 ± 0.272  -1.179 ± 0.189 
#>  [45] -0.359 ± 0.121  -0.581 ± 0.130  -0.301 ± 0.127  -0.323 ± 0.084 
#>  [49] -0.317 ± 0.080  -1.291 ± 0.346  -0.288 ± 0.097  -0.834 ± 0.146 
#>  [53] -0.399 ± 0.132  -0.373 ± 0.148  -0.383 ± 0.136  -0.319 ± 0.190 
#>  [57] -0.657 ± 0.118  -0.948 ± 0.355  -1.366 ± 0.348  -0.979 ± 0.163 
#>  [61] -0.541 ± 0.100  -0.873 ± 0.328  -0.114 ± 0.072  -0.903 ± 0.250 
#>  [65] -2.039 ± 0.612  -0.505 ± 0.135  -0.275 ± 0.083  -1.065 ± 0.240 
#>  [69] -0.434 ± 0.084  -0.639 ± 0.235  -0.615 ± 0.215  -0.465 ± 0.178 
#>  [73] -1.492 ± 0.373  -0.951 ± 0.195  -1.141 ± 0.408  -0.373 ± 0.142 
#>  [77] -0.881 ± 0.141  -0.492 ± 0.180  -0.764 ± 0.191  -0.539 ± 0.198 
#>  [81] -0.162 ± 0.098  -0.220 ± 0.138  -0.342 ± 0.082  -0.273 ± 0.090 
#>  [85] -0.129 ± 0.075  -1.133 ± 0.320  -0.824 ± 0.129  -0.784 ± 0.252 
#>  [89] -1.285 ± 0.321  -0.260 ± 0.134  -0.386 ± 0.132  -1.509 ± 0.346 
#>  [93] -0.738 ± 0.221  -0.316 ± 0.088  -0.386 ± 0.113  -1.581 ± 0.288 
#>  [97] -0.430 ± 0.102  -1.051 ± 0.377  -0.695 ± 0.144  -0.391 ± 0.100 
#> 

posterior::as_draws_list(fit)

#> # A draws_list: 1000 iterations, 4 chains, and 105 variables
#> 
#> [chain = 1]
#> $lp__
#>  [1] -68 -66 -66 -65 -65 -64 -65 -65 -68 -65
#> 
#> $alpha
#>  [1] 0.59 0.28 0.49 0.31 0.41 0.49 0.40 0.15 0.33 0.56
#> 
#> $`beta[1]`
#>  [1] -0.047 -1.181 -0.310 -0.513 -0.953 -0.660 -0.388 -0.661 -0.329 -0.615
#> 
#> $`beta[2]`
#>  [1] -0.360 -0.269 -0.082  0.032 -0.445 -0.227 -0.190 -0.549 -0.689 -0.528
#> 
#> 
#> [chain = 2]
#> $lp__
#>  [1] -69 -66 -65 -67 -68 -68 -65 -66 -67 -66
#> 
#> $alpha
#>  [1] 0.65 0.27 0.29 0.18 0.75 0.93 0.46 0.48 0.60 0.39
#> 
#> $`beta[1]`
#>  [1] -0.92 -0.68 -0.91 -1.03 -1.10 -0.71 -0.83 -0.84 -1.00 -0.16
#> 
#> $`beta[2]`
#>  [1]  0.138 -0.393 -0.245 -0.666 -0.469 -0.596  0.027 -0.707 -0.725 -0.209
#> 
#> # ... with 990 more iterations, and 2 more chains, and 101 more variables

# }