Fit models for use in examples

Fit models for use in examples

cmdstanr_example(
  example = c("logistic", "schools", "schools_ncp"),
  method = c("sample", "optimize", "variational", "diagnose"),
  ...,
  quiet = TRUE
)

print_example_program(example = c("logistic", "schools", "schools_ncp"))

Arguments

example
(string) The name of the example. The currently available examples are

"logistic": logistic regression with intercept and 3 predictors.

"schools": the so-called "eight schools" model, a hierarchical meta-analysis. Fitting this model will result in warnings about divergences.

"schools_ncp": non-centered parameterization of the "eight schools" model that fixes the problem with divergences.

To print the Stan code for a given example use print_example_program(example).

method
(string) Which fitting method should be used? The default is the "sample" method (MCMC).

...
Arguments passed to the chosen method. See the help pages for the individual methods for details.

quiet
(logical) If TRUE (the default) then fitting the model is wrapped in utils::capture.output().

Value

The fitted model object returned by the selected method.

Examples

# \dontrun{
print_example_program("logistic")

#> data {
#>   int<lower=0> N;
#>   int<lower=0> K;
#>   array[N] int<lower=0, upper=1> y;
#>   matrix[N, K] X;
#> }
#> parameters {
#>   real alpha;
#>   vector[K] beta;
#> }
#> model {
#>   target += normal_lpdf(alpha | 0, 1);
#>   target += normal_lpdf(beta | 0, 1);
#>   target += bernoulli_logit_glm_lpmf(y | X, alpha, beta);
#> }
#> generated quantities {
#>   vector[N] log_lik;
#>   for (n in 1 : N) {
#>     log_lik[n] = bernoulli_logit_lpmf(y[n] | alpha + X[n] * beta);
#>   }
#> }

fit_logistic_mcmc <- cmdstanr_example("logistic", chains = 2)
fit_logistic_mcmc$summary()

#> # A tibble: 105 × 10
#>    variable      mean  median    sd    mad       q5     q95  rhat ess_bulk
#>    <chr>        <dbl>   <dbl> <dbl>  <dbl>    <dbl>   <dbl> <dbl>    <dbl>
#>  1 lp__       -66.0   -65.6   1.48  1.26   -69.0    -64.3    1.00    1079.
#>  2 alpha        0.380   0.378 0.223 0.221    0.0248   0.754  1.00    1776.
#>  3 beta[1]     -0.661  -0.665 0.251 0.247   -1.08    -0.254  1.00    2485.
#>  4 beta[2]     -0.271  -0.270 0.231 0.226   -0.658    0.103  1.00    2084.
#>  5 beta[3]      0.674   0.666 0.259 0.257    0.262    1.11   1.00    2140.
#>  6 log_lik[1]  -0.516  -0.512 0.101 0.0988  -0.693   -0.359  1.00    1984.
#>  7 log_lik[2]  -0.406  -0.390 0.147 0.142   -0.662   -0.199  1.00    2317.
#>  8 log_lik[3]  -0.500  -0.463 0.223 0.208   -0.908   -0.199  1.00    2242.
#>  9 log_lik[4]  -0.454  -0.431 0.156 0.146   -0.734   -0.233  1.01    2139.
#> 10 log_lik[5]  -1.18   -1.16  0.278 0.273   -1.67    -0.750  1.00    1999.
#> # … with 95 more rows, and 1 more variable: ess_tail <dbl>

fit_logistic_optim <- cmdstanr_example("logistic", method = "optimize")
fit_logistic_optim$summary()

#> # A tibble: 105 × 2
#>    variable   estimate
#>    <chr>         <dbl>
#>  1 lp__        -63.9  
#>  2 alpha         0.365
#>  3 beta[1]      -0.632
#>  4 beta[2]      -0.259
#>  5 beta[3]       0.648
#>  6 log_lik[1]   -0.515
#>  7 log_lik[2]   -0.394
#>  8 log_lik[3]   -0.469
#>  9 log_lik[4]   -0.442
#> 10 log_lik[5]   -1.14 
#> # … with 95 more rows

fit_logistic_vb <- cmdstanr_example("logistic", method = "variational")
fit_logistic_vb$summary()

#> # A tibble: 106 × 7
#>    variable       mean  median    sd   mad       q5      q95
#>    <chr>         <dbl>   <dbl> <dbl> <dbl>    <dbl>    <dbl>
#>  1 lp__        -65.9   -65.6   1.42  1.24  -68.7    -64.2   
#>  2 lp_approx__  -1.99   -1.68  1.40  1.24   -4.76    -0.361 
#>  3 alpha         0.402   0.398 0.234 0.243   0.0390   0.785 
#>  4 beta[1]      -0.598  -0.605 0.233 0.249  -0.971   -0.230 
#>  5 beta[2]      -0.252  -0.256 0.199 0.192  -0.585    0.0849
#>  6 beta[3]       0.779   0.773 0.211 0.215   0.450    1.13  
#>  7 log_lik[1]   -0.520  -0.514 0.104 0.101  -0.711   -0.364 
#>  8 log_lik[2]   -0.370  -0.354 0.122 0.115  -0.604   -0.202 
#>  9 log_lik[3]   -0.470  -0.439 0.198 0.184  -0.841   -0.206 
#> 10 log_lik[4]   -0.440  -0.430 0.132 0.123  -0.699   -0.251 
#> # … with 96 more rows

print_example_program("schools")

#> data {
#>   int<lower=1> J;
#>   vector<lower=0>[J] sigma;
#>   vector[J] y;
#> }
#> parameters {
#>   real mu;
#>   real<lower=0> tau;
#>   vector[J] theta;
#> }
#> model {
#>   target += normal_lpdf(tau | 0, 10);
#>   target += normal_lpdf(mu | 0, 10);
#>   target += normal_lpdf(theta | mu, tau);
#>   target += normal_lpdf(y | theta, sigma);
#> }

fit_schools_mcmc <- cmdstanr_example("schools")

#> Warning: 486 of 4000 (12.0%) transitions ended with a divergence.
#> See https://mc-stan.org/misc/warnings for details.

fit_schools_mcmc$summary()

#> # A tibble: 11 × 10
#>    variable   mean median    sd   mad       q5   q95  rhat ess_bulk ess_tail
#>    <chr>     <dbl>  <dbl> <dbl> <dbl>    <dbl> <dbl> <dbl>    <dbl>    <dbl>
#>  1 lp__     -53.1  -55.8   9.50  9.03 -65.4    -38.8  1.53     7.47    NA   
#>  2 mu         6.00   4.90  3.77  2.69   0.0567  12.9  1.53   572.     662.  
#>  3 tau        4.02   3.07  3.67  3.86   0.466   11.2  1.53     7.48     4.03
#>  4 theta[1]   7.92   5.86  6.05  3.86  -0.134   19.3  1.18   120.     751.  
#>  5 theta[2]   6.32   4.73  4.93  3.51  -1.61    15.2  1.53   437.    1036.  
#>  6 theta[3]   5.30   4.72  5.90  3.97  -4.58    15.1  1.53  1151.    1090.  
#>  7 theta[4]   6.13   4.92  5.06  3.34  -1.99    15.3  1.52   703.    1088.  
#>  8 theta[5]   4.63   5.04  5.10  3.38  -4.69    12.7  1.53  1088.    1234.  
#>  9 theta[6]   5.38   5.04  5.10  3.37  -3.60    13.6  1.53  1063.    1095.  
#> 10 theta[7]   7.92   6.05  5.48  3.76   0.492   18.5  1.14   103.     828.  
#> 11 theta[8]   6.48   4.91  5.78  3.75  -2.36    17.0  1.53   951.     980.  

print_example_program("schools_ncp")

#> data {
#>   int<lower=1> J;
#>   vector<lower=0>[J] sigma;
#>   vector[J] y;
#> }
#> parameters {
#>   real mu;
#>   real<lower=0> tau;
#>   vector[J] theta_raw;
#> }
#> transformed parameters {
#>   vector[J] theta = mu + tau * theta_raw;
#> }
#> model {
#>   target += normal_lpdf(tau | 0, 10);
#>   target += normal_lpdf(mu | 0, 10);
#>   target += normal_lpdf(theta_raw | 0, 1);
#>   target += normal_lpdf(y | theta, sigma);
#> }

fit_schools_ncp_mcmc <- cmdstanr_example("schools_ncp")

#> Warning: 1 of 4000 (0.0%) transitions ended with a divergence.
#> See https://mc-stan.org/misc/warnings for details.

fit_schools_ncp_mcmc$summary()

#> # A tibble: 19 × 10
#>    variable         mean   median    sd   mad       q5    q95  rhat ess_bulk
#>    <chr>           <dbl>    <dbl> <dbl> <dbl>    <dbl>  <dbl> <dbl>    <dbl>
#>  1 lp__         -46.9    -46.6    2.46  2.40  -51.4    -43.4   1.00    1458.
#>  2 mu             6.47     6.50   4.10  3.95   -0.399   13.2   1.00    3104.
#>  3 tau            4.82     4.07   3.72  3.62    0.354   11.9   1.00    2010.
#>  4 theta_raw[1]   0.366    0.401  0.973 0.916  -1.25     1.98  1.00    4189.
#>  5 theta_raw[2]   0.0442   0.0644 0.917 0.937  -1.47     1.53  1.00    3836.
#>  6 theta_raw[3]  -0.138   -0.151  0.963 0.930  -1.70     1.48  1.00    4136.
#>  7 theta_raw[4]   0.0208   0.0354 0.893 0.883  -1.42     1.47  1.00    3921.
#>  8 theta_raw[5]  -0.270   -0.270  0.918 0.891  -1.78     1.25  1.00    2871.
#>  9 theta_raw[6]  -0.153   -0.156  0.908 0.913  -1.63     1.34  1.00    3449.
#> 10 theta_raw[7]   0.355    0.370  0.942 0.933  -1.23     1.88  1.00    3219.
#> 11 theta_raw[8]   0.0531   0.0601 0.935 0.955  -1.50     1.55  1.00    4099.
#> 12 theta[1]       9.01     8.28   6.92  5.57   -0.715   21.0   1.00    3218.
#> 13 theta[2]       6.74     6.66   5.71  5.30   -2.28    16.3   1.00    4062.
#> 14 theta[3]       5.52     5.96   6.46  5.56   -5.68    15.0   1.00    3718.
#> 15 theta[4]       6.60     6.67   5.56  4.92   -2.48    15.4   1.00    4581.
#> 16 theta[5]       4.77     5.23   5.53  5.08   -4.99    13.0   1.00    3625.
#> 17 theta[6]       5.59     5.89   5.86  5.29   -4.57    14.4   1.00    4589.
#> 18 theta[7]       8.84     8.37   5.85  5.45    0.0924  18.9   1.00    4020.
#> 19 theta[8]       6.92     6.77   6.44  5.48   -3.30    17.8   1.00    3704.
#> # … with 1 more variable: ess_tail <dbl>

# optimization fails for hierarchical model
cmdstanr_example("schools", "optimize", quiet = FALSE)

#> Initial log joint probability = -373.952 
#>     Iter      log prob        ||dx||      ||grad||       alpha      alpha0  # evals  Notes  
#>       99       114.115      0.330131   2.59089e+09      0.3881           1      169    
#>     Iter      log prob        ||dx||      ||grad||       alpha      alpha0  # evals  Notes  
#>      179       242.979      0.047195   2.42406e+16       1e-12       0.001      373  LS failed, Hessian reset  

#> Optimization terminated with error: 

#>   Line search failed to achieve a sufficient decrease, no more progress can be made

#> Finished in  0.1 seconds.

#>  variable estimate
#>  lp__       242.98
#>  mu           1.05
#>  tau          0.00
#>  theta[1]     1.05
#>  theta[2]     1.05
#>  theta[3]     1.05
#>  theta[4]     1.05
#>  theta[5]     1.05
#>  theta[6]     1.05
#>  theta[7]     1.05
#> 
#>  # showing 10 of 11 rows (change via 'max_rows' argument or 'cmdstanr_max_rows' option)

# }