Run CmdStan's stansummary and diagnose utilities

Run CmdStan's stansummary and diagnose utilities. These are documented in the CmdStan Guide:

https://mc-stan.org/docs/cmdstan-guide/stansummary.html

https://mc-stan.org/docs/cmdstan-guide/diagnose.html

Although these methods can be used for models fit using the $variational() method, much of the output is currently only relevant for models fit using the $sample() method.

See the $summary() for computing similar summaries in R rather than calling CmdStan's utilites.

cmdstan_summary(flags = NULL)

cmdstan_diagnose()

Arguments

flags
An optional character vector of flags (e.g. flags = c("--sig_figs=1")).

See also

Examples

# \dontrun{
fit <- cmdstanr_example("logistic")
fit$cmdstan_diagnose()

#> Processing csv files: /tmp/Rtmpi2xL70/logistic-202204262035-1-703fc2.csv, /tmp/Rtmpi2xL70/logistic-202204262035-2-703fc2.csv, /tmp/Rtmpi2xL70/logistic-202204262035-3-703fc2.csv, /tmp/Rtmpi2xL70/logistic-202204262035-4-703fc2.csv
#> 
#> Checking sampler transitions treedepth.
#> Treedepth satisfactory for all transitions.
#> 
#> Checking sampler transitions for divergences.
#> No divergent transitions found.
#> 
#> Checking E-BFMI - sampler transitions HMC potential energy.
#> E-BFMI satisfactory.
#> 
#> Effective sample size satisfactory.
#> 
#> Split R-hat values satisfactory all parameters.
#> 
#> Processing complete, no problems detected.

fit$cmdstan_summary()

#> Inference for Stan model: logistic_model
#> 4 chains: each with iter=(1000,1000,1000,1000); warmup=(0,0,0,0); thin=(1,1,1,1); 4000 iterations saved.
#> 
#> Warmup took (0.025, 0.022, 0.025, 0.022) seconds, 0.094 seconds total
#> Sampling took (0.066, 0.065, 0.067, 0.064) seconds, 0.26 seconds total
#> 
#>                     Mean     MCSE  StdDev     5%       50%       95%    N_Eff  N_Eff/s    R_hat
#> 
#> lp__            -6.6e+01  3.1e-02     1.5    -69  -6.6e+01  -6.4e+01     2250     8589      1.0
#> accept_stat__       0.92  1.5e-03   0.096   0.72      0.95       1.0  4.2e+03  1.6e+04  1.0e+00
#> stepsize__          0.70  2.6e-02   0.037   0.66      0.71      0.76  2.0e+00  7.7e+00  1.6e+13
#> treedepth__          2.5  1.1e-02    0.52    2.0       3.0       3.0  2.4e+03  9.1e+03  1.0e+00
#> n_leapfrog__         5.5  7.3e-02     2.0    3.0       7.0       7.0  7.2e+02  2.8e+03  1.0e+00
#> divergent__         0.00      nan    0.00   0.00      0.00      0.00      nan      nan      nan
#> energy__              68  4.7e-02     2.0     65        68        72  1.8e+03  7.0e+03  1.0e+00
#> 
#> alpha            3.8e-01  3.5e-03    0.22  0.028   3.8e-01   7.4e-01     3792    14473     1.00
#> beta[1]         -6.6e-01  4.1e-03    0.26   -1.1  -6.6e-01  -2.5e-01     3876    14792     1.00
#> beta[2]         -2.8e-01  3.9e-03    0.23  -0.67  -2.7e-01   9.3e-02     3659    13966      1.0
#> beta[3]          6.8e-01  4.2e-03    0.26   0.26   6.7e-01   1.1e+00     4011    15308      1.0
#> log_lik[1]      -5.2e-01  1.7e-03    0.10  -0.69  -5.1e-01  -3.6e-01     3502    13367     1.00
#> log_lik[2]      -4.0e-01  2.2e-03    0.15  -0.67  -3.9e-01  -2.0e-01     4360    16643     1.00
#> log_lik[3]      -5.1e-01  3.9e-03    0.23  -0.95  -4.7e-01  -2.0e-01     3464    13220     1.00
#> log_lik[4]      -4.5e-01  2.5e-03    0.15  -0.73  -4.3e-01  -2.3e-01     3635    13874     1.00
#> log_lik[5]      -1.2e+00  4.4e-03    0.28   -1.7  -1.2e+00  -7.6e-01     4095    15630      1.0
#> log_lik[6]      -5.9e-01  2.9e-03    0.19  -0.93  -5.7e-01  -3.2e-01     4244    16198      1.0
#> log_lik[7]      -6.4e-01  1.9e-03    0.12  -0.85  -6.3e-01  -4.6e-01     4344    16580     1.00
#> log_lik[8]      -2.8e-01  2.1e-03    0.13  -0.53  -2.5e-01  -1.0e-01     3992    15236     1.00
#> log_lik[9]      -6.9e-01  2.5e-03    0.17  -0.98  -6.7e-01  -4.4e-01     4315    16469      1.0
#> log_lik[10]     -7.4e-01  4.2e-03    0.25   -1.2  -7.2e-01  -3.9e-01     3331    12715     1.00
#> log_lik[11]     -2.8e-01  2.0e-03    0.12  -0.51  -2.6e-01  -1.2e-01     3850    14695      1.0
#> log_lik[12]     -5.0e-01  3.8e-03    0.24  -0.95  -4.6e-01  -1.9e-01     3913    14935     1.00
#> log_lik[13]     -6.6e-01  3.2e-03    0.21   -1.0  -6.4e-01  -3.6e-01     4274    16313      1.0
#> log_lik[14]     -3.7e-01  3.0e-03    0.18  -0.73  -3.4e-01  -1.3e-01     3629    13850     1.00
#> log_lik[15]     -2.8e-01  1.8e-03    0.11  -0.47  -2.7e-01  -1.3e-01     3730    14236      1.0
#> log_lik[16]     -2.7e-01  1.4e-03   0.086  -0.43  -2.6e-01  -1.5e-01     3704    14138      1.0
#> log_lik[17]     -1.6e+00  4.8e-03    0.29   -2.1  -1.6e+00  -1.2e+00     3679    14042     1.00
#> log_lik[18]     -4.8e-01  1.6e-03    0.11  -0.67  -4.7e-01  -3.2e-01     4244    16200      1.0
#> log_lik[19]     -2.3e-01  1.2e-03   0.076  -0.37  -2.2e-01  -1.3e-01     3740    14274     1.00
#> log_lik[20]     -1.2e-01  1.3e-03   0.081  -0.28  -9.6e-02  -2.8e-02     3626    13841     1.00
#> log_lik[21]     -2.1e-01  1.4e-03   0.087  -0.37  -2.0e-01  -9.5e-02     3668    14001     1.00
#> log_lik[22]     -5.7e-01  2.3e-03    0.15  -0.83  -5.5e-01  -3.5e-01     4148    15832      1.0
#> log_lik[23]     -3.3e-01  2.2e-03    0.14  -0.59  -3.1e-01  -1.4e-01     4140    15802     1.00
#> log_lik[24]     -1.4e-01  1.1e-03   0.068  -0.26  -1.2e-01  -5.2e-02     3613    13789     1.00
#> log_lik[25]     -4.6e-01  1.9e-03    0.12  -0.67  -4.4e-01  -2.8e-01     4104    15665     1.00
#> log_lik[26]     -1.5e+00  5.6e-03    0.34   -2.1  -1.5e+00  -1.0e+00     3785    14446     1.00
#> log_lik[27]     -3.0e-01  1.9e-03    0.12  -0.53  -2.9e-01  -1.4e-01     3839    14653      1.0
#> log_lik[28]     -4.4e-01  1.3e-03   0.082  -0.59  -4.4e-01  -3.2e-01     3782    14433     1.00
#> log_lik[29]     -7.2e-01  3.7e-03    0.23   -1.1  -7.0e-01  -3.9e-01     3903    14895     1.00
#> log_lik[30]     -6.9e-01  2.9e-03    0.19   -1.0  -6.7e-01  -4.1e-01     4350    16602      1.0
#> log_lik[31]     -4.8e-01  2.6e-03    0.16  -0.78  -4.6e-01  -2.5e-01     4022    15350      1.0
#> log_lik[32]     -4.2e-01  1.7e-03    0.11  -0.61  -4.1e-01  -2.7e-01     3958    15106      1.0
#> log_lik[33]     -4.1e-01  1.9e-03    0.13  -0.64  -4.0e-01  -2.3e-01     4380    16719     1.00
#> log_lik[34]     -6.4e-02  8.8e-04   0.052  -0.16  -5.0e-02  -1.3e-02     3435    13109     1.00
#> log_lik[35]     -5.9e-01  2.7e-03    0.18  -0.91  -5.7e-01  -3.3e-01     4382    16727     1.00
#> log_lik[36]     -3.2e-01  2.0e-03    0.13  -0.57  -3.0e-01  -1.4e-01     4159    15874     1.00
#> log_lik[37]     -7.0e-01  4.0e-03    0.23   -1.1  -6.7e-01  -3.7e-01     3470    13245     1.00
#> log_lik[38]     -3.1e-01  2.4e-03    0.15  -0.59  -2.9e-01  -1.2e-01     3807    14529     1.00
#> log_lik[39]     -1.8e-01  1.9e-03    0.11  -0.41  -1.6e-01  -5.1e-02     3722    14208     1.00
#> log_lik[40]     -6.8e-01  1.9e-03    0.13  -0.90  -6.7e-01  -4.8e-01     4452    16991     1.00
#> log_lik[41]     -1.1e+00  4.2e-03    0.26   -1.6  -1.1e+00  -7.5e-01     3766    14376      1.0
#> log_lik[42]     -9.3e-01  2.9e-03    0.20   -1.3  -9.1e-01  -6.3e-01     4473    17072     1.00
#> log_lik[43]     -4.1e-01  4.1e-03    0.26  -0.93  -3.5e-01  -1.1e-01     4195    16010     1.00
#> log_lik[44]     -1.2e+00  3.3e-03    0.19   -1.5  -1.2e+00  -8.8e-01     3362    12831     1.00
#> log_lik[45]     -3.6e-01  2.1e-03    0.12  -0.60  -3.5e-01  -1.9e-01     3565    13607     1.00
#> log_lik[46]     -5.8e-01  1.9e-03    0.13  -0.81  -5.8e-01  -3.9e-01     4415    16852     1.00
#> log_lik[47]     -3.1e-01  2.0e-03    0.13  -0.55  -2.9e-01  -1.3e-01     3938    15032     1.00
#> log_lik[48]     -3.2e-01  1.4e-03   0.083  -0.47  -3.2e-01  -2.0e-01     3704    14139     1.00
#> log_lik[49]     -3.2e-01  1.3e-03   0.079  -0.46  -3.1e-01  -2.0e-01     3618    13808      1.0
#> log_lik[50]     -1.3e+00  5.4e-03    0.33   -1.9  -1.3e+00  -7.9e-01     3829    14614     1.00
#> log_lik[51]     -2.9e-01  1.6e-03   0.095  -0.46  -2.8e-01  -1.6e-01     3785    14447     1.00
#> log_lik[52]     -8.3e-01  2.4e-03    0.14   -1.1  -8.3e-01  -6.1e-01     3541    13514     1.00
#> log_lik[53]     -4.0e-01  2.1e-03    0.13  -0.64  -3.8e-01  -2.1e-01     3949    15073      1.0
#> log_lik[54]     -3.7e-01  2.5e-03    0.15  -0.66  -3.5e-01  -1.7e-01     3671    14012     1.00
#> log_lik[55]     -3.9e-01  2.1e-03    0.14  -0.64  -3.7e-01  -2.0e-01     4337    16552     1.00
#> log_lik[56]     -3.2e-01  3.0e-03    0.19  -0.69  -2.8e-01  -9.1e-02     4044    15435     1.00
#> log_lik[57]     -6.6e-01  2.0e-03    0.12  -0.86  -6.5e-01  -4.8e-01     3642    13900     1.00
#> log_lik[58]     -9.5e-01  5.7e-03    0.36   -1.6  -9.0e-01  -4.6e-01     3840    14655      1.0
#> log_lik[59]     -1.4e+00  5.2e-03    0.34   -2.0  -1.3e+00  -8.4e-01     4305    16430     1.00
#> log_lik[60]     -9.8e-01  2.8e-03    0.16   -1.3  -9.7e-01  -7.3e-01     3395    12959     1.00
#> log_lik[61]     -5.4e-01  1.6e-03   0.098  -0.71  -5.3e-01  -3.9e-01     3723    14209     1.00
#> log_lik[62]     -8.7e-01  5.3e-03    0.32   -1.4  -8.4e-01  -4.3e-01     3651    13936     1.00
#> log_lik[63]     -1.2e-01  1.2e-03   0.073  -0.25  -1.0e-01  -3.2e-02     3613    13790     1.00
#> log_lik[64]     -9.1e-01  3.8e-03    0.25   -1.4  -8.8e-01  -5.4e-01     4354    16619      1.0
#> log_lik[65]     -2.0e+00  9.8e-03    0.59   -3.1  -2.0e+00  -1.1e+00     3688    14076      1.0
#> log_lik[66]     -5.1e-01  2.2e-03    0.13  -0.75  -4.9e-01  -3.1e-01     3882    14818      1.0
#> log_lik[67]     -2.8e-01  1.3e-03   0.081  -0.42  -2.7e-01  -1.6e-01     3857    14720     1.00
#> log_lik[68]     -1.1e+00  3.8e-03    0.24   -1.5  -1.0e+00  -7.1e-01     4016    15327      1.0
#> log_lik[69]     -4.3e-01  1.4e-03   0.083  -0.58  -4.3e-01  -3.1e-01     3551    13555      1.0
#> log_lik[70]     -6.3e-01  3.5e-03    0.24   -1.1  -6.1e-01  -3.1e-01     4427    16898     1.00
#> log_lik[71]     -6.1e-01  3.1e-03    0.21  -0.98  -5.9e-01  -3.2e-01     4437    16937      1.0
#> log_lik[72]     -4.7e-01  2.7e-03    0.17  -0.78  -4.5e-01  -2.3e-01     4335    16545     1.00
#> log_lik[73]     -1.5e+00  5.8e-03    0.36   -2.1  -1.5e+00  -9.3e-01     3871    14774     1.00
#> log_lik[74]     -9.5e-01  3.4e-03    0.20   -1.3  -9.3e-01  -6.4e-01     3536    13498     1.00
#> log_lik[75]     -1.1e+00  6.6e-03    0.40   -1.8  -1.1e+00  -5.6e-01     3593    13715     1.00
#> log_lik[76]     -3.8e-01  2.5e-03    0.15  -0.66  -3.6e-01  -1.7e-01     3552    13557     1.00
#> log_lik[77]     -8.8e-01  2.1e-03    0.14   -1.1  -8.7e-01  -6.7e-01     4046    15443      1.0
#> log_lik[78]     -4.9e-01  2.8e-03    0.18  -0.81  -4.7e-01  -2.5e-01     4034    15397     1.00
#> log_lik[79]     -7.7e-01  3.4e-03    0.20   -1.1  -7.5e-01  -4.8e-01     3402    12985     1.00
#> log_lik[80]     -5.4e-01  2.9e-03    0.19  -0.89  -5.2e-01  -2.7e-01     4491    17140     1.00
#> log_lik[81]     -1.6e-01  1.6e-03    0.10  -0.35  -1.4e-01  -4.9e-02     3825    14599     1.00
#> log_lik[82]     -2.3e-01  2.2e-03    0.14  -0.49  -2.0e-01  -6.3e-02     4083    15586     1.00
#> log_lik[83]     -3.4e-01  1.4e-03   0.082  -0.49  -3.4e-01  -2.2e-01     3418    13047     1.00
#> log_lik[84]     -2.8e-01  1.6e-03   0.093  -0.44  -2.7e-01  -1.4e-01     3547    13536     1.00
#> log_lik[85]     -1.3e-01  1.3e-03   0.077  -0.29  -1.2e-01  -4.1e-02     3578    13656     1.00
#> log_lik[86]     -1.1e+00  4.7e-03    0.32   -1.7  -1.1e+00  -6.5e-01     4467    17049     1.00
#> log_lik[87]     -8.2e-01  2.0e-03    0.13   -1.0  -8.2e-01  -6.3e-01     4081    15577     1.00
#> log_lik[88]     -7.8e-01  3.7e-03    0.24   -1.2  -7.6e-01  -4.4e-01     4178    15948      1.0
#> log_lik[89]     -1.3e+00  4.8e-03    0.31   -1.8  -1.2e+00  -8.0e-01     4256    16245     1.00
#> log_lik[90]     -2.6e-01  2.1e-03    0.13  -0.52  -2.4e-01  -9.3e-02     3950    15078     1.00
#> log_lik[91]     -3.9e-01  2.0e-03    0.13  -0.62  -3.7e-01  -2.1e-01     4364    16656     1.00
#> log_lik[92]     -1.5e+00  5.5e-03    0.34   -2.1  -1.5e+00  -9.8e-01     3882    14816      1.0
#> log_lik[93]     -7.4e-01  3.9e-03    0.23   -1.2  -7.1e-01  -4.1e-01     3338    12742     1.00
#> log_lik[94]     -3.2e-01  1.5e-03   0.090  -0.48  -3.1e-01  -1.9e-01     3435    13112     1.00
#> log_lik[95]     -3.9e-01  1.8e-03    0.11  -0.59  -3.7e-01  -2.3e-01     3905    14906      1.0
#> log_lik[96]     -1.6e+00  4.9e-03    0.29   -2.1  -1.6e+00  -1.1e+00     3423    13064     1.00
#> log_lik[97]     -4.3e-01  1.6e-03   0.100  -0.61  -4.2e-01  -2.8e-01     4060    15496     1.00
#> log_lik[98]     -1.1e+00  5.9e-03    0.37   -1.7  -1.0e+00  -5.3e-01     3930    14999      1.0
#> log_lik[99]     -7.0e-01  2.1e-03    0.14  -0.94  -6.9e-01  -4.8e-01     4466    17045      1.0
#> log_lik[100]    -3.9e-01  1.7e-03   0.099  -0.57  -3.8e-01  -2.5e-01     3610    13779     1.00
#> 
#> Samples were drawn using hmc with nuts.
#> For each parameter, N_Eff is a crude measure of effective sample size,
#> and R_hat is the potential scale reduction factor on split chains (at 
#> convergence, R_hat=1).

# }