Extract posterior draws

Extract posterior draws after MCMC or approximate posterior draws after variational approximation using formats provided by the posterior package.

The variables include the parameters, transformed parameters, and generated quantities from the Stan program as well as lp__, the total log probability (target) accumulated in the model block.

Usage

$draws(variables = NULL, inc_warmup = FALSE, ...)

Arguments

variables: (character vector) The variables (parameters and generated quantities) to read in. If NULL (the default) then the draws of all variables are included.

inc_warmup: (logical) For MCMC only, should warmup draws be included? Defaults to FALSE.

...: Arguments passed on to posterior::as_draws().

Value

For MCMC, a 3-D draws_array object (iteration x chain x variable).

For variational inference, a 2-D draws_matrix object (draw x variable). An additional variable lp_approx__ is also included, which is the log density of the variational approximation to the posterior evaluated at each of the draws.

For optimization, a 1-row draws_matrix with one column per variable. These are not actually draws, just point estimates stored in the draws_matrix format.

Examples

# \dontrun{
stan_program <- tempfile(fileext=".stan")
cat("
parameters {
  real<lower=0> alpha;
  vector[6] beta;
}
model {
 alpha ~ exponential(1);
 beta ~ std_normal();
}
", file = stan_program)

# only using capture.output to avoid too much printed output in example
out <- utils::capture.output(
  mod <- cmdstan_model(stan_program),
  fit <- mod$sample(save_warmup=TRUE)
)

#> Compiling Stan program...

# returned as 3-D array (see ?posterior::draws_array)
draws <- fit$draws()
dim(draws)

#> [1] 1000    4    8

str(draws)

#>  'draws_array' num [1:1000, 1:4, 1:8] -4.61 -3.21 -3.09 -9.81 -8.23 ...
#>  - attr(*, "dimnames")=List of 3
#>   ..$ iteration: chr [1:1000] "1" "2" "3" "4" ...
#>   ..$ chain    : chr [1:4] "1" "2" "3" "4"
#>   ..$ variable : chr [1:8] "lp__" "alpha" "beta[1]" "beta[2]" ...

# can easily convert to other formats that have chains combined,
# e.g. matrix, data frame, list
library("posterior")
draws <- fit$draws()
as_draws_matrix(draws)

#> # A draws_matrix: 4000 draws, and 8 variables
#>     variable
#> draw lp__ alpha beta[1] beta[2] beta[3] beta[4] beta[5] beta[6]
#>   1  -4.6 0.356    0.18  -0.865  -1.657    1.30  -0.349   1.052
#>   2  -3.2 0.182    0.67   0.356  -0.844    0.62  -0.093  -0.986
#>   3  -3.1 1.980   -1.15  -0.089  -0.120   -1.48  -0.174   0.126
#>   4  -9.8 0.057    2.02   0.618   0.444   -2.58  -1.273   0.901
#>   5  -8.2 1.837   -1.94  -0.835   0.735    2.78   1.026  -0.494
#>   6  -8.5 0.472    0.37  -0.299  -0.693   -3.44  -1.451   0.033
#>   7  -7.3 0.371   -1.07   1.063   0.394    2.65   1.502  -0.471
#>   8  -8.2 0.182    1.36   2.491  -0.183    0.66  -1.281  -1.559
#>   9  -6.4 0.363   -1.32  -2.461   0.055   -0.64   1.165   0.680
#>   10 -2.4 1.757    1.20   0.126  -0.101    0.52  -0.432  -0.759
#> # ... with 3990 more draws

as_draws_df(draws)

#> # A draws_df: 1000 iterations, 4 chains, and 8 variables
#>    lp__ alpha beta[1] beta[2] beta[3] beta[4] beta[5] beta[6]
#> 1  -4.6 0.356    0.18  -0.865  -1.657    1.30  -0.349   1.052
#> 2  -3.2 0.182    0.67   0.356  -0.844    0.62  -0.093  -0.986
#> 3  -3.1 1.980   -1.15  -0.089  -0.120   -1.48  -0.174   0.126
#> 4  -9.8 0.057    2.02   0.618   0.444   -2.58  -1.273   0.901
#> 5  -8.2 1.837   -1.94  -0.835   0.735    2.78   1.026  -0.494
#> 6  -8.5 0.472    0.37  -0.299  -0.693   -3.44  -1.451   0.033
#> 7  -7.3 0.371   -1.07   1.063   0.394    2.65   1.502  -0.471
#> 8  -8.2 0.182    1.36   2.491  -0.183    0.66  -1.281  -1.559
#> 9  -6.4 0.363   -1.32  -2.461   0.055   -0.64   1.165   0.680
#> 10 -2.4 1.757    1.20   0.126  -0.101    0.52  -0.432  -0.759
#> # ... with 3990 more draws
#> # ... hidden meta-columns {'.chain', '.iteration', '.draw'}

as_draws_list(draws)

#> # A draws_list: 1000 iterations, 4 chains, and 8 variables
#> 
#> [chain = 1]
#> $lp__
#>  [1] -4.6 -3.2 -3.1 -9.8 -8.2 -8.5 -7.3 -8.2 -6.4 -2.4
#> 
#> $alpha
#>  [1] 0.356 0.182 1.980 0.057 1.837 0.472 0.371 0.182 0.363 1.757
#> 
#> $`beta[1]`
#>  [1]  0.18  0.67 -1.15  2.02 -1.94  0.37 -1.07  1.36 -1.32  1.20
#> 
#> $`beta[2]`
#>  [1] -0.865  0.356 -0.089  0.618 -0.835 -0.299  1.063  2.491 -2.461  0.126
#> 
#> 
#> [chain = 2]
#> $lp__
#>  [1]  -4.8 -10.4  -3.6  -2.4  -5.3  -2.1  -3.0  -3.7  -3.2  -3.7
#> 
#> $alpha
#>  [1] 0.648 0.069 0.735 1.105 0.688 0.619 0.825 0.236 2.181 0.170
#> 
#> $`beta[1]`
#>  [1]  0.039  2.356 -0.742  0.857  0.574  0.708 -0.802 -1.351  0.165  0.126
#> 
#> $`beta[2]`
#>  [1]  1.23 -0.26 -1.28 -1.12 -1.76 -0.79 -0.26 -0.36 -0.52  0.59
#> 
#> # ... with 990 more iterations, and 2 more chains, and 4 more variables

# can select specific parameters
fit$draws("alpha")

#> # A draws_array: 1000 iterations, 4 chains, and 1 variables
#> , , variable = alpha
#> 
#>          chain
#> iteration     1     2    3    4
#>         1 0.356 0.648 0.13 0.87
#>         2 0.182 0.069 4.62 0.66
#>         3 1.980 0.735 2.08 1.71
#>         4 0.057 1.105 0.49 1.13
#>         5 1.837 0.688 1.43 0.19
#> 
#> # ... with 995 more iterations

fit$draws("beta")  # selects entire vector beta

#> # A draws_array: 1000 iterations, 4 chains, and 6 variables
#> , , variable = beta[1]
#> 
#>          chain
#> iteration     1      2      3     4
#>         1  0.18  0.039 -0.089 -0.13
#>         2  0.67  2.356  0.155 -1.59
#>         3 -1.15 -0.742  0.522 -0.16
#>         4  2.02  0.857  0.777  0.40
#>         5 -1.94  0.574 -0.108 -0.92
#> 
#> , , variable = beta[2]
#> 
#>          chain
#> iteration      1     2     3     4
#>         1 -0.865  1.23  1.11  0.40
#>         2  0.356 -0.26  0.45 -0.85
#>         3 -0.089 -1.28  0.40  0.53
#>         4  0.618 -1.12 -2.42 -0.14
#>         5 -0.835 -1.76 -1.15  0.46
#> 
#> , , variable = beta[3]
#> 
#>          chain
#> iteration     1    2      3     4
#>         1 -1.66 1.45  0.508  0.35
#>         2 -0.84 2.21 -0.037 -0.22
#>         3 -0.12 0.53  0.967 -1.26
#>         4  0.44 0.76 -1.024 -1.01
#>         5  0.73 1.67 -0.055  1.54
#> 
#> , , variable = beta[4]
#> 
#>          chain
#> iteration     1     2     3     4
#>         1  1.30  1.73 -1.77 -1.59
#>         2  0.62 -1.17  2.14  1.41
#>         3 -1.48 -0.71  1.12 -1.19
#>         4 -2.58  0.15 -0.36  0.69
#>         5  2.78 -1.25 -0.92  1.34
#> 
#> # ... with 995 more iterations, and 2 more variables

fit$draws(c("alpha", "beta[2]"))

#> # A draws_array: 1000 iterations, 4 chains, and 2 variables
#> , , variable = alpha
#> 
#>          chain
#> iteration     1     2    3    4
#>         1 0.356 0.648 0.13 0.87
#>         2 0.182 0.069 4.62 0.66
#>         3 1.980 0.735 2.08 1.71
#>         4 0.057 1.105 0.49 1.13
#>         5 1.837 0.688 1.43 0.19
#> 
#> , , variable = beta[2]
#> 
#>          chain
#> iteration      1     2     3     4
#>         1 -0.865  1.23  1.11  0.40
#>         2  0.356 -0.26  0.45 -0.85
#>         3 -0.089 -1.28  0.40  0.53
#>         4  0.618 -1.12 -2.42 -0.14
#>         5 -0.835 -1.76 -1.15  0.46
#> 
#> # ... with 995 more iterations

# }