Extract inverse metric (mass matrix) after MCMC

Extract the inverse metric (mass matrix) for each MCMC chain.

inv_metric(matrix = TRUE)

Arguments

matrix
(logical) If a diagonal metric was used, setting matrix = FALSE returns a list containing just the diagonals of the matrices instead of the full matrices. Setting matrix = FALSE has no effect for dense metrics.

Value

A list of length equal to the number of MCMC chains. See the matrix argument for details.

See also

Examples

# \dontrun{
fit <- cmdstanr_example("logistic")
fit$inv_metric()

#> $`1`
#>           [,1]      [,2]      [,3]      [,4]
#> [1,] 0.0427641 0.0000000 0.0000000 0.0000000
#> [2,] 0.0000000 0.0642607 0.0000000 0.0000000
#> [3,] 0.0000000 0.0000000 0.0541114 0.0000000
#> [4,] 0.0000000 0.0000000 0.0000000 0.0786657
#> 
#> $`2`
#>           [,1]      [,2]      [,3]      [,4]
#> [1,] 0.0462032 0.0000000 0.0000000 0.0000000
#> [2,] 0.0000000 0.0708086 0.0000000 0.0000000
#> [3,] 0.0000000 0.0000000 0.0509175 0.0000000
#> [4,] 0.0000000 0.0000000 0.0000000 0.0663431
#> 
#> $`3`
#>           [,1]      [,2]      [,3]      [,4]
#> [1,] 0.0413976 0.0000000 0.0000000 0.0000000
#> [2,] 0.0000000 0.0612544 0.0000000 0.0000000
#> [3,] 0.0000000 0.0000000 0.0516873 0.0000000
#> [4,] 0.0000000 0.0000000 0.0000000 0.0825751
#> 
#> $`4`
#>           [,1]      [,2]      [,3]      [,4]
#> [1,] 0.0483107 0.0000000 0.0000000 0.0000000
#> [2,] 0.0000000 0.0707139 0.0000000 0.0000000
#> [3,] 0.0000000 0.0000000 0.0558434 0.0000000
#> [4,] 0.0000000 0.0000000 0.0000000 0.0769705
#> 

fit$inv_metric(matrix=FALSE)

#> $`1`
#> [1] 0.0427641 0.0642607 0.0541114 0.0786657
#> 
#> $`2`
#> [1] 0.0462032 0.0708086 0.0509175 0.0663431
#> 
#> $`3`
#> [1] 0.0413976 0.0612544 0.0516873 0.0825751
#> 
#> $`4`
#> [1] 0.0483107 0.0707139 0.0558434 0.0769705
#> 

fit <- cmdstanr_example("logistic", metric = "dense_e")
fit$inv_metric()

#> $`1`
#>              [,1]         [,2]         [,3]        [,4]
#> [1,]  0.046573400 -6.21164e-04  1.01049e-03 -0.00618128
#> [2,] -0.000621164  6.75953e-02  3.15148e-05 -0.01506060
#> [3,]  0.001010490  3.15148e-05  4.99230e-02 -0.01078110
#> [4,] -0.006181280 -1.50606e-02 -1.07811e-02  0.07336490
#> 
#> $`2`
#>              [,1]        [,2]         [,3]        [,4]
#> [1,]  0.049113400 -0.00364659  0.000534645  0.00374945
#> [2,] -0.003646590  0.06232720 -0.002478890 -0.00767640
#> [3,]  0.000534645 -0.00247889  0.050718100 -0.01287720
#> [4,]  0.003749450 -0.00767640 -0.012877200  0.06211910
#> 
#> $`3`
#>              [,1]        [,2]        [,3]         [,4]
#> [1,]  0.049970300 -0.00263438 -0.00172118 -0.000169874
#> [2,] -0.002634380  0.05825500 -0.00535528 -0.005990200
#> [3,] -0.001721180 -0.00535528  0.04690110 -0.008007390
#> [4,] -0.000169874 -0.00599020 -0.00800739  0.067966800
#> 
#> $`4`
#>              [,1]         [,2]         [,3]         [,4]
#> [1,]  0.041539200 -0.005749800 -0.000964921  0.007615130
#> [2,] -0.005749800  0.069531600 -0.004916180 -0.000402701
#> [3,] -0.000964921 -0.004916180  0.066553600 -0.026309300
#> [4,]  0.007615130 -0.000402701 -0.026309300  0.077661500
#> 

# }