Extract (penalized) maximum likelihood estimate after optimization

The $mle() method is only available for CmdStanMLE objects. It returns the penalized maximum likelihood estimate (posterior mode) as a numeric vector with one element per variable. The returned vector does not include lp__, the total log probability (target) accumulated in the model block of the Stan program, which is available via the $lp() method and also included in the $draws() method.

Usage

$mle(variables = NULL)

Arguments

variables: (character vector) The variables (parameters and generated quantities) to include. If NULL (the default) then all variables are included.

Value

A numeric vector. See Examples.

See also

Examples

# \dontrun{
fit <- cmdstanr_example("logistic", method = "optimize")

#> Model executable is up to date!

fit$mle()

#>     alpha   beta[1]   beta[2]   beta[3] 
#>  0.364444 -0.631529 -0.258967  0.648531 

fit$mle("alpha")

#>    alpha 
#> 0.364444 

fit$mle("beta")

#>   beta[1]   beta[2]   beta[3] 
#> -0.631529 -0.258967  0.648531 

fit$mle("beta[2]")

#>   beta[2] 
#> -0.258967 

# }