Compute a summary table of estimates and diagnostics

The $summary() method runs summarise_draws() from the posterior package and returns the output. For MCMC, only post-warmup draws are included in the summary.

There is also a $print() method that prints the same summary stats but removes the extra formatting used for printing tibbles and returns the fitted model object itself. The $print() method may also be faster than $summary() because it is designed to only compute the summary statistics for the variables that will actually fit in the printed output whereas $summary() will compute them for all of the specified variables in order to be able to return them to the user. See Examples.

summary(variables = NULL, ...)

Arguments

variables
(character vector) The variables to include.

...
Optional arguments to pass to posterior::summarise_draws().

Value

The $summary() method returns the tibble data frame created by posterior::summarise_draws().

The $print() method returns the fitted model object itself (invisibly), which is the standard behavior for print methods in R.

See also

Examples

# \dontrun{
fit <- cmdstanr_example("logistic")
fit$summary()

#> # A tibble: 105 × 10
#>    variable      mean  median     sd   mad       q5      q95  rhat ess_bulk
#>    <chr>        <dbl>   <dbl>  <dbl> <dbl>    <dbl>    <dbl> <dbl>    <dbl>
#>  1 lp__       -65.9   -65.6   1.43   1.21  -68.7    -64.3    1.00     1939.
#>  2 alpha        0.382   0.381 0.221  0.221   0.0221   0.750  1.00     4337.
#>  3 beta[1]     -0.664  -0.654 0.250  0.250  -1.09    -0.260  1.00     4508.
#>  4 beta[2]     -0.274  -0.269 0.219  0.216  -0.634    0.0804 1.00     3743.
#>  5 beta[3]      0.679   0.671 0.267  0.260   0.247    1.14   1.00     4066.
#>  6 log_lik[1]  -0.515  -0.509 0.0993 0.100  -0.690   -0.363  1.00     4448.
#>  7 log_lik[2]  -0.404  -0.385 0.146  0.141  -0.667   -0.203  1.00     4695.
#>  8 log_lik[3]  -0.497  -0.465 0.214  0.200  -0.898   -0.208  1.00     4254.
#>  9 log_lik[4]  -0.450  -0.434 0.150  0.141  -0.718   -0.238  1.00     3721.
#> 10 log_lik[5]  -1.18   -1.16  0.284  0.276  -1.69    -0.758  0.999    4506.
#> # … with 95 more rows, and 1 more variable: ess_tail <dbl>

fit$print()

#>    variable   mean median   sd  mad     q5    q95 rhat ess_bulk ess_tail
#>  lp__       -65.94 -65.61 1.43 1.21 -68.71 -64.29 1.00     1939     3139
#>  alpha        0.38   0.38 0.22 0.22   0.02   0.75 1.00     4337     3142
#>  beta[1]     -0.66  -0.65 0.25 0.25  -1.09  -0.26 1.00     4508     3023
#>  beta[2]     -0.27  -0.27 0.22 0.22  -0.63   0.08 1.00     3742     2820
#>  beta[3]      0.68   0.67 0.27 0.26   0.25   1.14 1.00     4066     2894
#>  log_lik[1]  -0.52  -0.51 0.10 0.10  -0.69  -0.36 1.00     4447     2755
#>  log_lik[2]  -0.40  -0.39 0.15 0.14  -0.67  -0.20 1.00     4695     3425
#>  log_lik[3]  -0.50  -0.46 0.21 0.20  -0.90  -0.21 1.00     4253     2934
#>  log_lik[4]  -0.45  -0.43 0.15 0.14  -0.72  -0.24 1.00     3720     3080
#>  log_lik[5]  -1.18  -1.16 0.28 0.28  -1.69  -0.76 1.00     4506     3360
#> 
#>  # showing 10 of 105 rows (change via 'max_rows' argument or 'cmdstanr_max_rows' option)

fit$print(max_rows = 2) # same as print(fit, max_rows = 2)

#>  variable   mean median   sd  mad     q5    q95 rhat ess_bulk ess_tail
#>     lp__  -65.94 -65.61 1.43 1.21 -68.71 -64.29 1.00     1939     3139
#>     alpha   0.38   0.38 0.22 0.22   0.02   0.75 1.00     4337     3142
#> 
#>  # showing 2 of 105 rows (change via 'max_rows' argument or 'cmdstanr_max_rows' option)

# include only certain variables
fit$summary("beta")

#> # A tibble: 3 × 10
#>   variable   mean median    sd   mad     q5     q95  rhat ess_bulk ess_tail
#>   <chr>     <dbl>  <dbl> <dbl> <dbl>  <dbl>   <dbl> <dbl>    <dbl>    <dbl>
#> 1 beta[1]  -0.664 -0.654 0.250 0.250 -1.09  -0.260   1.00    4508.    3024.
#> 2 beta[2]  -0.274 -0.269 0.219 0.216 -0.634  0.0804  1.00    3743.    2821.
#> 3 beta[3]   0.679  0.671 0.267 0.260  0.247  1.14    1.00    4066.    2894.

fit$print(c("alpha", "beta[2]"))

#>  variable  mean median   sd  mad    q5  q95 rhat ess_bulk ess_tail
#>   alpha    0.38   0.38 0.22 0.22  0.02 0.75 1.00     4337     3142
#>   beta[2] -0.27  -0.27 0.22 0.22 -0.63 0.08 1.00     3742     2820

# include all variables but only certain summaries
fit$summary(NULL, c("mean", "sd"))

#> # A tibble: 105 × 3
#>    variable      mean     sd
#>    <chr>        <dbl>  <dbl>
#>  1 lp__       -65.9   1.43  
#>  2 alpha        0.382 0.221 
#>  3 beta[1]     -0.664 0.250 
#>  4 beta[2]     -0.274 0.219 
#>  5 beta[3]      0.679 0.267 
#>  6 log_lik[1]  -0.515 0.0993
#>  7 log_lik[2]  -0.404 0.146 
#>  8 log_lik[3]  -0.497 0.214 
#>  9 log_lik[4]  -0.450 0.150 
#> 10 log_lik[5]  -1.18  0.284 
#> # … with 95 more rows

# can use functions created from formulas
# for example, calculate Pr(beta > 0)
fit$summary("beta", prob_gt_0 = ~ mean(. > 0))

#> # A tibble: 3 × 2
#>   variable prob_gt_0
#>   <chr>        <dbl>
#> 1 beta[1]     0.0025
#> 2 beta[2]     0.0992
#> 3 beta[3]     0.997 

# }