Report timing of CmdStan runs

Report the run time in seconds. For MCMC additional information is provided about the run times of individual chains and the warmup and sampling phases.

Usage

$time()

Value

A list with elements

total: (scalar) the total run time.

chains: (data frame) for MCMC only, timing info for the individual chains. The data frame has columns "chain_id", "warmup", "sampling", and "total".

See also

Examples

# \dontrun{
fit_mcmc <- cmdstanr_example("logistic", method = "sample")

#> Model executable is up to date!

fit_mcmc$time()

#> $total
#> [1] 0.552681
#> 
#> $chains
#>   chain_id      warmup    sampling     total
#> 1        1 0.004246950 0.003657103 0.1272390
#> 2        2 0.004976034 0.003442049 0.1288078
#> 3        3 0.003338099 0.002820015 0.1225679
#> 4        4 0.003430843 0.002821922 0.1242170
#> 

fit_mle <- cmdstanr_example("logistic", method = "optimize")

#> Model executable is up to date!

fit_mle$time()

#> $total
#> [1] 0.01351094
#> 

fit_vb <- cmdstanr_example("logistic", method = "variational")

#> Model executable is up to date!

fit_vb$time()

#> $total
#> [1] 0.03305221
#> 

# }