Check syntax of a Stan program

The $check_syntax() method of a CmdStanModel object checks the Stan program for syntax errors and returns TRUE (invisibly) if parsing succeeds. If invalid syntax in found an error is thrown.

Usage

$check_syntax(
  quiet = TRUE,
  include_paths = NULL,
  stanc_options = list()
)

Arguments

quiet: (logical) Should informational messages be printed? Default is TRUE, which will print a message if the model is valid or the compiler error message if there are syntax errors in the model. If FALSE, only the error message will be printed.

include_paths: (character vector) Paths to directories where Stan should look for files specified in #include directives in the Stan program.

stanc_options: (list) Any Stan-to-C++ transpiler options to be used when compiling the model. See the documentation for the $compile() method for details.

Value

The $check_syntax() method returns TRUE (invisibly) if the model is valid.

See also

Examples

# \dontrun{
file <- file.path(cmdstan_path(), "examples/bernoulli/bernoulli.stan")

mod <- cmdstan_model(file, compile = FALSE)
mod$check_syntax()

#> Stan program is syntactically correct

# }