Run Stan's diagnose method

The $diagnose() method of a CmdStanModel object runs Stan's basic diagnostic feature that will calculate the gradients of the initial state and compare them with gradients calculated by finite differences. Discrepancies between the two indicate that there is a problem with the model or initial states or else there is a bug in Stan.

diagnose(
  data = NULL,
  seed = NULL,
  init = NULL,
  output_dir = NULL,
  output_basename = NULL,
  epsilon = NULL,
  error = NULL
)

Arguments

data
(multiple options) The data to use for the variables specified in the data block of the Stan program. One of the following:

A named list of R objects with the names corresponding to variables declared in the data block of the Stan program. Internally this list is then written to JSON for CmdStan using write_stan_json(). See write_stan_json() for details on the conversions performed on R objects before they are passed to Stan.

A path to a data file compatible with CmdStan (JSON or R dump). See the appendices in the CmdStan guide for details on using these formats.

NULL or an empty list if the Stan program has no data block.

seed
(positive integer(s)) A seed for the (P)RNG to pass to CmdStan. In the case of multi-chain sampling the single seed will automatically be augmented by the the run (chain) ID so that each chain uses a different seed. The exception is the transformed data block, which defaults to using same seed for all chains so that the same data is generated for all chains if RNG functions are used. The only time seed should be specified as a vector (one element per chain) is if RNG functions are used in transformed data and the goal is to generate different data for each chain.

init
(multiple options) The initialization method to use for the variables declared in the parameters block of the Stan program. One of the following:

A real number x>0. This initializes all parameters randomly between [-x,x] on the unconstrained parameter space.;

The number 0. This initializes all parameters to 0;

A character vector of paths (one per chain) to JSON or Rdump files containing initial values for all or some parameters. See write_stan_json() to write R objects to JSON files compatible with CmdStan.

A list of lists containing initial values for all or some parameters. For MCMC the list should contain a sublist for each chain. For optimization and variational inference there should be just one sublist. The sublists should have named elements corresponding to the parameters for which you are specifying initial values. See Examples.

A function that returns a single list with names corresponding to the parameters for which you are specifying initial values. The function can take no arguments or a single argument chain_id. For MCMC, if the function has argument chain_id it will be supplied with the chain id (from 1 to number of chains) when called to generate the initial values. See Examples.

output_dir
(string) A path to a directory where CmdStan should write its output CSV files. For interactive use this can typically be left at NULL (temporary directory) since CmdStanR makes the CmdStan output (posterior draws and diagnostics) available in R via methods of the fitted model objects. The behavior of output_dir is as follows:

If NULL (the default), then the CSV files are written to a temporary directory and only saved permanently if the user calls one of the $save_* methods of the fitted model object (e.g., $save_output_files()). These temporary files are removed when the fitted model object is garbage collected (manually or automatically).

If a path, then the files are created in output_dir with names corresponding to the defaults used by $save_output_files().

output_basename
(string) A string to use as a prefix for the names of the output CSV files of CmdStan. If NULL (the default), the basename of the output CSV files will be comprised from the model name, timestamp, and 5 random characters.

epsilon
(positive real) The finite difference step size. Default value is 1e-6.

error
(positive real) The error threshold. Default value is 1e-6.

Value

A CmdStanDiagnose object.

See also

Examples

# \dontrun{
test <- cmdstanr_example("logistic", method = "diagnose")

# retrieve the gradients
test$gradients()

#>   param_idx      value     model finite_diff        error
#> 1         0  1.3362100  -9.33442    -9.33442  2.88445e-08
#> 2         1  1.7816500 -32.65590   -32.65590 -2.89809e-10
#> 3         2 -0.0718337  -2.56116    -2.56116 -2.13148e-08
#> 4         3 -1.9334500  29.16060    29.16060 -3.65847e-08

# }