Run stanc's auto-formatter on the model code.

The $format() method of a CmdStanModel object runs stanc's auto-formatter on the model code. Either saves the formatted model directly back to the file or prints it for inspection.

format(
  overwrite_file = FALSE,
  canonicalize = FALSE,
  backup = TRUE,
  max_line_length = NULL,
  quiet = FALSE
)

Arguments

overwrite_file
(logical) Should the formatted code be written back to the input model file. The default is FALSE.

canonicalize
(list or logical) Defines whether or not the compiler should 'canonicalize' the Stan model, removing things like deprecated syntax. Default is FALSE. If TRUE, all canonicalizations are run. You can also supply a list of strings which represent options. In that case the options are passed to stanc (new in Stan 2.29). See the User's guide section for available canonicalization options.

backup
(logical) If TRUE, create stanfile.bak backups before writing to the file. Disable this option if you're sure you have other copies of the file or are using a version control system like Git. Defaults to TRUE. The value is ignored if overwrite_file = FALSE.

max_line_length
(integer) The maximum length of a line when formatting. The default is NULL, which defers to the default line length of stanc.

quiet
(logical) Should informational messages be suppressed? The default is FALSE.

Value

The $format() method returns TRUE (invisibly) if the model is valid.

See also

Examples

# \dontrun{
file <- write_stan_file("
data {
  int N;
  int y[N];
}
parameters {
  real                     lambda;
}
model {
  target +=
 poisson_log(y | lambda);
}
")
mod <- cmdstan_model(file, compile = FALSE)
mod$format(canonicalize = TRUE)

#> Semantic error in '/var/folders/s0/zfzm55px2nd2v__zlw5xfj2h0000gn/T/RtmpmzUYEz/model_44779dbadddd7550b5164157b77f4d11.stan', line 11, column 1 to column 24:
#>    -------------------------------------------------
#>      9:  model {
#>     10:    target +=
#>     11:   poisson_log(y | lambda);
#>           ^
#>     12:  }
#>     13:  
#>    -------------------------------------------------
#> 
#> Only functions with names ending in _lpdf, _lupdf, _lpmf, _lupmf, _cdf, _lcdf, _lccdf can make use of conditional notation.

#> Error: Syntax error found! See the message above for more information.

# }