Run Stan's MCMC algorithms

The $sample() method of a CmdStanModel object runs the default MCMC algorithm in CmdStan (algorithm=hmc engine=nuts), to produce a set of draws from the posterior distribution of a model conditioned on some data.

Usage

$sample(
  data = NULL,
  seed = NULL,
  refresh = NULL,
  init = NULL,
  save_extra_diagnostics = FALSE,
  output_dir = NULL,
  num_chains = 4,
  num_cores = getOption("mc.cores", 1),
  num_warmup = NULL,
  num_samples = NULL,
  save_warmup = FALSE,
  thin = NULL,
  max_depth = NULL,
  adapt_engaged = TRUE,
  adapt_delta = NULL,
  stepsize = NULL,
  metric = NULL,
  metric_file = NULL,
  inv_metric = NULL,
  init_buffer = NULL,
  term_buffer = NULL,
  window = NULL
)

Arguments shared by all fitting methods

The following arguments can be specified for any of the fitting methods (sample, optimize, variational). Arguments left at NULL default to the default used by the installed version of CmdStan.

data: (multiple options) The data to use. One of the following:

A named list of R objects (like for RStan). Internally this list is then written to JSON for CmdStan using write_stan_json().

A path to a data file compatible with CmdStan (JSON or R dump). See the appendices in the CmdStan manual for details on using these formats.

seed: (positive integer) A seed for the (P)RNG to pass to CmdStan.

refresh: (non-negative integer) The number of iterations between printed screen updates.

init: (multiple options) The initialization method:

A real number x>0 initializes randomly between [-x,x] (on the unconstrained parameter space);

0 initializes to 0;

A character vector of paths (one per chain) to JSON or Rdump files. See write_stan_json() to write R objects to JSON files compatible with CmdStan.

save_extra_diagnostics: (logical) Should auxiliary diagnostic information (beyond standard diagnostics) be written to temporary diagnostic CSV files? This argument replaces CmdStan's diagnostic_file argument and the content written to CSV is controlled by the user's CmdStan installation and not CmdStanR (and for some algorithms no content may be written). The default is save_extra_diagnostics=FALSE, which is appropriate for almost every use case (all diagnostics recommended for users to check are always saved, e.g., divergences for HMC). To save the temporary files created when save_extra_diagnostics=TRUE see the $save_diagnostic_files() method.

output_dir: (string) A path to a directory where CmdStan should write its output CSV files. For interactive use this can typically be left at NULL (temporary directory) since CmdStanR makes the CmdStan output (e.g., posterior draws and diagnostics) available in R via methods of the fitted model objects. The behavior of output_dir is as follows:

If NULL (the default) then the CSV files are written to a temporary directory and only saved permanently if the user calls one of the $save_* methods of the fitted model object (e.g., $save_output_files()).

If a path then the files are created in output_dir with names corresponding the defaults used by $save_output_files() (and similar methods like $save_diagnostic_files()).

Arguments unique to the sample method

In addition to the arguments above, the $sample() method also has its own set of arguments.

The following two arguments are offered by CmdStanR but do not correspond to arguments in CmdStan because all CmdStan arguments pertain to the execution of a single run only.

num_chains: (positive integer) The number of Markov chains to run. The default is 4.

num_cores: (positive integer) The maximum number of cores to use for running parallel chains. If num_cores is not specified then the default is to look for the option "mc.cores", which can be set for an entire R session by options(mc.cores=value). If the "mc.cores" option has not been set then the default is 1.

The rest of the arguments correspond to arguments offered by CmdStan. They are described briefly here and in greater detail in the CmdStan manual. Arguments left at NULL default to the default used by the installed version of CmdStan.

num_samples: (positive integer) The number of post-warmup iterations to run per chain.

num_warmup: (positive integer) The number of warmup iterations to run per chain.

save_warmup: (logical) Should warmup iterations be saved? The default is FALSE.

thin: (positive integer) The period between saved samples. This should typically be left at its default (no thinning) unless memory is a problem.

max_depth: (positive integer) The maximum allowed tree depth for the NUTS engine. See the Tree Depth section of the CmdStan manual for more details.

adapt_engaged: (logical) Do warmup adaptation? The default is TRUE. If a precomputed inverse metric is specified via the inv_metric argument (or metric_file) then, if adapt_engaged=TRUE, Stan will use the provided inverse metric just as an initial guess during adaptation. To turn off adaptation when using a precomputed inverse metric set adapt_engaged=FALSE.

adapt_delta: (real in (0,1)) The adaptation target acceptance statistic.

stepsize: (positive real) The initial step size for the discrete approximation to continuous Hamiltonian dynamics. This is further tuned during warmup.

metric: (character) One of "diag_e", "dense_e", or "unit_e", specifying the geometry of the base manifold. See the Euclidean Metric section of the CmdStan documentation for more details. To specify a precomputed (inverse) metric, see the inv_metric argument below.

metric_file: (character) A character vector containing paths to JSON or Rdump files (one per chain) compatible with CmdStan that contain precomputed inverse metrics. The metric_file argument is inherited from CmdStan but is confusing in that the entry in JSON or Rdump file(s) must be named inv_metric, referring to the inverse metric. We recommend instead using CmdStanR's inv_metric argument (see below) to specify an inverse metric directly using a vector or matrix from your R session.

inv_metric: (vector, matrix) A vector (if metric='diag_e') or a matrix (if metric='dense_e') for initializing the inverse metric, which can be used as an alternative to the metric_file argument. A vector is interpreted as a diagonal metric. The inverse metric is usually set to an estimate of the posterior covariance. See the adapt_engaged argument above for details on (and control over) how specifying a precomputed inverse metric interacts with adaptation.

init_buffer: (nonnegative integer) Width of initial fast timestep adaptation interval during warmup.

term_buffer: (nonnegative integer) Width of final fast timestep adaptation interval during warmup.

window: (nonnegative integer) Initial width of slow timestep/metric adaptation interval.

fixed_param: (logical) When True, call CmdStan with argument "algorithm=fixed_param".

validate_csv: (logical) When True, validate the sampling results in the csv files. Disable if you wish to manually read in the sampling results and validate them.

Value

The $sample() method returns a CmdStanMCMC object.

See also

Examples

if (FALSE) {
# Set path to cmdstan
# (Note: if you installed CmdStan via install_cmdstan() with default settings
# then setting the path is unnecessary but the default below should still work.
# Otherwise use the `path` argument to specify the location of your
# CmdStan installation.)

set_cmdstan_path(path = NULL)

# Create a CmdStanModel object from a Stan program,
# here using the example model that comes with CmdStan
stan_program <- file.path(cmdstan_path(), "examples/bernoulli/bernoulli.stan")
mod <- cmdstan_model(stan_program)
mod$print()

# data as a named list (like RStan)
stan_data <- list(N = 10, y = c(0,1,0,0,0,0,0,0,0,1))

# run MCMC using the 'sample' method
fit_mcmc <- mod$sample(
  data = stan_data,
  seed = 123,
  num_chains = 2,
  num_cores = 2
)

# Use 'posterior' package for summaries
fit_mcmc$summary()

# Call CmdStan's diagnose and stansummary utilities
fit_mcmc$cmdstan_diagnose()
fit_mcmc$cmdstan_summary()

# For models fit using MCMC, if you like working with RStan's stanfit objects
# then you can create one with rstan::read_stan_csv()
if (require(rstan, quietly = TRUE)) {
  stanfit <- rstan::read_stan_csv(fit_mcmc$output_files())
  print(stanfit)
}

# Run 'optimize' method to get a point estimate (default is Stan's LBFGS algorithm)
# and also demonstrate specifying data as a path to a file instead of a list
my_data_file <- file.path(cmdstan_path(), "examples/bernoulli/bernoulli.data.json")
fit_optim <- mod$optimize(data = my_data_file, seed = 123)
fit_optim$summary()

# Run 'variational' to approximate the posterior (default is meanfield ADVI)
fit_vb <- mod$variational(data = stan_data, seed = 123)
fit_vb$summary()
}