Run Stan's variational approximation algorithms

The $variational() method of a CmdStanModel object runs Stan's variational Bayes (ADVI) algorithms.

Details

CmdStan can fit a variational approximation to the posterior. The approximation is a Gaussian in the unconstrained variable space. Stan implements two variational algorithms. The algorithm="meanfield" option uses a fully factorized Gaussian for the approximation. The algorithm="fullrank" option uses a Gaussian with a full-rank covariance matrix for the approximation.

-- CmdStan Interface User's Guide

Usage

$variational(
  data = NULL,
  seed = NULL,
  refresh = NULL,
  init = NULL,
  save_extra_diagnostics = FALSE,
  output_dir = NULL,
  algorithm = NULL,
  iter = NULL,
  grad_samples = NULL,
  elbo_samples = NULL,
  eta = NULL,
  adapt_engaged = NULL,
  adapt_iter = NULL,
  tol_rel_obj = NULL,
  eval_elbo = NULL,
  output_samples = NULL
)

Arguments shared by all fitting methods

The following arguments can be specified for any of the fitting methods (sample, optimize, variational). Arguments left at NULL default to the default used by the installed version of CmdStan.

data: (multiple options) The data to use. One of the following:

A named list of R objects (like for RStan). Internally this list is then written to JSON for CmdStan using write_stan_json().

A path to a data file compatible with CmdStan (JSON or R dump). See the appendices in the CmdStan manual for details on using these formats.

seed: (positive integer) A seed for the (P)RNG to pass to CmdStan.

refresh: (non-negative integer) The number of iterations between printed screen updates.

init: (multiple options) The initialization method:

A real number x>0 initializes randomly between [-x,x] (on the unconstrained parameter space);

0 initializes to 0;

A character vector of paths (one per chain) to JSON or Rdump files. See write_stan_json() to write R objects to JSON files compatible with CmdStan.

save_extra_diagnostics: (logical) Should auxiliary diagnostic information (beyond standard diagnostics) be written to temporary diagnostic CSV files? This argument replaces CmdStan's diagnostic_file argument and the content written to CSV is controlled by the user's CmdStan installation and not CmdStanR (and for some algorithms no content may be written). The default is save_extra_diagnostics=FALSE, which is appropriate for almost every use case (all diagnostics recommended for users to check are always saved, e.g., divergences for HMC). To save the temporary files created when save_extra_diagnostics=TRUE see the $save_diagnostic_files() method.

output_dir: (string) A path to a directory where CmdStan should write its output CSV files. For interactive use this can typically be left at NULL (temporary directory) since CmdStanR makes the CmdStan output (e.g., posterior draws and diagnostics) available in R via methods of the fitted model objects. The behavior of output_dir is as follows:

If NULL (the default) then the CSV files are written to a temporary directory and only saved permanently if the user calls one of the $save_* methods of the fitted model object (e.g., $save_output_files()).

If a path then the files are created in output_dir with names corresponding the defaults used by $save_output_files() (and similar methods like $save_diagnostic_files()).

Arguments unique to the variational method

In addition to the arguments above, the $variational() method also has its own set of arguments. These arguments are described briefly here and in greater detail in the CmdStan manual. Arguments left at NULL default to the default used by the installed version of CmdStan.

algorithm: (string) The algorithm. Either "meanfield" or "fullrank".

iter: (positive integer) The maximum number of iterations.

grad_samples: (positive integer) The number of samples for Monte Carlo estimate of gradients.

elbo_samples: (positive integer) The number of samples for Monte Carlo estimate of ELBO (objective function).

eta: (positive real) The stepsize weighting parameter for adaptive stepsize sequence.

adapt_engaged: (logical) Do warmup adaptation?

adapt_iter: (positive integer) The maximum number of adaptation iterations.

tol_rel_obj: (positive real) Convergence tolerance on the relative norm of the objective.

eval_elbo: (positive integer) Evaluate ELBO every Nth iteration.

output_samples: (positive integer) Number of posterior samples to draw and save.

Value

The $variational() method returns a CmdStanVB object.

See also

Examples

if (FALSE) {
# Set path to cmdstan
# (Note: if you installed CmdStan via install_cmdstan() with default settings
# then setting the path is unnecessary but the default below should still work.
# Otherwise use the `path` argument to specify the location of your
# CmdStan installation.)

set_cmdstan_path(path = NULL)

# Create a CmdStanModel object from a Stan program,
# here using the example model that comes with CmdStan
stan_program <- file.path(cmdstan_path(), "examples/bernoulli/bernoulli.stan")
mod <- cmdstan_model(stan_program)
mod$print()

# data as a named list (like RStan)
stan_data <- list(N = 10, y = c(0,1,0,0,0,0,0,0,0,1))

# run MCMC using the 'sample' method
fit_mcmc <- mod$sample(
  data = stan_data,
  seed = 123,
  num_chains = 2,
  num_cores = 2
)

# Use 'posterior' package for summaries
fit_mcmc$summary()

# Call CmdStan's diagnose and stansummary utilities
fit_mcmc$cmdstan_diagnose()
fit_mcmc$cmdstan_summary()

# For models fit using MCMC, if you like working with RStan's stanfit objects
# then you can create one with rstan::read_stan_csv()
if (require(rstan, quietly = TRUE)) {
  stanfit <- rstan::read_stan_csv(fit_mcmc$output_files())
  print(stanfit)
}

# Run 'optimize' method to get a point estimate (default is Stan's LBFGS algorithm)
# and also demonstrate specifying data as a path to a file instead of a list
my_data_file <- file.path(cmdstan_path(), "examples/bernoulli/bernoulli.data.json")
fit_optim <- mod$optimize(data = my_data_file, seed = 123)
fit_optim$summary()

# Run 'variational' to approximate the posterior (default is meanfield ADVI)
fit_vb <- mod$variational(data = stan_data, seed = 123)
fit_vb$summary()
}