CmdStanModel objects

A CmdStanModel object is an R6 object returned by the cmdstan_model() function. The object stores the path to a Stan program as well as a path to a compiled executable once created, and provides methods for fitting the model. See Details section for available methods.

Details

CmdStanModel objects have the following associated methods:

Method	Description

code	Return Stan program as a string

print	Print readable version of Stan program

compile	Compile Stan program

sample	Run CmdStan's "sample" method, return CmdStanMCMC object

optimize	Run CmdStan's "optimize" method, return CmdStanMLE object

variational	Run CmdStan's "variational" method, return CmdStanVB object

Examples

# \dontrun{
# Set path to cmdstan
# Note: if you installed CmdStan via install_cmdstan() with default settings
# then default below should work. Otherwise use the `path` argument to
# specify the location of your CmdStan installation.

set_cmdstan_path(path = NULL)

#> CmdStan path set to: /Users/jgabry/.cmdstanr/cmdstan-2.20.0

# Create a CmdStan model object from a Stan program,
# here using the example model that comes with CmdStan
stan_program <- file.path(cmdstan_path(), "examples/bernoulli/bernoulli.stan")
mod <- cmdstan_model(stan_program)
mod$print()

#> data { 
#>   int<lower=0> N; 
#>   int<lower=0,upper=1> y[N];
#> } 
#> parameters {
#>   real<lower=0,upper=1> theta;
#> } 
#> model {
#>   theta ~ beta(1,1);
#>   for (n in 1:N) 
#>     y[n] ~ bernoulli(theta);
#> }

# Compile to create executable
mod$compile()

#> Running make \
#>   /Users/jgabry/.cmdstanr/cmdstan-2.20.0/examples/bernoulli/bernoulli
#> make: `/Users/jgabry/.cmdstanr/cmdstan-2.20.0/examples/bernoulli/bernoulli' is up to date.

# Run MCMC (Stan's dynamic HMC/NUTS),
# specifying data as a named list (like RStan)
standata <- list(N = 10, y =c(0,1,0,0,0,0,0,0,0,1))
fit <- mod$sample(data = standata, seed = 123, num_chains = 2)

#> Running ./bernoulli 'id=1' random 'seed=123' data \
#>   'file=/var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T//RtmpwnZmc3/standata-103e0c12326a.data.R' \
#>   output \
#>   'file=/var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T//RtmpwnZmc3/bernoulli-stan-sample-1.csv' \
#>   'method=sample' 'save_warmup=0' 'algorithm=hmc' 'engine=nuts'
#> method = sample (Default)
#>   sample
#>     num_samples = 1000 (Default)
#>     num_warmup = 1000 (Default)
#>     save_warmup = 0 (Default)
#>     thin = 1 (Default)
#>     adapt
#>       engaged = 1 (Default)
#>       gamma = 0.050000000000000003 (Default)
#>       delta = 0.80000000000000004 (Default)
#>       kappa = 0.75 (Default)
#>       t0 = 10 (Default)
#>       init_buffer = 75 (Default)
#>       term_buffer = 50 (Default)
#>       window = 25 (Default)
#>     algorithm = hmc (Default)
#>       hmc
#>         engine = nuts (Default)
#>           nuts
#>             max_depth = 10 (Default)
#>         metric = diag_e (Default)
#>         metric_file =  (Default)
#>         stepsize = 1 (Default)
#>         stepsize_jitter = 0 (Default)
#> id = 1
#> data
#>   file = /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T//RtmpwnZmc3/standata-103e0c12326a.data.R
#> init = 2 (Default)
#> random
#>   seed = 123
#> output
#>   file = /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T//RtmpwnZmc3/bernoulli-stan-sample-1.csv
#>   diagnostic_file =  (Default)
#>   refresh = 100 (Default)
#> 
#> 
#> Gradient evaluation took 1.9e-05 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 0.19 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:    1 / 2000 [  0%]  (Warmup)
#> Iteration:  100 / 2000 [  5%]  (Warmup)
#> Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Iteration:  300 / 2000 [ 15%]  (Warmup)
#> Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Iteration:  500 / 2000 [ 25%]  (Warmup)
#> Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Iteration:  700 / 2000 [ 35%]  (Warmup)
#> Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Iteration:  900 / 2000 [ 45%]  (Warmup)
#> Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Iteration: 1100 / 2000 [ 55%]  (Sampling)
#> Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Iteration: 1300 / 2000 [ 65%]  (Sampling)
#> Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Iteration: 1500 / 2000 [ 75%]  (Sampling)
#> Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Iteration: 1700 / 2000 [ 85%]  (Sampling)
#> Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Iteration: 1900 / 2000 [ 95%]  (Sampling)
#> Iteration: 2000 / 2000 [100%]  (Sampling)
#> 
#>  Elapsed Time: 0.011945 seconds (Warm-up)
#>                0.018921 seconds (Sampling)
#>                0.030866 seconds (Total)
#> 
#> Running ./bernoulli 'id=2' random 'seed=124' data \
#>   'file=/var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T//RtmpwnZmc3/standata-103e0c12326a.data.R' \
#>   output \
#>   'file=/var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T//RtmpwnZmc3/bernoulli-stan-sample-2.csv' \
#>   'method=sample' 'save_warmup=0' 'algorithm=hmc' 'engine=nuts'
#> method = sample (Default)
#>   sample
#>     num_samples = 1000 (Default)
#>     num_warmup = 1000 (Default)
#>     save_warmup = 0 (Default)
#>     thin = 1 (Default)
#>     adapt
#>       engaged = 1 (Default)
#>       gamma = 0.050000000000000003 (Default)
#>       delta = 0.80000000000000004 (Default)
#>       kappa = 0.75 (Default)
#>       t0 = 10 (Default)
#>       init_buffer = 75 (Default)
#>       term_buffer = 50 (Default)
#>       window = 25 (Default)
#>     algorithm = hmc (Default)
#>       hmc
#>         engine = nuts (Default)
#>           nuts
#>             max_depth = 10 (Default)
#>         metric = diag_e (Default)
#>         metric_file =  (Default)
#>         stepsize = 1 (Default)
#>         stepsize_jitter = 0 (Default)
#> id = 2
#> data
#>   file = /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T//RtmpwnZmc3/standata-103e0c12326a.data.R
#> init = 2 (Default)
#> random
#>   seed = 124
#> output
#>   file = /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T//RtmpwnZmc3/bernoulli-stan-sample-2.csv
#>   diagnostic_file =  (Default)
#>   refresh = 100 (Default)
#> 
#> 
#> Gradient evaluation took 1.8e-05 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 0.18 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:    1 / 2000 [  0%]  (Warmup)
#> Iteration:  100 / 2000 [  5%]  (Warmup)
#> Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Iteration:  300 / 2000 [ 15%]  (Warmup)
#> Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Iteration:  500 / 2000 [ 25%]  (Warmup)
#> Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Iteration:  700 / 2000 [ 35%]  (Warmup)
#> Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Iteration:  900 / 2000 [ 45%]  (Warmup)
#> Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Iteration: 1100 / 2000 [ 55%]  (Sampling)
#> Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Iteration: 1300 / 2000 [ 65%]  (Sampling)
#> Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Iteration: 1500 / 2000 [ 75%]  (Sampling)
#> Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Iteration: 1700 / 2000 [ 85%]  (Sampling)
#> Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Iteration: 1900 / 2000 [ 95%]  (Sampling)
#> Iteration: 2000 / 2000 [100%]  (Sampling)
#> 
#>  Elapsed Time: 0.011552 seconds (Warm-up)
#>                0.018995 seconds (Sampling)
#>                0.030547 seconds (Total)
#> 

fit$summary()

#> Running bin/stansummary \
#>   /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T//RtmpwnZmc3/bernoulli-stan-sample-1.csv \
#>   /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T//RtmpwnZmc3/bernoulli-stan-sample-2.csv
#> Inference for Stan model: bernoulli_model
#> 2 chains: each with iter=(1000,1000); warmup=(0,0); thin=(1,1); 2000 iterations saved.
#> 
#> Warmup took (0.012, 0.012) seconds, 0.023 seconds total
#> Sampling took (0.019, 0.019) seconds, 0.038 seconds total
#> 
#>                 Mean     MCSE   StdDev     5%   50%   95%  N_Eff  N_Eff/s    R_hat
#> lp__            -7.3  2.7e-02  7.3e-01   -8.9  -7.0  -6.8    738    19476  1.0e+00
#> accept_stat__   0.92  3.1e-03  1.3e-01   0.64  0.97   1.0   1670    44032  1.0e+00
#> stepsize__      0.92  1.7e-03  1.7e-03   0.92  0.92  0.92    1.0       26  1.4e+12
#> treedepth__      1.3  1.1e-02  4.7e-01    1.0   1.0   2.0   1968    51898  1.0e+00
#> n_leapfrog__     2.4  2.5e-02  1.0e+00    1.0   3.0   3.0   1640    43258  1.0e+00
#> divergent__     0.00  0.0e+00  0.0e+00   0.00  0.00  0.00   1000    26374      nan
#> energy__         7.8  4.0e-02  1.0e+00    6.8   7.5   9.8    647    17058  1.0e+00
#> theta           0.24  4.6e-03  1.2e-01  0.077  0.22  0.47    720    18992  1.0e+00
#> 
#> Samples were drawn using hmc with nuts.
#> For each parameter, N_Eff is a crude measure of effective sample size,
#> and R_hat is the potential scale reduction factor on split chains (at 
#> convergence, R_hat=1).
#> 

# Can also specify data as a path to a file readable by CmdStan
my_data_file <- file.path(cmdstan_path(), "examples/bernoulli/bernoulli.data.R")
fit2 <- mod$sample(data = my_data_file, seed = 123)

#> Running ./bernoulli 'id=1' random 'seed=123' data \
#>   'file=/Users/jgabry/.cmdstanr/cmdstan-2.20.0/examples/bernoulli/bernoulli.data.R' \
#>   output \
#>   'file=/var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T//RtmpwnZmc3/bernoulli-stan-sample-1.csv' \
#>   'method=sample' 'save_warmup=0' 'algorithm=hmc' 'engine=nuts'
#> method = sample (Default)
#>   sample
#>     num_samples = 1000 (Default)
#>     num_warmup = 1000 (Default)
#>     save_warmup = 0 (Default)
#>     thin = 1 (Default)
#>     adapt
#>       engaged = 1 (Default)
#>       gamma = 0.050000000000000003 (Default)
#>       delta = 0.80000000000000004 (Default)
#>       kappa = 0.75 (Default)
#>       t0 = 10 (Default)
#>       init_buffer = 75 (Default)
#>       term_buffer = 50 (Default)
#>       window = 25 (Default)
#>     algorithm = hmc (Default)
#>       hmc
#>         engine = nuts (Default)
#>           nuts
#>             max_depth = 10 (Default)
#>         metric = diag_e (Default)
#>         metric_file =  (Default)
#>         stepsize = 1 (Default)
#>         stepsize_jitter = 0 (Default)
#> id = 1
#> data
#>   file = /Users/jgabry/.cmdstanr/cmdstan-2.20.0/examples/bernoulli/bernoulli.data.R
#> init = 2 (Default)
#> random
#>   seed = 123
#> output
#>   file = /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T//RtmpwnZmc3/bernoulli-stan-sample-1.csv
#>   diagnostic_file =  (Default)
#>   refresh = 100 (Default)
#> 
#> 
#> Gradient evaluation took 1.9e-05 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 0.19 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Iteration:    1 / 2000 [  0%]  (Warmup)
#> Iteration:  100 / 2000 [  5%]  (Warmup)
#> Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Iteration:  300 / 2000 [ 15%]  (Warmup)
#> Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Iteration:  500 / 2000 [ 25%]  (Warmup)
#> Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Iteration:  700 / 2000 [ 35%]  (Warmup)
#> Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Iteration:  900 / 2000 [ 45%]  (Warmup)
#> Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Iteration: 1100 / 2000 [ 55%]  (Sampling)
#> Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Iteration: 1300 / 2000 [ 65%]  (Sampling)
#> Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Iteration: 1500 / 2000 [ 75%]  (Sampling)
#> Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Iteration: 1700 / 2000 [ 85%]  (Sampling)
#> Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Iteration: 1900 / 2000 [ 95%]  (Sampling)
#> Iteration: 2000 / 2000 [100%]  (Sampling)
#> 
#>  Elapsed Time: 0.011623 seconds (Warm-up)
#>                0.018366 seconds (Sampling)
#>                0.029989 seconds (Total)
#> 

fit2$summary()

#> Running bin/stansummary \
#>   /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T//RtmpwnZmc3/bernoulli-stan-sample-1.csv
#> Inference for Stan model: bernoulli_model
#> 1 chains: each with iter=(1000); warmup=(0); thin=(1); 1000 iterations saved.
#> 
#> Warmup took (0.012) seconds, 0.012 seconds total
#> Sampling took (0.018) seconds, 0.018 seconds total
#> 
#>                 Mean     MCSE   StdDev     5%   50%   95%  N_Eff  N_Eff/s    R_hat
#> lp__            -7.4  3.8e-02  7.8e-01   -9.1  -7.1  -6.8    423    23009  1.0e+00
#> accept_stat__   0.91  5.0e-03  1.4e-01   0.61  0.97   1.0    763    41551  1.0e+00
#> stepsize__      0.92  4.2e-15  4.2e-15   0.92  0.92  0.92    1.0       55  1.0e+00
#> treedepth__      1.3  1.5e-02  4.7e-01    1.0   1.0   2.0    971    52891  1.0e+00
#> n_leapfrog__     2.3  3.2e-02  9.9e-01    1.0   3.0   3.0    943    51322  1.0e+00
#> divergent__     0.00  0.0e+00  0.0e+00   0.00  0.00  0.00    500    27224      nan
#> energy__         7.9  5.7e-02  1.1e+00    6.8   7.6  10.0    356    19407  1.0e+00
#> theta           0.24  6.6e-03  1.2e-01  0.072  0.21  0.46    360    19595  1.0e+00
#> 
#> Samples were drawn using hmc with nuts.
#> For each parameter, N_Eff is a crude measure of effective sample size,
#> and R_hat is the potential scale reduction factor on split chains (at 
#> convergence, R_hat=1).
#> 

# If you like working with RStan's stanfit objects then you can
# also create a stanfit object with rstan::read_stan_csv()
if (require(rstan, quietly = TRUE)) {
  stanfit <- rstan::read_stan_csv(fit$output_files())
}

# }