CmdStanModel objects

A CmdStanModel object is an R6 object returned by the cmdstan_model() function. The object stores the path to a Stan program as well as a path to a compiled executable once created, and provides methods for fitting the model. See Details section for available methods.

Details

CmdStanModel objects have the following associated methods:

Method	Description

code	Return Stan program as a string.

print	Print readable version of Stan program.

stan_file	Return the file path to the Stan program.

exe_file	Return the file path to the compiled executable once compiled.

compile	Compile Stan program.

sample	Run CmdStan's "sample" method, return CmdStanMCMC object.

optimize	Run CmdStan's "optimize" method, return CmdStanMLE object.

variational	Run CmdStan's "variational" method, return CmdStanVB object.

See also

Examples

# \dontrun{
# Set path to cmdstan
# (Note: if you installed CmdStan via install_cmdstan() with default settings
# then setting the path is unnecessary but the default below should still work.
# Otherwise use the `path` argument to specify the location of your
# CmdStan installation.)

set_cmdstan_path(path = NULL)

#> CmdStan path set to: /Users/jgabry/.cmdstanr/cmdstan

# Create a CmdStanModel object from a Stan program,
# here using the example model that comes with CmdStan
stan_program <- file.path(cmdstan_path(), "examples/bernoulli/bernoulli.stan")
mod <- cmdstan_model(stan_program)

#> Compiling Stan program...

mod$print()

#> data { 
#>   int<lower=0> N; 
#>   int<lower=0,upper=1> y[N];
#> } 
#> parameters {
#>   real<lower=0,upper=1> theta;
#> } 
#> model {
#>   theta ~ beta(1,1);  // uniform prior on interval 0,1
#>   y ~ bernoulli(theta);
#> }

# data as a named list (like RStan)
stan_data <- list(N = 10, y = c(0,1,0,0,0,0,0,0,0,1))
fit_mcmc <- mod$sample(
  data = stan_data,
  seed = 123,
  num_chains = 2,
  num_cores = 2
)

#> Running MCMC with 2 chain(s) on 2 core(s)...
#> 
#> Running ./bernoulli 'id=1' random 'seed=123' data \
#>   'file=/var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpYuY7Z7/standata-1264f416e3f82.json' \
#>   output \
#>   'file=/var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T//RtmpYuY7Z7/bernoulli-stan-sample-1-1264f2d0b915d.csv' \
#>   'method=sample' 'save_warmup=0' 'algorithm=hmc' 'engine=nuts' adapt \
#>   'engaged=1'
#> Running ./bernoulli 'id=2' random 'seed=124' data \
#>   'file=/var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpYuY7Z7/standata-1264f416e3f82.json' \
#>   output \
#>   'file=/var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T//RtmpYuY7Z7/bernoulli-stan-sample-2-1264f527487c5.csv' \
#>   'method=sample' 'save_warmup=0' 'algorithm=hmc' 'engine=nuts' adapt \
#>   'engaged=1'
#> Chain 1 Iteration:    1 / 2000 [  0%]  (Warmup) 
#> Chain 1 Iteration:  100 / 2000 [  5%]  (Warmup) 
#> Chain 1 Iteration:  200 / 2000 [ 10%]  (Warmup) 
#> Chain 1 Iteration:  300 / 2000 [ 15%]  (Warmup) 
#> Chain 1 Iteration:  400 / 2000 [ 20%]  (Warmup) 
#> Chain 1 Iteration:  500 / 2000 [ 25%]  (Warmup) 
#> Chain 1 Iteration:  600 / 2000 [ 30%]  (Warmup) 
#> Chain 1 Iteration:  700 / 2000 [ 35%]  (Warmup) 
#> Chain 1 Iteration:  800 / 2000 [ 40%]  (Warmup) 
#> Chain 1 Iteration:  900 / 2000 [ 45%]  (Warmup) 
#> Chain 1 Iteration: 1000 / 2000 [ 50%]  (Warmup) 
#> Chain 1 Iteration: 1001 / 2000 [ 50%]  (Sampling) 
#> Chain 1 Iteration: 1100 / 2000 [ 55%]  (Sampling) 
#> Chain 1 Iteration: 1200 / 2000 [ 60%]  (Sampling) 
#> Chain 1 Iteration: 1300 / 2000 [ 65%]  (Sampling) 
#> Chain 1 Iteration: 1400 / 2000 [ 70%]  (Sampling) 
#> Chain 1 Iteration: 1500 / 2000 [ 75%]  (Sampling) 
#> Chain 1 Iteration: 1600 / 2000 [ 80%]  (Sampling) 
#> Chain 1 Iteration: 1700 / 2000 [ 85%]  (Sampling) 
#> Chain 1 Iteration: 1800 / 2000 [ 90%]  (Sampling) 
#> Chain 1 Iteration: 1900 / 2000 [ 95%]  (Sampling) 
#> Chain 1 Iteration: 2000 / 2000 [100%]  (Sampling) 
#> Chain 2 Iteration:    1 / 2000 [  0%]  (Warmup) 
#> Chain 2 Iteration:  100 / 2000 [  5%]  (Warmup) 
#> Chain 2 Iteration:  200 / 2000 [ 10%]  (Warmup) 
#> Chain 2 Iteration:  300 / 2000 [ 15%]  (Warmup) 
#> Chain 2 Iteration:  400 / 2000 [ 20%]  (Warmup) 
#> Chain 2 Iteration:  500 / 2000 [ 25%]  (Warmup) 
#> Chain 2 Iteration:  600 / 2000 [ 30%]  (Warmup) 
#> Chain 2 Iteration:  700 / 2000 [ 35%]  (Warmup) 
#> Chain 2 Iteration:  800 / 2000 [ 40%]  (Warmup) 
#> Chain 2 Iteration:  900 / 2000 [ 45%]  (Warmup) 
#> Chain 2 Iteration: 1000 / 2000 [ 50%]  (Warmup) 
#> Chain 2 Iteration: 1001 / 2000 [ 50%]  (Sampling) 
#> Chain 2 Iteration: 1100 / 2000 [ 55%]  (Sampling) 
#> Chain 2 Iteration: 1200 / 2000 [ 60%]  (Sampling) 
#> Chain 2 Iteration: 1300 / 2000 [ 65%]  (Sampling) 
#> Chain 2 Iteration: 1400 / 2000 [ 70%]  (Sampling) 
#> Chain 2 Iteration: 1500 / 2000 [ 75%]  (Sampling) 
#> Chain 2 Iteration: 1600 / 2000 [ 80%]  (Sampling) 
#> Chain 2 Iteration: 1700 / 2000 [ 85%]  (Sampling) 
#> Chain 2 Iteration: 1800 / 2000 [ 90%]  (Sampling) 
#> Chain 2 Iteration: 1900 / 2000 [ 95%]  (Sampling) 
#> Chain 2 Iteration: 2000 / 2000 [100%]  (Sampling) 
#> Chain 1 finished in 0.2 seconds.
#> Chain 2 finished in 0.2 seconds.
#> 
#> Both chains finished succesfully.
#> Mean chain execution time: 0.2 seconds.
#> Total execution time: 0.2 seconds.

# Call CmdStan's bin/stansummary
fit_mcmc$summary()

#> Running bin/stansummary \
#>   /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T//RtmpYuY7Z7/bernoulli-stan-sample-1-1264f2d0b915d.csv \
#>   /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T//RtmpYuY7Z7/bernoulli-stan-sample-2-1264f527487c5.csv
#> Inference for Stan model: bernoulli_model
#> 2 chains: each with iter=(1000,1000); warmup=(0,0); thin=(1,1); 2000 iterations saved.
#> 
#> Warmup took (0.0077, 0.0077) seconds, 0.015 seconds total
#> Sampling took (0.016, 0.016) seconds, 0.032 seconds total
#> 
#>                 Mean     MCSE  StdDev     5%   50%   95%    N_Eff  N_Eff/s    R_hat
#> lp__            -7.3  3.1e-02    0.78   -8.9  -7.0  -6.8  6.5e+02  2.0e+04  1.0e+00
#> accept_stat__   0.92  3.3e-03    0.12   0.68  0.97   1.0  1.3e+03  4.1e+04  1.0e+00
#> stepsize__      0.95  1.3e-01    0.14   0.81   1.1   1.1  1.0e+00  3.1e+01  4.2e+13
#> treedepth__      1.4  1.2e-02    0.48    1.0   1.0   2.0  1.7e+03  5.1e+04  1.0e+00
#> n_leapfrog__     2.5  2.0e-01     1.3    1.0   3.0   3.0  4.4e+01  1.4e+03  1.0e+00
#> divergent__     0.00      nan    0.00   0.00  0.00  0.00      nan      nan      nan
#> energy__         7.8  4.2e-02     1.0    6.8   7.5   9.8  6.1e+02  1.9e+04  1.0e+00
#> theta           0.24  4.5e-03    0.12  0.077  0.23  0.47  7.4e+02  2.3e+04  1.0e+00
#> 
#> Samples were drawn using hmc with nuts.
#> For each parameter, N_Eff is a crude measure of effective sample size,
#> and R_hat is the potential scale reduction factor on split chains (at 
#> convergence, R_hat=1).
#> 

# Call CmdStan's bin/diagnose
fit_mcmc$diagnose()

#> Running bin/diagnose \
#>   /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T//RtmpYuY7Z7/bernoulli-stan-sample-1-1264f2d0b915d.csv \
#>   /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T//RtmpYuY7Z7/bernoulli-stan-sample-2-1264f527487c5.csv
#> Processing csv files: /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T//RtmpYuY7Z7/bernoulli-stan-sample-1-1264f2d0b915d.csv, /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T//RtmpYuY7Z7/bernoulli-stan-sample-2-1264f527487c5.csv
#> 
#> Checking sampler transitions treedepth.
#> Treedepth satisfactory for all transitions.
#> 
#> Checking sampler transitions for divergences.
#> No divergent transitions found.
#> 
#> Checking E-BFMI - sampler transitions HMC potential energy.
#> E-BFMI satisfactory for all transitions.
#> 
#> Effective sample size satisfactory.
#> 
#> Split R-hat values satisfactory all parameters.
#> 
#> Processing complete, no problems detected.

# For models fit using MCMC, if you like working with RStan's stanfit objects
# then you can create one with rstan::read_stan_csv()
if (require(rstan, quietly = TRUE)) {
  stanfit <- rstan::read_stan_csv(fit_mcmc$output_files())
  print(stanfit)
}

#> rstan (Version 2.19.9, GitRev: 2e1f913d3ca3)

#> For execution on a local, multicore CPU with excess RAM we recommend calling
#> options(mc.cores = parallel::detectCores()).
#> To avoid recompilation of unchanged Stan programs, we recommend calling
#> rstan_options(auto_write = TRUE)

#> Inference for Stan model: bernoulli-stan-sample-1-1264f2d0b915d.
#> 2 chains, each with iter=2000; warmup=1000; thin=1; 
#> post-warmup draws per chain=1000, total post-warmup draws=2000.
#> 
#>        mean se_mean   sd  2.5%   25%   50%   75% 97.5% n_eff Rhat
#> theta  0.24    0.00 0.12  0.06  0.15  0.23  0.32  0.52   736    1
#> lp__  -7.31    0.03 0.78 -9.60 -7.51 -7.04 -6.82 -6.75   650    1
#> 
#> Samples were drawn using NUTS(diag_e) at Sat Nov 09 18:30:48 2019.
#> For each parameter, n_eff is a crude measure of effective sample size,
#> and Rhat is the potential scale reduction factor on split chains (at 
#> convergence, Rhat=1).

# Run optimization method (default is Stan's LBFGS algorithm)
# and also demonstrate specifying data as a path to a file (readable by CmdStan)
my_data_file <- file.path(cmdstan_path(), "examples/bernoulli/bernoulli.data.json")
fit_optim <- mod$optimize(data = my_data_file, seed = 123)

#> Warning: Optimization method is experimental and the structure of returned object may change.

#> method = optimize
#>   optimize
#>     algorithm = lbfgs (Default)
#>       lbfgs
#>         init_alpha = 0.001 (Default)
#>         tol_obj = 9.9999999999999998e-13 (Default)
#>         tol_rel_obj = 10000 (Default)
#>         tol_grad = 1e-08 (Default)
#>         tol_rel_grad = 10000000 (Default)
#>         tol_param = 1e-08 (Default)
#>         history_size = 5 (Default)
#>     iter = 2000 (Default)
#>     save_iterations = 0 (Default)
#> id = 1
#> data
#>   file = /Users/jgabry/.cmdstanr/cmdstan/examples/bernoulli/bernoulli.data.json
#> init = 2 (Default)
#> random
#>   seed = 123
#> output
#>   file = /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T//RtmpYuY7Z7/bernoulli-stan-optimize-1-1264f5ffd0438.csv
#>   diagnostic_file =  (Default)
#>   refresh = 100 (Default)
#> 
#> Initial log joint probability = -9.51104
#>     Iter      log prob        ||dx||      ||grad||       alpha      alpha0  # evals  Notes 
#>        6      -5.00402   0.000103557   2.55661e-07           1           1        9   
#> Optimization terminated normally: 
#>   Convergence detected: relative gradient magnitude is below tolerance

#' Print estimates
fit_optim$summary()

#> Estimates from optimization:

#>    theta     lp__ 
#>  0.20000 -5.00402 

# Run variational Bayes method (default is meanfield ADVI)
fit_vb <- mod$variational(data = stan_data, seed = 123)

#> Warning: Variational inference method is experimental and the structure of returned object may change.

#> method = variational
#>   variational
#>     algorithm = meanfield (Default)
#>       meanfield
#>     iter = 10000 (Default)
#>     grad_samples = 1 (Default)
#>     elbo_samples = 100 (Default)
#>     eta = 1 (Default)
#>     adapt
#>       engaged = 1 (Default)
#>       iter = 50 (Default)
#>     tol_rel_obj = 0.01 (Default)
#>     eval_elbo = 100 (Default)
#>     output_samples = 1000 (Default)
#> id = 1
#> data
#>   file = /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpYuY7Z7/standata-1264f5c2029c3.json
#> init = 2 (Default)
#> random
#>   seed = 123
#> output
#>   file = /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T//RtmpYuY7Z7/bernoulli-stan-variational-1-1264f4395f075.csv
#>   diagnostic_file =  (Default)
#>   refresh = 100 (Default)
#> 
#> ------------------------------------------------------------
#> EXPERIMENTAL ALGORITHM:
#>   This procedure has not been thoroughly tested and may be unstable
#>   or buggy. The interface is subject to change.
#> ------------------------------------------------------------
#> 
#> 
#> 
#> Gradient evaluation took 1.8e-05 seconds
#> 1000 transitions using 10 leapfrog steps per transition would take 0.18 seconds.
#> Adjust your expectations accordingly!
#> 
#> 
#> Begin eta adaptation.
#> Iteration:   1 / 250 [  0%]  (Adaptation)
#> Iteration:  50 / 250 [ 20%]  (Adaptation)
#> Iteration: 100 / 250 [ 40%]  (Adaptation)
#> Iteration: 150 / 250 [ 60%]  (Adaptation)
#> Iteration: 200 / 250 [ 80%]  (Adaptation)
#> Success! Found best value [eta = 1] earlier than expected.
#> 
#> Begin stochastic gradient ascent.
#>   iter             ELBO   delta_ELBO_mean   delta_ELBO_med   notes 
#>    100           -6.258             1.000            1.000
#>    200           -6.475             0.517            1.000
#>    300           -6.228             0.358            0.040
#>    400           -6.220             0.269            0.040
#>    500           -6.379             0.220            0.034
#>    600           -6.195             0.188            0.034
#>    700           -6.262             0.163            0.030
#>    800           -6.345             0.144            0.030
#>    900           -6.201             0.131            0.025
#>   1000           -6.307             0.119            0.025
#>   1100           -6.290             0.020            0.023
#>   1200           -6.238             0.017            0.017
#>   1300           -6.182             0.014            0.013
#>   1400           -6.167             0.014            0.013
#>   1500           -6.219             0.012            0.011
#>   1600           -6.164             0.010            0.009   MEDIAN ELBO CONVERGED
#> 
#> Drawing a sample of size 1000 from the approximate posterior... 
#> COMPLETED.

# Call CmdStan's bin/summary
fit_vb$summary()

#> Running bin/stansummary \
#>   /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T//RtmpYuY7Z7/bernoulli-stan-variational-1-1264f4395f075.csv
#> Warning: non-fatal error reading adapation data
#> Inference for Stan model: bernoulli_model
#> 1 chains: each with iter=(1001); warmup=(0); thin=(0); 1001 iterations saved.
#> 
#> Warmup took (0.00) seconds, 0.00 seconds total
#> Sampling took (0.00) seconds, 0.00 seconds total
#> 
#>            Mean     MCSE  StdDev     5%    50%       95%    N_Eff  N_Eff/s    R_hat
#> lp__       0.00      nan    0.00   0.00   0.00   0.0e+00      nan      nan      nan
#> log_p__    -7.2  2.5e-02    0.72   -8.6   -7.0  -6.8e+00  7.9e+02      inf  1.0e+00
#> log_g__   -0.54  2.9e-02    0.76   -2.1  -0.27  -1.5e-03  6.8e+02      inf  1.0e+00
#> theta      0.26  4.2e-03    0.12  0.091   0.23   4.9e-01  8.2e+02      inf  1.0e+00
#> 
#> Samples were drawn using meanfield with .
#> For each parameter, N_Eff is a crude measure of effective sample size,
#> and R_hat is the potential scale reduction factor on split chains (at 
#> convergence, R_hat=1).
#> 

# }