Create a draws object from a CmdStanR fitted model object

Create a draws object supported by the posterior package. These methods are just wrappers around CmdStanR's $draws() method provided for convenience.

# S3 method for CmdStanMCMC
as_draws(x, ...)

# S3 method for CmdStanMLE
as_draws(x, ...)

# S3 method for CmdStanVB
as_draws(x, ...)

# S3 method for CmdStanGQ
as_draws(x, ...)

Arguments

x
A CmdStanR fitted model object.

...
Optional arguments passed to the $draws() method (e.g., variables, inc_warmup, etc.).

Details

To subset iterations, chains, or draws, use the posterior::subset_draws() method after creating the draws object.

Examples

# \dontrun{
fit <- cmdstanr_example()
as_draws(fit)

#> # A draws_array: 1000 iterations, 4 chains, and 105 variables
#> , , variable = lp__
#> 
#>          chain
#> iteration   1   2   3   4
#>         1 -67 -65 -65 -67
#>         2 -65 -66 -65 -64
#>         3 -67 -68 -65 -67
#>         4 -65 -67 -66 -68
#>         5 -64 -67 -68 -66
#> 
#> , , variable = alpha
#> 
#>          chain
#> iteration    1     2    3    4
#>         1 0.46 0.450 0.50 0.49
#>         2 0.21 0.098 0.30 0.34
#>         3 0.31 0.712 0.40 0.26
#>         4 0.45 0.242 0.39 0.56
#>         5 0.46 0.848 0.28 0.24
#> 
#> , , variable = beta[1]
#> 
#>          chain
#> iteration     1     2     3     4
#>         1 -0.43 -0.55 -0.38 -1.19
#>         2 -0.64 -0.87 -0.91 -0.76
#>         3 -1.16 -0.35 -0.31 -0.16
#>         4 -0.62 -1.15 -0.89 -1.25
#>         5 -0.56 -0.37 -0.69 -1.10
#> 
#> , , variable = beta[2]
#> 
#>          chain
#> iteration     1      2      3      4
#>         1 -0.33 -0.575 -0.167 -0.291
#>         2 -0.51  0.026 -0.341 -0.179
#>         3 -0.36 -0.126 -0.401 -0.464
#>         4 -0.49 -0.319 -0.023 -0.047
#>         5 -0.39 -0.397  0.022 -0.209
#> 
#> # ... with 995 more iterations, and 101 more variables

# posterior's as_draws_*() methods will also work
posterior::as_draws_rvars(fit)

#> # A draws_rvars: 1000 iterations, 4 chains, and 4 variables
#> $lp__: rvar<1000,4>[1] mean ± sd:
#> [1] -66 ± 1.5 
#> 
#> $alpha: rvar<1000,4>[1] mean ± sd:
#> [1] 0.38 ± 0.22 
#> 
#> $beta: rvar<1000,4>[3] mean ± sd:
#> [1] -0.66 ± 0.25  -0.28 ± 0.23   0.69 ± 0.27 
#> 
#> $log_lik: rvar<1000,4>[100] mean ± sd:
#>   [1] -0.517 ± 0.101  -0.401 ± 0.148  -0.501 ± 0.220  -0.446 ± 0.151 
#>   [5] -1.187 ± 0.287  -0.586 ± 0.189  -0.640 ± 0.125  -0.276 ± 0.135 
#>   [9] -0.690 ± 0.166  -0.744 ± 0.235  -0.277 ± 0.125  -0.491 ± 0.234 
#>  [13] -0.660 ± 0.207  -0.364 ± 0.175  -0.279 ± 0.110  -0.274 ± 0.089 
#>  [17] -1.597 ± 0.300  -0.478 ± 0.109  -0.232 ± 0.078  -0.114 ± 0.081 
#>  [21] -0.211 ± 0.090  -0.565 ± 0.147  -0.328 ± 0.140  -0.136 ± 0.068 
#>  [25] -0.454 ± 0.122  -1.519 ± 0.347  -0.304 ± 0.122  -0.445 ± 0.086 
#>  [29] -0.723 ± 0.226  -0.693 ± 0.191  -0.483 ± 0.161  -0.423 ± 0.110 
#>  [33] -0.409 ± 0.128  -0.064 ± 0.052  -0.587 ± 0.182  -0.324 ± 0.133 
#>  [37] -0.701 ± 0.227  -0.310 ± 0.149  -0.180 ± 0.112  -0.680 ± 0.128 
#>  [41] -1.138 ± 0.260  -0.931 ± 0.198  -0.414 ± 0.265  -1.176 ± 0.192 
#>  [45] -0.361 ± 0.119  -0.581 ± 0.130  -0.302 ± 0.128  -0.324 ± 0.085 
#>  [49] -0.319 ± 0.082  -1.286 ± 0.332  -0.289 ± 0.096  -0.833 ± 0.144 
#>  [53] -0.399 ± 0.130  -0.373 ± 0.143  -0.384 ± 0.136  -0.319 ± 0.193 
#>  [57] -0.658 ± 0.119  -0.950 ± 0.352  -1.362 ± 0.345  -0.977 ± 0.162 
#>  [61] -0.542 ± 0.100  -0.868 ± 0.309  -0.116 ± 0.075  -0.904 ± 0.248 
#>  [65] -2.034 ± 0.615  -0.507 ± 0.138  -0.276 ± 0.084  -1.065 ± 0.239 
#>  [69] -0.435 ± 0.086  -0.638 ± 0.236  -0.613 ± 0.206  -0.464 ± 0.171 
#>  [73] -1.490 ± 0.370  -0.948 ± 0.197  -1.133 ± 0.386  -0.375 ± 0.141 
#>  [77] -0.880 ± 0.140  -0.491 ± 0.175  -0.766 ± 0.191  -0.540 ± 0.195 
#>  [81] -0.163 ± 0.101  -0.223 ± 0.139  -0.343 ± 0.084  -0.275 ± 0.094 
#>  [85] -0.131 ± 0.077  -1.130 ± 0.323  -0.823 ± 0.129  -0.782 ± 0.241 
#>  [89] -1.283 ± 0.323  -0.260 ± 0.135  -0.386 ± 0.130  -1.504 ± 0.354 
#>  [93] -0.735 ± 0.217  -0.318 ± 0.090  -0.386 ± 0.112  -1.576 ± 0.295 
#>  [97] -0.431 ± 0.102  -1.052 ± 0.371  -0.695 ± 0.142  -0.392 ± 0.099 
#> 

posterior::as_draws_list(fit)

#> # A draws_list: 1000 iterations, 4 chains, and 105 variables
#> 
#> [chain = 1]
#> $lp__
#>  [1] -67 -65 -67 -65 -64 -67 -65 -66 -66 -65
#> 
#> $alpha
#>  [1] 0.46 0.21 0.31 0.45 0.46 0.27 0.34 0.56 0.28 0.52
#> 
#> $`beta[1]`
#>  [1] -0.43 -0.64 -1.16 -0.62 -0.56 -1.13 -0.66 -0.88 -0.42 -0.91
#> 
#> $`beta[2]`
#>  [1] -0.327 -0.506 -0.364 -0.489 -0.387  0.114 -0.647 -0.539  0.094 -0.208
#> 
#> 
#> [chain = 2]
#> $lp__
#>  [1] -65 -66 -68 -67 -67 -66 -67 -68 -67 -65
#> 
#> $alpha
#>  [1] 0.450 0.098 0.712 0.242 0.848 0.382 0.178 0.062 0.682 0.147
#> 
#> $`beta[1]`
#>  [1] -0.55 -0.87 -0.35 -1.15 -0.37 -0.23 -1.14 -0.71 -0.93 -0.65
#> 
#> $`beta[2]`
#>  [1] -0.575  0.026 -0.126 -0.319 -0.397 -0.109 -0.369 -0.415 -0.210 -0.334
#> 
#> # ... with 990 more iterations, and 2 more chains, and 101 more variables

# }