Create a draws object from a CmdStanR fitted model object

Create a draws object supported by the posterior package. These methods are just wrappers around CmdStanR's $draws() method provided for convenience.

# S3 method for CmdStanMCMC
as_draws(x, ...)

# S3 method for CmdStanMLE
as_draws(x, ...)

# S3 method for CmdStanLaplace
as_draws(x, ...)

# S3 method for CmdStanVB
as_draws(x, ...)

# S3 method for CmdStanGQ
as_draws(x, ...)

# S3 method for CmdStanPathfinder
as_draws(x, ...)

Arguments

x

A CmdStanR fitted model object.

...

Optional arguments passed to the $draws() method (e.g., variables, inc_warmup, etc.).

Details

To subset iterations, chains, or draws, use the posterior::subset_draws() method after creating the draws object.

Examples

# \dontrun{
fit <- cmdstanr_example()
as_draws(fit)
#> # A draws_array: 1000 iterations, 4 chains, and 105 variables
#> , , variable = lp__
#> 
#>          chain
#> iteration   1   2   3   4
#>         1 -66 -68 -66 -65
#>         2 -65 -67 -65 -66
#>         3 -68 -66 -65 -66
#>         4 -68 -65 -65 -67
#>         5 -67 -64 -65 -67
#> 
#> , , variable = alpha
#> 
#>          chain
#> iteration    1      2    3    4
#>         1 0.44  0.414 0.71 0.46
#>         2 0.30 -0.045 0.44 0.17
#>         3 0.65  0.397 0.14 0.55
#>         4 0.29  0.644 0.47 0.54
#>         5 0.53  0.435 0.48 0.20
#> 
#> , , variable = beta[1]
#> 
#>          chain
#> iteration     1     2     3     4
#>         1 -0.91 -0.17 -0.50 -0.59
#>         2 -0.57 -0.97 -0.78 -0.65
#>         3 -0.90 -0.28 -0.38 -0.43
#>         4 -1.20 -0.58 -0.92 -1.04
#>         5 -0.59 -0.74 -0.70 -0.98
#> 
#> , , variable = beta[2]
#> 
#>          chain
#> iteration      1      2     3      4
#>         1  0.088 -0.192 -0.43 -0.621
#>         2 -0.515 -0.088 -0.35 -0.591
#>         3 -0.031 -0.444 -0.11  0.017
#>         4 -0.653 -0.229 -0.37  0.066
#>         5 -0.345 -0.277 -0.13 -0.678
#> 
#> # ... with 995 more iterations, and 101 more variables

# posterior's as_draws_*() methods will also work
posterior::as_draws_rvars(fit)
#> # A draws_rvars: 1000 iterations, 4 chains, and 4 variables
#> $lp__: rvar<1000,4>[1] mean ± sd:
#> [1] -66 ± 1.5 
#> 
#> $alpha: rvar<1000,4>[1] mean ± sd:
#> [1] 0.38 ± 0.22 
#> 
#> $beta: rvar<1000,4>[3] mean ± sd:
#> [1] -0.67 ± 0.26  -0.27 ± 0.23   0.69 ± 0.27 
#> 
#> $log_lik: rvar<1000,4>[100] mean ± sd:
#>   [1] -0.517 ± 0.098  -0.401 ± 0.152  -0.501 ± 0.223  -0.448 ± 0.154 
#>   [5] -1.185 ± 0.289  -0.590 ± 0.195  -0.640 ± 0.126  -0.278 ± 0.135 
#>   [9] -0.693 ± 0.170  -0.742 ± 0.238  -0.279 ± 0.127  -0.494 ± 0.244 
#>  [13] -0.659 ± 0.215  -0.363 ± 0.176  -0.279 ± 0.108  -0.275 ± 0.087 
#>  [17] -1.593 ± 0.288  -0.479 ± 0.107  -0.232 ± 0.076  -0.113 ± 0.080 
#>  [21] -0.211 ± 0.089  -0.568 ± 0.149  -0.330 ± 0.140  -0.136 ± 0.067 
#>  [25] -0.454 ± 0.123  -1.517 ± 0.342  -0.306 ± 0.123  -0.445 ± 0.083 
#>  [29] -0.725 ± 0.232  -0.696 ± 0.191  -0.487 ± 0.164  -0.423 ± 0.109 
#>  [33] -0.409 ± 0.130  -0.064 ± 0.052  -0.587 ± 0.189  -0.325 ± 0.132 
#>  [37] -0.700 ± 0.231  -0.312 ± 0.151  -0.180 ± 0.111  -0.681 ± 0.128 
#>  [41] -1.133 ± 0.258  -0.932 ± 0.200  -0.416 ± 0.265  -1.175 ± 0.185 
#>  [45] -0.360 ± 0.118  -0.580 ± 0.131  -0.303 ± 0.131  -0.324 ± 0.082 
#>  [49] -0.319 ± 0.079  -1.285 ± 0.333  -0.289 ± 0.094  -0.832 ± 0.142 
#>  [53] -0.401 ± 0.131  -0.373 ± 0.143  -0.384 ± 0.139  -0.321 ± 0.192 
#>  [57] -0.658 ± 0.119  -0.951 ± 0.363  -1.364 ± 0.348  -0.976 ± 0.158 
#>  [61] -0.542 ± 0.097  -0.871 ± 0.318  -0.116 ± 0.075  -0.901 ± 0.250 
#>  [65] -2.026 ± 0.610  -0.508 ± 0.138  -0.276 ± 0.082  -1.061 ± 0.237 
#>  [69] -0.435 ± 0.083  -0.641 ± 0.235  -0.613 ± 0.214  -0.464 ± 0.177 
#>  [73] -1.489 ± 0.374  -0.948 ± 0.198  -1.136 ± 0.394  -0.374 ± 0.140 
#>  [77] -0.878 ± 0.139  -0.491 ± 0.174  -0.765 ± 0.194  -0.539 ± 0.200 
#>  [81] -0.163 ± 0.102  -0.223 ± 0.145  -0.343 ± 0.080  -0.275 ± 0.092 
#>  [85] -0.130 ± 0.075  -1.132 ± 0.322  -0.822 ± 0.126  -0.780 ± 0.249 
#>  [89] -1.283 ± 0.323  -0.261 ± 0.137  -0.387 ± 0.133  -1.501 ± 0.352 
#>  [93] -0.737 ± 0.221  -0.318 ± 0.087  -0.388 ± 0.112  -1.573 ± 0.284 
#>  [97] -0.432 ± 0.100  -1.053 ± 0.383  -0.693 ± 0.143  -0.392 ± 0.096 
#> 
posterior::as_draws_list(fit)
#> # A draws_list: 1000 iterations, 4 chains, and 105 variables
#> 
#> [chain = 1]
#> $lp__
#>  [1] -66 -65 -68 -68 -67 -66 -65 -66 -67 -66
#> 
#> $alpha
#>  [1] 0.44 0.30 0.65 0.29 0.53 0.53 0.47 0.38 0.32 0.51
#> 
#> $`beta[1]`
#>  [1] -0.91 -0.57 -0.90 -1.20 -0.59 -0.51 -0.36 -0.19 -1.17 -1.02
#> 
#> $`beta[2]`
#>  [1]  0.088 -0.515 -0.031 -0.653 -0.345 -0.646 -0.154 -0.488 -0.013  0.034
#> 
#> 
#> [chain = 2]
#> $lp__
#>  [1] -68 -67 -66 -65 -64 -65 -65 -64 -65 -65
#> 
#> $alpha
#>  [1]  0.414 -0.045  0.397  0.644  0.435  0.266  0.415  0.369  0.229  0.185
#> 
#> $`beta[1]`
#>  [1] -0.17 -0.97 -0.28 -0.58 -0.74 -0.48 -0.83 -0.72 -0.44 -0.58
#> 
#> $`beta[2]`
#>  [1] -0.192 -0.088 -0.444 -0.229 -0.277 -0.314  0.046 -0.240 -0.080 -0.090
#> 
#> # ... with 990 more iterations, and 2 more chains, and 101 more variables
# }