Fit models for use in examples

Fit models for use in examples

cmdstanr_example(
  example = c("logistic", "schools", "schools_ncp"),
  method = c("sample", "optimize", "variational"),
  ...,
  quiet = TRUE
)

print_example_program(example = c("logistic", "schools", "schools_ncp"))

Arguments

example
Name of the example. The currently available examples are:

"logistic": logistic regression with intercept and 3 predictors.

"schools": the so-called "eight schools" model, a hierarchical meta-analysis. Fitting this model will result in warnings about divergences.

"schools_ncp": non-centered parameterization eight schools model that fixes the problem with divergences.

To print the Stan code for a given example use print_example_program().

method
Which fitting method should be used? The default is the "sample" method (MCMC).

...
Arguments passed to the chosen method.

quiet
If TRUE (the default) then fitting the model is wrapped in utils::capture.output().

Value

The fitted model object returned by the selected method.

Examples

# \dontrun{
print_example_program("logistic")

#> data {
#>   int<lower=0> N;
#>   int<lower=0> K;
#>   int<lower=0,upper=1> y[N];
#>   matrix[N, K] X;
#> }
#> parameters {
#>   real alpha;
#>   vector[K] beta;
#> }
#> model {
#>   target += normal_lpdf(alpha | 0, 1);
#>   target += normal_lpdf(beta | 0, 1);
#>   target += bernoulli_logit_glm_lpmf(y | X, alpha, beta);
#> }

fit_logistic_mcmc <- cmdstanr_example("logistic", chains = 2)

#> Compiling Stan program...

fit_logistic_mcmc$summary()

#> # A tibble: 5 x 10
#>   variable    mean  median    sd   mad       q5      q95  rhat ess_bulk ess_tail
#>   <chr>      <dbl>   <dbl> <dbl> <dbl>    <dbl>    <dbl> <dbl>    <dbl>    <dbl>
#> 1 lp__     -65.9   -65.6   1.43  1.23  -68.7    -64.3    0.999    1012.    1162.
#> 2 alpha      0.380   0.372 0.215 0.225   0.0254   0.730  1.00     2402.    1749.
#> 3 beta[1]   -0.657  -0.655 0.240 0.240  -1.06    -0.263  1.00     2282.    1520.
#> 4 beta[2]   -0.276  -0.267 0.231 0.219  -0.673    0.0950 0.999    2160.    1525.
#> 5 beta[3]    0.677   0.663 0.273 0.268   0.254    1.14   0.999    2190.    1515.

fit_logistic_optim <- cmdstanr_example("logistic", method = "optimize")

#> Model executable is up to date!

fit_logistic_optim$summary()

#> # A tibble: 5 x 2
#>   variable estimate
#>   <chr>       <dbl>
#> 1 lp__      -63.9  
#> 2 alpha       0.364
#> 3 beta[1]    -0.632
#> 4 beta[2]    -0.259
#> 5 beta[3]     0.648

fit_logistic_vb <- cmdstanr_example("logistic", method = "variational")

#> Model executable is up to date!

fit_logistic_vb$summary()

#> # A tibble: 6 x 7
#>   variable       mean  median    sd   mad      q5      q95
#>   <chr>         <dbl>   <dbl> <dbl> <dbl>   <dbl>    <dbl>
#> 1 lp__        -66.2   -65.9   1.62  1.39  -69.3   -64.4   
#> 2 lp_approx__  -1.98   -1.66  1.37  1.22   -4.56   -0.352 
#> 3 alpha         0.237   0.239 0.219 0.223  -0.124   0.617 
#> 4 beta[1]      -0.611  -0.598 0.270 0.266  -1.07   -0.183 
#> 5 beta[2]      -0.276  -0.276 0.220 0.222  -0.641   0.0859
#> 6 beta[3]       0.706   0.709 0.231 0.231   0.329   1.09  

print_example_program("schools")

#> data {
#>   int<lower=1> J;
#>   vector<lower=0>[J] sigma;
#>   vector[J] y;
#> }
#> parameters {
#>   real mu;
#>   real<lower=0> tau;
#>   vector[J] theta;
#> }
#> model {
#>   target += normal_lpdf(tau | 0, 10);
#>   target += normal_lpdf(mu | 0, 10);
#>   target += normal_lpdf(theta | mu, tau);
#>   target += normal_lpdf(y | theta, sigma);
#> }

fit_schools_mcmc <- cmdstanr_example("schools")

#> Compiling Stan program...

#> Chain 2 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:

#> Chain 2 Exception: normal_lpdf: Scale parameter is 0, but must be > 0! (in '/var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpXEch8A/model-6aac67d946dd.stan', line 14, column 2 to column 41)

#> Chain 2 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,

#> Chain 2 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.

#> Chain 2 

#> 330 of 4000 (8.0%) transitions ended with a divergence.
#> These divergent transitions indicate that HMC is not fully able to explore the posterior distribution.
#> Try increasing adapt delta closer to 1.
#> If this doesn't remove all divergences, try to reparameterize the model.

fit_schools_mcmc$summary()

#> # A tibble: 11 x 10
#>    variable   mean median    sd   mad       q5   q95  rhat ess_bulk ess_tail
#>    <chr>     <dbl>  <dbl> <dbl> <dbl>    <dbl> <dbl> <dbl>    <dbl>    <dbl>
#>  1 lp__     -54.4  -55.6   7.24  9.10 -65.3    -45.3  1.41     8.51    26.6 
#>  2 mu         6.72   7.44  3.63  2.73   0.0174  12.2  1.46   402.     959.  
#>  3 tau        3.95   2.95  3.54  3.43   0.634   11.0  1.41     8.51     4.81
#>  4 theta[1]   8.73   7.52  5.87  3.98   0.0713  19.8  1.47   791.    1023.  
#>  5 theta[2]   7.28   8.17  4.86  3.69  -0.927   15.2  1.45   575.    1443.  
#>  6 theta[3]   6.00   7.01  5.54  3.82  -3.84    14.4  1.46   836.    1178.  
#>  7 theta[4]   6.48   6.02  5.04  3.41  -2.09    15.0  1.47   987.    1167.  
#>  8 theta[5]   5.28   7.01  4.93  3.60  -3.94    11.9  1.29   199.    1316.  
#>  9 theta[6]   6.09   7.57  5.12  3.75  -3.40    13.6  1.47   438.    1594.  
#> 10 theta[7]   8.79   8.43  5.19  3.70   0.722   18.5  1.46   783.     984.  
#> 11 theta[8]   7.05   7.08  5.64  3.72  -2.32    16.2  1.49  1060.    1166.  

print_example_program("schools_ncp")

#> data {
#>   int<lower=1> J;
#>   vector<lower=0>[J] sigma;
#>   vector[J] y;
#> }
#> parameters {
#>   real mu;
#>   real<lower=0> tau;
#>   vector[J] theta_raw;
#> }
#> transformed parameters {
#>   vector[J] theta = mu + tau * theta_raw;
#> }
#> model {
#>   target += normal_lpdf(tau | 0, 10);
#>   target += normal_lpdf(mu | 0, 10);
#>   target += normal_lpdf(theta_raw | 0, 1);
#>   target += normal_lpdf(y | theta, sigma);
#> }

fit_schools_ncp_mcmc <- cmdstanr_example("schools_ncp")

#> Compiling Stan program...

fit_schools_ncp_mcmc$summary()

#> # A tibble: 19 x 10
#>    variable     mean   median    sd   mad       q5    q95  rhat ess_bulk
#>    <chr>       <dbl>    <dbl> <dbl> <dbl>    <dbl>  <dbl> <dbl>    <dbl>
#>  1 lp__     -46.9    -46.6    2.36  2.33  -5.10e+1 -43.4   1.00    1655.
#>  2 mu         6.38     6.40   4.11  3.99  -3.11e-1  13.1   1.00    4083.
#>  3 tau        4.76     3.95   3.70  3.46   3.78e-1  11.9   1.00    2091.
#>  4 theta_r…   0.350    0.356  0.969 0.983 -1.24e+0   1.95  1.00    4263.
#>  5 theta_r…   0.0546   0.0697 0.906 0.911 -1.44e+0   1.52  1.00    4765.
#>  6 theta_r…  -0.125   -0.140  0.935 0.932 -1.65e+0   1.43  1.00    4984.
#>  7 theta_r…   0.0531   0.0546 0.909 0.868 -1.48e+0   1.56  1.00    4933.
#>  8 theta_r…  -0.258   -0.263  0.922 0.916 -1.73e+0   1.29  1.00    4527.
#>  9 theta_r…  -0.180   -0.168  0.929 0.933 -1.72e+0   1.35  1.00    4169.
#> 10 theta_r…   0.360    0.373  0.913 0.919 -1.14e+0   1.80  1.00    3777.
#> 11 theta_r…   0.0844   0.0725 0.965 0.967 -1.49e+0   1.67  1.01    4741.
#> 12 theta[1]   8.84     8.16   6.86  5.82  -9.87e-1  21.4   1.00    4232.
#> 13 theta[2]   6.76     6.62   5.57  5.15  -2.24e+0  16.1   1.00    5451.
#> 14 theta[3]   5.40     5.81   6.29  5.50  -5.60e+0  14.7   1.00    4297.
#> 15 theta[4]   6.66     6.66   5.72  5.10  -2.51e+0  16.1   1.00    5318.
#> 16 theta[5]   4.70     5.16   5.70  5.24  -5.22e+0  13.4   1.00    4988.
#> 17 theta[6]   5.33     5.49   5.86  5.38  -4.83e+0  14.4   1.00    4727.
#> 18 theta[7]   8.70     8.19   5.90  5.35   9.55e-3  19.3   1.00    4299.
#> 19 theta[8]   6.98     6.87   6.56  5.64  -3.45e+0  17.5   1.00    4798.
#> # … with 1 more variable: ess_tail <dbl>

# optimization fails for hierarchical model
cmdstanr_example("schools", "optimize", quiet = FALSE)

#> Model executable is up to date!

#> method = optimize 
#>   optimize 
#>     algorithm = lbfgs (Default) 
#>       lbfgs 
#>         init_alpha = 0.001 (Default) 
#>         tol_obj = 9.9999999999999998e-13 (Default) 
#>         tol_rel_obj = 10000 (Default) 
#>         tol_grad = 1e-08 (Default) 
#>         tol_rel_grad = 10000000 (Default) 
#>         tol_param = 1e-08 (Default) 
#>         history_size = 5 (Default) 
#>     iter = 2000 (Default) 
#>     save_iterations = 0 (Default) 
#> id = 1 
#> data 
#>   file = /private/var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpWfRZTx/temp_libpath5a083f451490/cmdstanr/schools.data.json 
#> init = 2 (Default) 
#> random 
#>   seed = 1245460385 
#> output 
#>   file = /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpXEch8A/schools-202006301358-1-da537b.csv 
#>   diagnostic_file =  (Default) 
#>   refresh = 100 (Default) 
#>  
#> Initial log joint probability = -51.7356 
#>     Iter      log prob        ||dx||      ||grad||       alpha      alpha0  # evals  Notes  
#>       99       126.564      0.164463   2.53023e+09      0.6979      0.6979      162    
#>     Iter      log prob        ||dx||      ||grad||       alpha      alpha0  # evals  Notes  
#>      179       243.381    0.00904858        8.0032       1e-12       0.001      396  LS failed, Hessian reset  
#> Optimization terminated with error:  
#>   Line search failed to achieve a sufficient decrease, no more progress can be made 

#> Warning: Fitting finished unexpectedly!

#> Error in if (metadata$method == "sample") {    if (!is.null(warmup_draws)) {        posterior::variables(warmup_draws) <- repaired_variables    }    if (!is.null(post_warmup_draws)) {        posterior::variables(post_warmup_draws) <- repaired_variables    }    list(metadata = metadata, inv_metric = inv_metric, step_size = step_size,         warmup_draws = warmup_draws, post_warmup_draws = post_warmup_draws,         warmup_sampler_diagnostics = warmup_sampler_diagnostics_draws,         post_warmup_sampler_diagnostics = post_warmup_sampler_diagnostics_draws)} else if (metadata$method == "variational") {    metadata$model_params <- metadata$model_params[metadata$model_params !=         "lp__"]    metadata$model_params <- gsub("log_p__", "lp__", metadata$model_params)    metadata$model_params <- gsub("log_g__", "lp_approx__", metadata$model_params)    repaired_variables <- repaired_variables[repaired_variables !=         "lp__"]    repaired_variables <- gsub("log_p__", "lp__", repaired_variables)    repaired_variables <- gsub("log_g__", "lp_approx__", repaired_variables)    if (!is.null(variational_draws)) {        posterior::variables(variational_draws) <- repaired_variables    }    list(metadata = metadata, draws = variational_draws)} else if (metadata$method == "optimize") {    if (!is.null(point_estimates)) {        posterior::variables(point_estimates) <- repaired_variables    }    list(metadata = metadata, point_estimates = point_estimates)} else if (metadata$method == "generate_quantities") {    if (!is.null(generated_quantities)) {        posterior::variables(generated_quantities) <- repaired_variables    }    list(metadata = metadata, generated_quantities = generated_quantities)}: argument is of length zero

# }