Fit models for use in examples

Fit models for use in examples

cmdstanr_example(
  example = c("logistic", "schools", "schools_ncp"),
  method = c("sample", "optimize", "laplace", "variational", "pathfinder", "diagnose"),
  ...,
  quiet = TRUE,
  force_recompile = getOption("cmdstanr_force_recompile", default = FALSE)
)

print_example_program(example = c("logistic", "schools", "schools_ncp"))

Arguments

example

(string) The name of the example. The currently available examples are

"logistic": logistic regression with intercept and 3 predictors.

"schools": the so-called "eight schools" model, a hierarchical meta-analysis. Fitting this model will result in warnings about divergences.

"schools_ncp": non-centered parameterization of the "eight schools" model that fixes the problem with divergences.

To print the Stan code for a given example use print_example_program(example).

method

(string) Which fitting method should be used? The default is the "sample" method (MCMC).

...

Arguments passed to the chosen method. See the help pages for the individual methods for details.

quiet

(logical) If TRUE (the default) then fitting the model is wrapped in utils::capture.output().

force_recompile

Passed to the $compile() method.

Value

The fitted model object returned by the selected method.

Examples

# \dontrun{
print_example_program("logistic")
#> data {
#>   int<lower=0> N;
#>   int<lower=0> K;
#>   array[N] int<lower=0, upper=1> y;
#>   matrix[N, K] X;
#> }
#> parameters {
#>   real alpha;
#>   vector[K] beta;
#> }
#> model {
#>   target += normal_lpdf(alpha | 0, 1);
#>   target += normal_lpdf(beta | 0, 1);
#>   target += bernoulli_logit_glm_lpmf(y | X, alpha, beta);
#> }
#> generated quantities {
#>   vector[N] log_lik;
#>   for (n in 1 : N) {
#>     log_lik[n] = bernoulli_logit_lpmf(y[n] | alpha + X[n] * beta);
#>   }
#> }
fit_logistic_mcmc <- cmdstanr_example("logistic", chains = 2)
fit_logistic_mcmc$summary()
#> # A tibble: 105 × 10
#>    variable      mean  median    sd   mad       q5      q95  rhat ess_bulk
#>    <chr>        <dbl>   <dbl> <dbl> <dbl>    <dbl>    <dbl> <dbl>    <dbl>
#>  1 lp__       -66.1   -65.7   1.53  1.26  -69.1    -64.3    1.00     1090.
#>  2 alpha        0.373   0.371 0.223 0.215   0.0169   0.729  1.00     1944.
#>  3 beta[1]     -0.678  -0.676 0.260 0.261  -1.11    -0.265  1.00     1936.
#>  4 beta[2]     -0.276  -0.274 0.234 0.239  -0.674    0.0978 1.00     1841.
#>  5 beta[3]      0.684   0.681 0.268 0.270   0.270    1.13   1.00     2070.
#>  6 log_lik[1]  -0.518  -0.509 0.102 0.101  -0.694   -0.366  0.999    1983.
#>  7 log_lik[2]  -0.398  -0.380 0.149 0.143  -0.665   -0.194  1.00     2108.
#>  8 log_lik[3]  -0.501  -0.461 0.228 0.210  -0.939   -0.204  1.00     1995.
#>  9 log_lik[4]  -0.446  -0.428 0.157 0.148  -0.740   -0.226  1.00     1806.
#> 10 log_lik[5]  -1.18   -1.15  0.285 0.286  -1.67    -0.754  1.00     2253.
#> # ℹ 95 more rows
#> # ℹ 1 more variable: ess_tail <dbl>

fit_logistic_optim <- cmdstanr_example("logistic", method = "optimize")
fit_logistic_optim$summary()
#> # A tibble: 105 × 2
#>    variable   estimate
#>    <chr>         <dbl>
#>  1 lp__        -63.9  
#>  2 alpha         0.364
#>  3 beta[1]      -0.632
#>  4 beta[2]      -0.259
#>  5 beta[3]       0.649
#>  6 log_lik[1]   -0.515
#>  7 log_lik[2]   -0.394
#>  8 log_lik[3]   -0.469
#>  9 log_lik[4]   -0.442
#> 10 log_lik[5]   -1.14 
#> # ℹ 95 more rows

fit_logistic_vb <- cmdstanr_example("logistic", method = "variational")
fit_logistic_vb$summary()
#> # A tibble: 106 × 7
#>    variable       mean  median    sd   mad      q5      q95
#>    <chr>         <dbl>   <dbl> <dbl> <dbl>   <dbl>    <dbl>
#>  1 lp__        -66.5   -66.1   1.98  1.74  -70.5   -64.4   
#>  2 lp_approx__  -2.01   -1.63  1.53  1.27   -5.07   -0.341 
#>  3 alpha         0.305   0.311 0.248 0.234  -0.116   0.716 
#>  4 beta[1]      -0.623  -0.633 0.230 0.239  -0.987  -0.255 
#>  5 beta[2]      -0.377  -0.369 0.263 0.258  -0.821   0.0339
#>  6 beta[3]       0.681   0.671 0.291 0.288   0.215   1.16  
#>  7 log_lik[1]   -0.565  -0.558 0.118 0.108  -0.766  -0.383 
#>  8 log_lik[2]   -0.412  -0.388 0.172 0.156  -0.740  -0.186 
#>  9 log_lik[3]   -0.641  -0.589 0.294 0.280  -1.22   -0.265 
#> 10 log_lik[4]   -0.389  -0.367 0.145 0.134  -0.662  -0.190 
#> # ℹ 96 more rows

print_example_program("schools")
#> data {
#>   int<lower=1> J;
#>   vector<lower=0>[J] sigma;
#>   vector[J] y;
#> }
#> parameters {
#>   real mu;
#>   real<lower=0> tau;
#>   vector[J] theta;
#> }
#> model {
#>   target += normal_lpdf(tau | 0, 10);
#>   target += normal_lpdf(mu | 0, 10);
#>   target += normal_lpdf(theta | mu, tau);
#>   target += normal_lpdf(y | theta, sigma);
#> }
fit_schools_mcmc <- cmdstanr_example("schools")
#> Warning: 91 of 4000 (2.0%) transitions ended with a divergence.
#> See https://mc-stan.org/misc/warnings for details.
fit_schools_mcmc$summary()
#> # A tibble: 11 × 10
#>    variable   mean median    sd   mad      q5   q95  rhat ess_bulk ess_tail
#>    <chr>     <dbl>  <dbl> <dbl> <dbl>   <dbl> <dbl> <dbl>    <dbl>    <dbl>
#>  1 lp__     -57.8  -58.2   5.50  5.71 -66.3   -48.3  1.04     151.     248.
#>  2 mu         6.47   6.50  4.26  4.17  -1.05   13.3  1.01     464.     141.
#>  3 tau        5.22   4.49  3.59  3.40   0.984  12.2  1.04     115.     122.
#>  4 theta[1]   9.21   8.56  7.05  6.14  -1.40   21.8  1.01     427.     239.
#>  5 theta[2]   6.77   6.75  5.62  5.20  -2.79   15.9  1.01     718.     443.
#>  6 theta[3]   5.35   5.83  6.82  5.87  -6.13   15.5  1.01    1103.    1908.
#>  7 theta[4]   6.84   6.85  5.86  5.44  -2.42   16.6  1.01     923.    2076.
#>  8 theta[5]   4.64   4.91  5.62  5.40  -4.88   13.4  1.01     775.    1603.
#>  9 theta[6]   5.51   5.70  6.05  5.48  -4.39   15.1  1.01     639.    1878.
#> 10 theta[7]   8.94   8.42  6.14  5.47  -0.805  19.6  1.02     272.     164.
#> 11 theta[8]   7.08   6.94  6.75  5.70  -3.52   18.0  1.01     969.    1764.

print_example_program("schools_ncp")
#> data {
#>   int<lower=1> J;
#>   vector<lower=0>[J] sigma;
#>   vector[J] y;
#> }
#> parameters {
#>   real mu;
#>   real<lower=0> tau;
#>   vector[J] theta_raw;
#> }
#> transformed parameters {
#>   vector[J] theta = mu + tau * theta_raw;
#> }
#> model {
#>   target += normal_lpdf(tau | 0, 10);
#>   target += normal_lpdf(mu | 0, 10);
#>   target += normal_lpdf(theta_raw | 0, 1);
#>   target += normal_lpdf(y | theta, sigma);
#> }
fit_schools_ncp_mcmc <- cmdstanr_example("schools_ncp")
fit_schools_ncp_mcmc$summary()
#> # A tibble: 19 × 10
#>    variable          mean    median    sd   mad       q5    q95  rhat ess_bulk
#>    <chr>            <dbl>     <dbl> <dbl> <dbl>    <dbl>  <dbl> <dbl>    <dbl>
#>  1 lp__         -46.9     -46.6     2.47  2.39  -51.5    -43.5   1.00    1443.
#>  2 mu             6.50      6.57    4.25  4.13   -0.446   13.4   1.00    2941.
#>  3 tau            4.78      3.91    3.74  3.46    0.374   12.1   1.00    1743.
#>  4 theta_raw[1]   0.372     0.381   0.989 0.951  -1.28     1.97  1.00    3537.
#>  5 theta_raw[2]   0.0565    0.0620  0.915 0.928  -1.45     1.53  1.00    4450.
#>  6 theta_raw[3]  -0.148    -0.163   0.938 0.932  -1.65     1.40  1.00    3407.
#>  7 theta_raw[4]   0.00702   0.00718 0.937 0.912  -1.55     1.53  1.00    3544.
#>  8 theta_raw[5]  -0.264    -0.271   0.896 0.870  -1.70     1.23  1.00    3495.
#>  9 theta_raw[6]  -0.148    -0.157   0.950 0.938  -1.68     1.43  1.00    4517.
#> 10 theta_raw[7]   0.351     0.374   0.924 0.876  -1.23     1.82  1.00    3596.
#> 11 theta_raw[8]   0.0685    0.0540  0.960 0.970  -1.52     1.64  1.00    3657.
#> 12 theta[1]       8.96      8.30    6.79  5.56   -0.975   21.5   1.00    3374.
#> 13 theta[2]       6.90      6.72    5.56  5.25   -2.07    16.0   1.00    4663.
#> 14 theta[3]       5.58      5.91    6.35  5.42   -5.10    15.2   1.00    3392.
#> 15 theta[4]       6.67      6.72    6.02  5.40   -2.96    16.2   1.00    3971.
#> 16 theta[5]       4.81      5.19    5.50  5.21   -4.81    12.9   1.00    4119.
#> 17 theta[6]       5.55      5.86    5.93  5.29   -4.38    14.7   1.00    4010.
#> 18 theta[7]       8.73      8.24    5.88  5.28   -0.0503  19.3   1.00    3955.
#> 19 theta[8]       7.01      6.84    6.39  5.49   -3.28    17.9   1.00    3861.
#> # ℹ 1 more variable: ess_tail <dbl>

# optimization fails for hierarchical model
cmdstanr_example("schools", "optimize", quiet = FALSE)
#> Initial log joint probability = -319.937 
#>     Iter      log prob        ||dx||      ||grad||       alpha      alpha0  # evals  Notes  
#>       99       120.913      0.248985   3.76333e+09           1           1      183    
#>     Iter      log prob        ||dx||      ||grad||       alpha      alpha0  # evals  Notes  
#>      180       242.851     0.0892321   2.24452e+16       1e-12       0.001      378  LS failed, Hessian reset  
#>   Line search failed to achieve a sufficient decrease, no more progress can be made 
#> Chain 1 Optimization terminated with error: 
#> Warning: Fitting finished unexpectedly! Use the $output() method for more information.
#> Error: Fitting failed. Unable to print.
# }