Fit models for use in examples

Fit models for use in examples

cmdstanr_example(
  example = c("logistic", "schools", "schools_ncp"),
  method = c("sample", "optimize", "variational"),
  ...,
  quiet = TRUE
)

print_example_program(example = c("logistic", "schools", "schools_ncp"))

Arguments

example
Name of the example. The currently available examples are:

"logistic": logistic regression with intercept and 3 predictors.

"schools": the so-called "eight schools" model, a hierarchical meta-analysis. Fitting this model will result in warnings about divergences.

"schools_ncp": non-centered parameterization eight schools model that fixes the problem with divergences.

To print the Stan code for a given example use print_example_program().

method
Which fitting method should be used? The default is the "sample" method (MCMC).

...
Arguments passed to the chosen method.

quiet
If TRUE (the default) then fitting the model is wrapped in utils::capture.output().

Value

The fitted model object returned by the selected method.

Examples

# \dontrun{
print_example_program("logistic")

#> data {
#>   int<lower=0> N;
#>   int<lower=0> K;
#>   int<lower=0,upper=1> y[N];
#>   matrix[N, K] X;
#> }
#> parameters {
#>   real alpha;
#>   vector[K] beta;
#> }
#> model {
#>   target += normal_lpdf(alpha | 0, 1);
#>   target += normal_lpdf(beta | 0, 1);
#>   target += bernoulli_logit_glm_lpmf(y | X, alpha, beta);
#> }

fit_logistic_mcmc <- cmdstanr_example("logistic", chains = 2)

#> Compiling Stan program...

fit_logistic_mcmc$summary()

#> # A tibble: 5 x 10
#>   variable    mean  median    sd   mad       q5      q95  rhat ess_bulk ess_tail
#>   <chr>      <dbl>   <dbl> <dbl> <dbl>    <dbl>    <dbl> <dbl>    <dbl>    <dbl>
#> 1 lp__     -66.0   -65.6   1.48  1.29  -6.90e+1 -64.3     1.00    1123.    1526.
#> 2 alpha      0.380   0.378 0.224 0.214  9.52e-3   0.750   1.00    2082.    1591.
#> 3 beta[1]   -0.674  -0.674 0.256 0.259 -1.09e+0  -0.257   1.00    2169.    1526.
#> 4 beta[2]   -0.274  -0.276 0.227 0.231 -6.51e-1   0.0992  1.00    1760.    1415.
#> 5 beta[3]    0.677   0.665 0.270 0.269  2.57e-1   1.16    1.00    2241.    1630.

fit_logistic_optim <- cmdstanr_example("logistic", method = "optimize")

#> Model executable is up to date!

fit_logistic_optim$summary()

#> # A tibble: 5 x 2
#>   variable estimate
#>   <chr>       <dbl>
#> 1 lp__      -63.9  
#> 2 alpha       0.364
#> 3 beta[1]    -0.632
#> 4 beta[2]    -0.259
#> 5 beta[3]     0.649

fit_logistic_vb <- cmdstanr_example("logistic", method = "variational")

#> Model executable is up to date!

fit_logistic_vb$summary()

#> # A tibble: 6 x 7
#>   variable       mean  median    sd   mad        q5      q95
#>   <chr>         <dbl>   <dbl> <dbl> <dbl>     <dbl>    <dbl>
#> 1 lp__        -65.9   -65.6   1.42  1.25  -68.5     -64.2   
#> 2 lp_approx__  -2.00   -1.66  1.44  1.22   -4.77     -0.321 
#> 3 alpha         0.356   0.362 0.214 0.212   0.00182   0.718 
#> 4 beta[1]      -0.699  -0.688 0.231 0.232  -1.07     -0.329 
#> 5 beta[2]      -0.333  -0.334 0.218 0.222  -0.686     0.0451
#> 6 beta[3]       0.621   0.630 0.234 0.232   0.218     0.992 

print_example_program("schools")

#> data {
#>   int<lower=1> J;
#>   vector<lower=0>[J] sigma;
#>   vector[J] y;
#> }
#> parameters {
#>   real mu;
#>   real<lower=0> tau;
#>   vector[J] theta;
#> }
#> model {
#>   target += normal_lpdf(tau | 0, 10);
#>   target += normal_lpdf(mu | 0, 10);
#>   target += normal_lpdf(theta | mu, tau);
#>   target += normal_lpdf(y | theta, sigma);
#> }

fit_schools_mcmc <- cmdstanr_example("schools")

#> Compiling Stan program...

#> 81 of 4000 (2.0%) transitions ended with a divergence.
#> These divergent transitions indicate that HMC is not fully able to explore the posterior distribution.
#> Try increasing adapt delta closer to 1.
#> If this doesn't remove all divergences, try to reparameterize the model.

fit_schools_mcmc$summary()

#> # A tibble: 11 x 10
#>    variable   mean median    sd   mad       q5   q95  rhat ess_bulk ess_tail
#>    <chr>     <dbl>  <dbl> <dbl> <dbl>    <dbl> <dbl> <dbl>    <dbl>    <dbl>
#>  1 lp__     -57.9  -58.1   5.40  5.74 -66.6    -48.7  1.01     301.     396.
#>  2 mu         6.56   6.55  4.17  4.18  -0.175   13.3  1.00     807.    1199.
#>  3 tau        5.21   4.34  3.53  3.28   1.21    12.2  1.01     332.     357.
#>  4 theta[1]   9.33   8.53  7.01  5.86  -0.610   22.2  1.00    1208.    1720.
#>  5 theta[2]   6.93   6.89  5.64  5.25  -2.13    16.4  1.00    1307.    2024.
#>  6 theta[3]   5.48   5.73  6.82  5.64  -6.12    15.7  1.00    1164.    1739.
#>  7 theta[4]   6.61   6.47  5.73  5.15  -2.49    16.2  1.00    1386.    1903.
#>  8 theta[5]   4.62   4.97  5.65  5.22  -5.52    13.2  1.00    1072.    1776.
#>  9 theta[6]   5.55   5.60  5.88  5.36  -4.49    15.0  1.00    1145.    1822.
#> 10 theta[7]   8.96   8.44  5.96  5.47  -0.0548  19.5  1.00    1127.    1723.
#> 11 theta[8]   7.08   6.81  6.85  5.50  -4.04    18.4  1.00    1178.    1513.

print_example_program("schools_ncp")

#> data {
#>   int<lower=1> J;
#>   vector<lower=0>[J] sigma;
#>   vector[J] y;
#> }
#> parameters {
#>   real mu;
#>   real<lower=0> tau;
#>   vector[J] theta_raw;
#> }
#> transformed parameters {
#>   vector[J] theta = mu + tau * theta_raw;
#> }
#> model {
#>   target += normal_lpdf(tau | 0, 10);
#>   target += normal_lpdf(mu | 0, 10);
#>   target += normal_lpdf(theta_raw | 0, 1);
#>   target += normal_lpdf(y | theta, sigma);
#> }

fit_schools_ncp_mcmc <- cmdstanr_example("schools_ncp")

#> Compiling Stan program...

#> 1 of 4000 (0.0%) transitions ended with a divergence.
#> These divergent transitions indicate that HMC is not fully able to explore the posterior distribution.
#> Try increasing adapt delta closer to 1.
#> If this doesn't remove all divergences, try to reparameterize the model.

fit_schools_ncp_mcmc$summary()

#> # A tibble: 19 x 10
#>    variable     mean   median    sd   mad       q5    q95  rhat ess_bulk
#>    <chr>       <dbl>    <dbl> <dbl> <dbl>    <dbl>  <dbl> <dbl>    <dbl>
#>  1 lp__     -4.69e+1 -46.6    2.41  2.28  -51.3    -43.6   1.00    1601.
#>  2 mu        6.58e+0   6.51   4.21  4.18   -0.190   13.4   1.00    3765.
#>  3 tau       4.67e+0   3.90   3.53  3.38    0.367   11.5   1.00    2257.
#>  4 theta_r…  3.64e-1   0.390  0.967 0.971  -1.27     1.89  1.00    4478.
#>  5 theta_r…  4.41e-2   0.0447 0.923 0.931  -1.48     1.53  1.00    4439.
#>  6 theta_r… -1.35e-1  -0.149  0.997 0.974  -1.79     1.49  1.00    5175.
#>  7 theta_r…  4.61e-3  -0.0222 0.921 0.916  -1.48     1.52  1.00    4735.
#>  8 theta_r… -2.65e-1  -0.285  0.892 0.882  -1.71     1.20  1.00    4221.
#>  9 theta_r… -1.51e-1  -0.172  0.932 0.913  -1.66     1.41  1.00    4971.
#> 10 theta_r…  3.49e-1   0.362  0.919 0.895  -1.25     1.83  1.00    4395.
#> 11 theta_r…  5.05e-2   0.0755 0.958 0.903  -1.52     1.64  1.00    5184.
#> 12 theta[1]  8.96e+0   8.25   6.67  5.67   -0.351   21.3   1.00    3848.
#> 13 theta[2]  6.86e+0   6.74   5.48  5.19   -1.78    15.9   1.00    5245.
#> 14 theta[3]  5.53e+0   5.91   6.44  5.43   -5.64    15.2   1.00    4698.
#> 15 theta[4]  6.62e+0   6.58   5.66  5.22   -2.47    15.8   1.00    5038.
#> 16 theta[5]  4.91e+0   5.21   5.48  5.17   -4.50    13.0   1.00    4608.
#> 17 theta[6]  5.60e+0   5.84   5.56  5.00   -3.72    14.4   1.00    4624.
#> 18 theta[7]  8.74e+0   8.29   5.98  5.46    0.0295  19.4   1.00    4448.
#> 19 theta[8]  6.97e+0   6.76   6.34  5.41   -2.61    17.0   1.00    4343.
#> # … with 1 more variable: ess_tail <dbl>

# optimization fails for hierarchical model
cmdstanr_example("schools", "optimize", quiet = FALSE)

#> Model executable is up to date!

#> method = optimize 
#>   optimize 
#>     algorithm = lbfgs (Default) 
#>       lbfgs 
#>         init_alpha = 0.001 (Default) 
#>         tol_obj = 9.9999999999999998e-13 (Default) 
#>         tol_rel_obj = 10000 (Default) 
#>         tol_grad = 1e-08 (Default) 
#>         tol_rel_grad = 10000000 (Default) 
#>         tol_param = 1e-08 (Default) 
#>         history_size = 5 (Default) 
#>     iter = 2000 (Default) 
#>     save_iterations = 0 (Default) 
#> id = 1 
#> data 
#>   file = /private/var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpyTlvaJ/temp_libpath8e7b5ac5cf48/cmdstanr/schools.data.json 
#> init = 2 (Default) 
#> random 
#>   seed = 850275945 
#> output 
#>   file = /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpvSr5hc/schools-202007281357-1-d99d52.csv 
#>   diagnostic_file =  (Default) 
#>   refresh = 100 (Default) 
#>  
#> Initial log joint probability = -176.05 
#>     Iter      log prob        ||dx||      ||grad||       alpha      alpha0  # evals  Notes  
#>       99       124.074      0.201276   6.88626e+09      0.3374           1      176    
#>     Iter      log prob        ||dx||      ||grad||       alpha      alpha0  # evals  Notes  
#>      180       249.032     0.0338013   3.31169e+16       1e-12       0.001      421  LS failed, Hessian reset  
#> Optimization terminated with error:  
#>   Line search failed to achieve a sufficient decrease, no more progress can be made 

#> Warning: Fitting finished unexpectedly!

#> Error: Fitting failed. Unable to print.

# }