Fit models for use in examples

Fit models for use in examples

cmdstanr_example(
  example = c("logistic", "schools", "schools_ncp"),
  method = c("sample", "optimize", "variational"),
  ...,
  quiet = TRUE
)

print_example_program(example = c("logistic", "schools", "schools_ncp"))

Arguments

example
Name of the example. The currently available examples are:

"logistic": logistic regression with intercept and 3 predictors.

"schools": the so-called "eight schools" model, a hierarchical meta-analysis. Fitting this model will result in warnings about divergences.

"schools_ncp": non-centered parameterization eight schools model that fixes the problem with divergences.

To print the Stan code for a given example use print_example_program().

method
Which fitting method should be used? The default is the "sample" method (MCMC).

...
Arguments passed to the chosen method.

quiet
If TRUE (the default) then fitting the model is wrapped in utils::capture.output().

Value

The fitted model object returned by the selected method.

Examples

# \dontrun{
print_example_program("logistic")

#> data {
#>   int<lower=0> N;
#>   int<lower=0> K;
#>   int<lower=0,upper=1> y[N];
#>   matrix[N, K] X;
#> }
#> parameters {
#>   real alpha;
#>   vector[K] beta;
#> }
#> model {
#>   target += normal_lpdf(alpha | 0, 1);
#>   target += normal_lpdf(beta | 0, 1);
#>   target += bernoulli_logit_glm_lpmf(y | X, alpha, beta);
#> }

fit_logistic_mcmc <- cmdstanr_example("logistic", chains = 2)

#> Compiling Stan program...

fit_logistic_mcmc$summary()

#> # A tibble: 5 x 10
#>   variable    mean  median    sd   mad       q5      q95  rhat ess_bulk ess_tail
#>   <chr>      <dbl>   <dbl> <dbl> <dbl>    <dbl>    <dbl> <dbl>    <dbl>    <dbl>
#> 1 lp__     -66.0   -65.6   1.43  1.23  -68.6    -64.3     1.00    1143.    1550.
#> 2 alpha      0.377   0.370 0.217 0.219   0.0249   0.750   1.00    2141.    1631.
#> 3 beta[1]   -0.668  -0.656 0.254 0.262  -1.09    -0.271   1.00    2315.    1578.
#> 4 beta[2]   -0.267  -0.268 0.226 0.228  -0.633    0.0956  1.00    2028.    1564.
#> 5 beta[3]    0.668   0.668 0.260 0.264   0.241    1.10    1.00    1897.    1539.

fit_logistic_optim <- cmdstanr_example("logistic", method = "optimize")

#> Model executable is up to date!

fit_logistic_optim$summary()

#> # A tibble: 5 x 2
#>   variable estimate
#>   <chr>       <dbl>
#> 1 lp__      -63.9  
#> 2 alpha       0.364
#> 3 beta[1]    -0.632
#> 4 beta[2]    -0.259
#> 5 beta[3]     0.648

fit_logistic_vb <- cmdstanr_example("logistic", method = "variational")

#> Model executable is up to date!

fit_logistic_vb$summary()

#> # A tibble: 6 x 7
#>   variable       mean  median    sd   mad      q5     q95
#>   <chr>         <dbl>   <dbl> <dbl> <dbl>   <dbl>   <dbl>
#> 1 lp__        -66.1   -65.7   1.53  1.25  -69.1   -64.3  
#> 2 lp_approx__  -1.94   -1.63  1.33  1.18   -4.66   -0.370
#> 3 alpha         0.451   0.455 0.201 0.200   0.111   0.772
#> 4 beta[1]      -0.604  -0.610 0.220 0.228  -0.951  -0.236
#> 5 beta[2]      -0.228  -0.220 0.277 0.282  -0.703   0.218
#> 6 beta[3]       0.672   0.665 0.233 0.236   0.299   1.04 

print_example_program("schools")

#> data {
#>   int<lower=1> J;
#>   vector<lower=0>[J] sigma;
#>   vector[J] y;
#> }
#> parameters {
#>   real mu;
#>   real<lower=0> tau;
#>   vector[J] theta;
#> }
#> model {
#>   target += normal_lpdf(tau | 0, 10);
#>   target += normal_lpdf(mu | 0, 10);
#>   target += normal_lpdf(theta | mu, tau);
#>   target += normal_lpdf(y | theta, sigma);
#> }

fit_schools_mcmc <- cmdstanr_example("schools")

#> Compiling Stan program...

#> 102 of 4000 (3.0%) transitions ended with a divergence.
#> These divergent transitions indicate that HMC is not fully able to explore the posterior distribution.
#> Try increasing adapt delta closer to 1.
#> If this doesn't remove all divergences, try to reparameterize the model.

fit_schools_mcmc$summary()

#> # A tibble: 11 x 10
#>    variable   mean median    sd   mad      q5   q95  rhat ess_bulk ess_tail
#>    <chr>     <dbl>  <dbl> <dbl> <dbl>   <dbl> <dbl> <dbl>    <dbl>    <dbl>
#>  1 lp__     -58.2  -58.5   5.31  5.38 -66.6   -49.2  1.02     162.     207.
#>  2 mu         6.33   6.28  4.22  4.03  -0.753  13.1  1.00     839.    1019.
#>  3 tau        5.40   4.62  3.62  3.30   1.28   12.4  1.01     168.     172.
#>  4 theta[1]   9.20   8.33  7.40  6.17  -1.01   22.4  1.00    1119.    1579.
#>  5 theta[2]   6.75   6.56  5.65  5.25  -2.32   16.1  1.00    1574.    1996.
#>  6 theta[3]   5.27   5.58  6.58  5.88  -5.92   15.4  1.00    1528.    2516.
#>  7 theta[4]   6.43   6.39  5.93  5.46  -2.86   16.3  1.01    1301.    1483.
#>  8 theta[5]   4.41   4.76  5.59  5.10  -5.07   13.2  1.00    1138.    1884.
#>  9 theta[6]   5.29   5.50  6.12  5.19  -5.44   14.9  1.00    1509.    2030.
#> 10 theta[7]   8.88   8.48  6.13  5.70  -0.408  19.5  1.00     998.    1899.
#> 11 theta[8]   6.98   6.78  6.78  5.89  -3.93   18.5  1.00    1455.    1582.

print_example_program("schools_ncp")

#> data {
#>   int<lower=1> J;
#>   vector<lower=0>[J] sigma;
#>   vector[J] y;
#> }
#> parameters {
#>   real mu;
#>   real<lower=0> tau;
#>   vector[J] theta_raw;
#> }
#> transformed parameters {
#>   vector[J] theta = mu + tau * theta_raw;
#> }
#> model {
#>   target += normal_lpdf(tau | 0, 10);
#>   target += normal_lpdf(mu | 0, 10);
#>   target += normal_lpdf(theta_raw | 0, 1);
#>   target += normal_lpdf(y | theta, sigma);
#> }

fit_schools_ncp_mcmc <- cmdstanr_example("schools_ncp")

#> Compiling Stan program...

#> 1 of 4000 (0.0%) transitions ended with a divergence.
#> These divergent transitions indicate that HMC is not fully able to explore the posterior distribution.
#> Try increasing adapt delta closer to 1.
#> If this doesn't remove all divergences, try to reparameterize the model.

fit_schools_ncp_mcmc$summary()

#> # A tibble: 19 x 10
#>    variable     mean   median    sd   mad      q5    q95  rhat ess_bulk ess_tail
#>    <chr>       <dbl>    <dbl> <dbl> <dbl>   <dbl>  <dbl> <dbl>    <dbl>    <dbl>
#>  1 lp__     -46.9    -46.7    2.46  2.41  -51.4   -43.5  1.00     1479.    2220.
#>  2 mu         6.42     6.45   4.18  4.04   -0.361  13.4  1.00     3155.    2639.
#>  3 tau        4.58     3.81   3.56  3.37    0.335  11.3  1.00     2051.    1737.
#>  4 theta_r…   0.338    0.342  0.977 1.01   -1.25    1.91 1.00     3873.    2478.
#>  5 theta_r…   0.0594   0.0712 0.932 0.925  -1.49    1.56 1.00     3930.    2904.
#>  6 theta_r…  -0.101   -0.117  0.985 0.993  -1.69    1.52 1.00     3920.    2672.
#>  7 theta_r…   0.0263   0.0343 0.917 0.897  -1.50    1.48 1.00     3784.    2691.
#>  8 theta_r…  -0.263   -0.266  0.909 0.905  -1.73    1.24 1.00     4009.    2458.
#>  9 theta_r…  -0.145   -0.135  0.901 0.896  -1.61    1.30 1.00     3917.    2878.
#> 10 theta_r…   0.371    0.389  0.945 0.927  -1.23    1.87 1.00     3323.    2766.
#> 11 theta_r…   0.0703   0.0624 0.946 0.945  -1.45    1.63 1.00     4154.    2889.
#> 12 theta[1]   8.76     8.14   6.71  5.63   -0.988  20.8  1.00     3531.    2926.
#> 13 theta[2]   6.79     6.77   5.49  5.02   -2.17   15.8  0.999    4200.    3027.
#> 14 theta[3]   5.69     5.98   6.41  5.54   -5.15   15.6  1.00     3751.    2844.
#> 15 theta[4]   6.62     6.72   5.67  5.10   -2.65   15.9  1.00     3977.    3094.
#> 16 theta[5]   4.85     5.20   5.50  5.09   -4.81   13.2  1.00     4280.    3236.
#> 17 theta[6]   5.52     5.82   5.67  5.29   -4.04   14.1  1.00     3907.    3306.
#> 18 theta[7]   8.69     8.11   5.75  5.06    0.135  18.9  1.00     3579.    2975.
#> 19 theta[8]   6.98     6.91   6.44  5.38   -3.33   17.5  1.00     3626.    2605.

# optimization fails for hierarchical model
cmdstanr_example("schools", "optimize", quiet = FALSE)

#> Model executable is up to date!

#> method = optimize
#>   optimize
#>     algorithm = lbfgs (Default)
#>       lbfgs
#>         init_alpha = 0.001 (Default)
#>         tol_obj = 9.9999999999999998e-13 (Default)
#>         tol_rel_obj = 10000 (Default)
#>         tol_grad = 1e-08 (Default)
#>         tol_rel_grad = 10000000 (Default)
#>         tol_param = 1e-08 (Default)
#>         history_size = 5 (Default)
#>     iter = 2000 (Default)
#>     save_iterations = 0 (Default)
#> id = 1
#> data
#>   file = /private/var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpiX7M4F/temp_libpath1565047f7dc58/cmdstanr/schools.data.json
#> init = 2 (Default)
#> random
#>   seed = 1998422484
#> output
#>   file = /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpKbdtzw/schools-202006151402-1-73f8f5.csv
#>   diagnostic_file =  (Default)
#>   refresh = 100 (Default)
#> 
#> Initial log joint probability = -472.235
#>     Iter      log prob        ||dx||      ||grad||       alpha      alpha0  # evals  Notes 
#>       99        106.99      0.162734   4.09264e+08      0.2769      0.2769      176   
#>     Iter      log prob        ||dx||      ||grad||       alpha      alpha0  # evals  Notes 
#>      194       245.522       0.18154   2.16497e+16       1e-12       0.001      395  LS failed, Hessian reset 
#> Optimization terminated with error: 
#>   Line search failed to achieve a sufficient decrease, no more progress can be made

#> Error in processx::run(command = self$command(), args = self$command_args()[[1]],     wd = dirname(self$exe_file()), echo_cmd = FALSE, echo = TRUE,     error_on_status = TRUE): System command 'schools' failed

# }