Run CmdStan's stansummary and diagnose utilities

Run CmdStan's stansummary and diagnose utilities. These are documented in the CmdStan Guide:

https://mc-stan.org/docs/cmdstan-guide/stansummary.html

https://mc-stan.org/docs/cmdstan-guide/diagnose.html

Although these methods can be used for models fit using the $variational() method, much of the output is currently only relevant for models fit using the $sample() method.

See the $summary() for computing similar summaries in R rather than calling CmdStan's utilites.

cmdstan_summary(flags = NULL)

cmdstan_diagnose()

Arguments

flags

An optional character vector of flags (e.g. flags = c("--sig_figs=1")).

See also

Examples

# \dontrun{
fit <- cmdstanr_example("logistic")
fit$cmdstan_diagnose()
#> Processing csv files: /var/folders/s0/zfzm55px2nd2v__zlw5xfj2h0000gn/T/RtmpAMFHpW/logistic-202312131001-1-2718e9.csv, /var/folders/s0/zfzm55px2nd2v__zlw5xfj2h0000gn/T/RtmpAMFHpW/logistic-202312131001-2-2718e9.csv, /var/folders/s0/zfzm55px2nd2v__zlw5xfj2h0000gn/T/RtmpAMFHpW/logistic-202312131001-3-2718e9.csv, /var/folders/s0/zfzm55px2nd2v__zlw5xfj2h0000gn/T/RtmpAMFHpW/logistic-202312131001-4-2718e9.csv
#> 
#> Checking sampler transitions treedepth.
#> Treedepth satisfactory for all transitions.
#> 
#> Checking sampler transitions for divergences.
#> No divergent transitions found.
#> 
#> Checking E-BFMI - sampler transitions HMC potential energy.
#> E-BFMI satisfactory.
#> 
#> Effective sample size satisfactory.
#> 
#> Split R-hat values satisfactory all parameters.
#> 
#> Processing complete, no problems detected.
fit$cmdstan_summary()
#> Inference for Stan model: logistic_model
#> 4 chains: each with iter=(1000,1000,1000,1000); warmup=(0,0,0,0); thin=(1,1,1,1); 4000 iterations saved.
#> 
#> Warmup took (0.022, 0.023, 0.023, 0.023) seconds, 0.091 seconds total
#> Sampling took (0.071, 0.070, 0.070, 0.070) seconds, 0.28 seconds total
#> 
#>                     Mean     MCSE  StdDev     5%       50%       95%  N_Eff  N_Eff/s    R_hat
#> 
#> lp__            -6.6e+01  3.1e-02     1.4    -69  -6.6e+01  -6.4e+01   2120     7545      1.0
#> accept_stat__       0.91  1.4e-03    0.10   0.69      0.95       1.0   5437    19350  1.0e+00
#> stepsize__          0.74  2.1e-02   0.030   0.70      0.75      0.78    2.0      7.1  1.1e+13
#> treedepth__          2.4  8.8e-03    0.53    2.0       2.0       3.0   3601    12815  1.0e+00
#> n_leapfrog__         5.3  3.3e-02     2.0    3.0       7.0       7.0   3773    13428  1.0e+00
#> divergent__         0.00      nan    0.00   0.00      0.00      0.00    nan      nan      nan
#> energy__              68  4.9e-02     2.0     65        68        72   1750     6227  1.0e+00
#> 
#> alpha            3.8e-01  3.5e-03    0.22  0.026   3.8e-01   7.5e-01   3978    14156     1.00
#> beta[1]         -6.7e-01  4.1e-03    0.26   -1.1  -6.6e-01  -2.6e-01   3779    13447      1.0
#> beta[2]         -2.7e-01  3.5e-03    0.23  -0.64  -2.7e-01   1.0e-01   4061    14450      1.0
#> beta[3]          6.7e-01  4.3e-03    0.27   0.24   6.7e-01   1.1e+00   3879    13804     1.00
#> log_lik[1]      -5.1e-01  1.5e-03   0.097  -0.68  -5.1e-01  -3.7e-01   3951    14060     1.00
#> log_lik[2]      -4.1e-01  2.3e-03    0.15  -0.69  -3.9e-01  -1.9e-01   4167    14830     1.00
#> log_lik[3]      -4.9e-01  3.3e-03    0.21  -0.88  -4.6e-01  -2.1e-01   4010    14272     1.00
#> log_lik[4]      -4.5e-01  2.5e-03    0.15  -0.72  -4.4e-01  -2.4e-01   3751    13350      1.0
#> log_lik[5]      -1.2e+00  4.4e-03    0.28   -1.7  -1.2e+00  -7.6e-01   4068    14476      1.0
#> log_lik[6]      -5.9e-01  3.0e-03    0.19  -0.96  -5.7e-01  -3.2e-01   4275    15213      1.0
#> log_lik[7]      -6.4e-01  2.0e-03    0.13  -0.86  -6.4e-01  -4.5e-01   4183    14885      1.0
#> log_lik[8]      -2.8e-01  2.3e-03    0.14  -0.55  -2.5e-01  -1.1e-01   3659    13021      1.0
#> log_lik[9]      -6.9e-01  2.7e-03    0.17   -1.0  -6.7e-01  -4.4e-01   4128    14690      1.0
#> log_lik[10]     -7.3e-01  3.6e-03    0.22   -1.1  -7.1e-01  -4.1e-01   3894    13857     1.00
#> log_lik[11]     -2.8e-01  2.1e-03    0.13  -0.54  -2.6e-01  -1.1e-01   3822    13601     1.00
#> log_lik[12]     -5.0e-01  3.8e-03    0.24  -0.95  -4.6e-01  -1.9e-01   3952    14065     1.00
#> log_lik[13]     -6.6e-01  3.3e-03    0.21   -1.0  -6.4e-01  -3.5e-01   4181    14878      1.0
#> log_lik[14]     -3.6e-01  2.6e-03    0.17  -0.67  -3.2e-01  -1.4e-01   4021    14308     1.00
#> log_lik[15]     -2.8e-01  1.7e-03    0.10  -0.47  -2.6e-01  -1.3e-01   3858    13730     1.00
#> log_lik[16]     -2.8e-01  1.5e-03   0.091  -0.44  -2.7e-01  -1.5e-01   3706    13189     1.00
#> log_lik[17]     -1.6e+00  5.0e-03    0.30   -2.1  -1.6e+00  -1.1e+00   3612    12854      1.0
#> log_lik[18]     -4.8e-01  1.8e-03    0.11  -0.68  -4.7e-01  -3.1e-01   4081    14523      1.0
#> log_lik[19]     -2.3e-01  1.3e-03   0.079  -0.37  -2.2e-01  -1.2e-01   3751    13347      1.0
#> log_lik[20]     -1.1e-01  1.3e-03   0.078  -0.26  -9.4e-02  -2.8e-02   3798    13516     1.00
#> log_lik[21]     -2.1e-01  1.5e-03   0.091  -0.39  -2.0e-01  -9.3e-02   3626    12906     1.00
#> log_lik[22]     -5.7e-01  2.4e-03    0.15  -0.86  -5.5e-01  -3.5e-01   4061    14452      1.0
#> log_lik[23]     -3.3e-01  2.3e-03    0.15  -0.61  -3.1e-01  -1.4e-01   3872    13780      1.0
#> log_lik[24]     -1.4e-01  1.1e-03   0.069  -0.27  -1.2e-01  -5.1e-02   3743    13321     1.00
#> log_lik[25]     -4.6e-01  1.9e-03    0.12  -0.68  -4.5e-01  -2.8e-01   3975    14145      1.0
#> log_lik[26]     -1.5e+00  5.8e-03    0.35   -2.1  -1.5e+00  -1.0e+00   3699    13164      1.0
#> log_lik[27]     -3.1e-01  2.0e-03    0.13  -0.55  -2.9e-01  -1.4e-01   3831    13633     1.00
#> log_lik[28]     -4.4e-01  1.4e-03   0.086  -0.59  -4.4e-01  -3.1e-01   3970    14130      1.0
#> log_lik[29]     -7.3e-01  3.4e-03    0.22   -1.1  -7.1e-01  -4.0e-01   4258    15153      1.0
#> log_lik[30]     -6.9e-01  3.1e-03    0.20   -1.1  -6.8e-01  -4.0e-01   4121    14666     1.00
#> log_lik[31]     -4.9e-01  2.6e-03    0.17  -0.81  -4.7e-01  -2.5e-01   4046    14398      1.0
#> log_lik[32]     -4.2e-01  1.7e-03    0.11  -0.62  -4.1e-01  -2.6e-01   3986    14185      1.0
#> log_lik[33]     -4.1e-01  2.0e-03    0.13  -0.65  -4.0e-01  -2.2e-01   4122    14669      1.0
#> log_lik[34]     -6.5e-02  9.1e-04   0.054  -0.17  -5.0e-02  -1.3e-02   3544    12612      1.0
#> log_lik[35]     -5.9e-01  2.8e-03    0.18  -0.92  -5.7e-01  -3.2e-01   4411    15696      1.0
#> log_lik[36]     -3.3e-01  2.2e-03    0.14  -0.58  -3.0e-01  -1.4e-01   4001    14239      1.0
#> log_lik[37]     -6.9e-01  3.4e-03    0.22   -1.1  -6.7e-01  -3.7e-01   4045    14394      1.0
#> log_lik[38]     -3.2e-01  2.4e-03    0.15  -0.60  -2.9e-01  -1.2e-01   3854    13717      1.0
#> log_lik[39]     -1.8e-01  1.7e-03    0.11  -0.38  -1.5e-01  -5.2e-02   3915    13934     1.00
#> log_lik[40]     -6.8e-01  2.1e-03    0.13  -0.92  -6.7e-01  -4.8e-01   4145    14752      1.0
#> log_lik[41]     -1.1e+00  4.2e-03    0.26   -1.6  -1.1e+00  -7.3e-01   3929    13983      1.0
#> log_lik[42]     -9.3e-01  3.2e-03    0.20   -1.3  -9.1e-01  -6.3e-01   3967    14116      1.0
#> log_lik[43]     -4.1e-01  4.0e-03    0.26  -0.91  -3.4e-01  -1.0e-01   4191    14913      1.0
#> log_lik[44]     -1.2e+00  3.1e-03    0.19   -1.5  -1.2e+00  -9.0e-01   3709    13201     1.00
#> log_lik[45]     -3.6e-01  1.8e-03    0.11  -0.56  -3.4e-01  -1.9e-01   4009    14267     1.00
#> log_lik[46]     -5.9e-01  2.0e-03    0.13  -0.82  -5.8e-01  -3.8e-01   4248    15119     1.00
#> log_lik[47]     -3.1e-01  2.2e-03    0.13  -0.55  -2.9e-01  -1.3e-01   3696    13153      1.0
#> log_lik[48]     -3.2e-01  1.4e-03   0.085  -0.47  -3.2e-01  -2.0e-01   3847    13691      1.0
#> log_lik[49]     -3.2e-01  1.3e-03   0.082  -0.46  -3.1e-01  -2.0e-01   3716    13223     1.00
#> log_lik[50]     -1.3e+00  5.4e-03    0.33   -1.9  -1.3e+00  -8.1e-01   3859    13734      1.0
#> log_lik[51]     -2.9e-01  1.6e-03   0.096  -0.46  -2.8e-01  -1.5e-01   3714    13218      1.0
#> log_lik[52]     -8.4e-01  2.2e-03    0.14   -1.1  -8.3e-01  -6.2e-01   3985    14182     1.00
#> log_lik[53]     -4.0e-01  2.1e-03    0.13  -0.66  -3.9e-01  -2.1e-01   3970    14127     1.00
#> log_lik[54]     -3.7e-01  2.2e-03    0.14  -0.62  -3.5e-01  -1.8e-01   4008    14263     1.00
#> log_lik[55]     -3.9e-01  2.2e-03    0.14  -0.65  -3.7e-01  -1.9e-01   4107    14614     1.00
#> log_lik[56]     -3.2e-01  3.2e-03    0.20  -0.70  -2.8e-01  -9.3e-02   3745    13326      1.0
#> log_lik[57]     -6.5e-01  1.8e-03    0.12  -0.85  -6.5e-01  -4.8e-01   3918    13943      1.0
#> log_lik[58]     -9.4e-01  5.3e-03    0.35   -1.6  -9.0e-01  -4.5e-01   4317    15364      1.0
#> log_lik[59]     -1.4e+00  5.7e-03    0.35   -2.0  -1.3e+00  -8.4e-01   3801    13527      1.0
#> log_lik[60]     -9.8e-01  2.5e-03    0.16   -1.3  -9.7e-01  -7.4e-01   3902    13886     1.00
#> log_lik[61]     -5.4e-01  1.5e-03   0.097  -0.70  -5.3e-01  -3.9e-01   4078    14513     1.00
#> log_lik[62]     -8.9e-01  5.0e-03    0.31   -1.5  -8.5e-01  -4.4e-01   3852    13708     1.00
#> log_lik[63]     -1.2e-01  1.3e-03   0.077  -0.27  -9.9e-02  -3.2e-02   3330    11849     1.00
#> log_lik[64]     -9.0e-01  3.8e-03    0.25   -1.4  -8.8e-01  -5.3e-01   4445    15817     1.00
#> log_lik[65]     -2.0e+00  9.8e-03    0.61   -3.1  -2.0e+00  -1.1e+00   3783    13462     1.00
#> log_lik[66]     -5.1e-01  2.1e-03    0.13  -0.74  -4.9e-01  -3.1e-01   4154    14782      1.0
#> log_lik[67]     -2.8e-01  1.4e-03   0.085  -0.43  -2.7e-01  -1.6e-01   3827    13618      1.0
#> log_lik[68]     -1.1e+00  3.8e-03    0.24   -1.5  -1.0e+00  -6.8e-01   4102    14597     1.00
#> log_lik[69]     -4.3e-01  1.4e-03   0.085  -0.58  -4.3e-01  -3.0e-01   3842    13673      1.0
#> log_lik[70]     -6.4e-01  3.6e-03    0.24   -1.1  -6.1e-01  -3.1e-01   4371    15555     1.00
#> log_lik[71]     -6.1e-01  3.2e-03    0.21  -0.99  -5.9e-01  -3.1e-01   4401    15663      1.0
#> log_lik[72]     -4.6e-01  2.7e-03    0.17  -0.79  -4.5e-01  -2.2e-01   4214    14996      1.0
#> log_lik[73]     -1.5e+00  6.1e-03    0.37   -2.1  -1.4e+00  -9.2e-01   3671    13065      1.0
#> log_lik[74]     -9.5e-01  3.0e-03    0.19   -1.3  -9.4e-01  -6.7e-01   3944    14035     1.00
#> log_lik[75]     -1.2e+00  6.4e-03    0.39   -1.8  -1.1e+00  -5.9e-01   3678    13088     1.00
#> log_lik[76]     -3.7e-01  2.1e-03    0.13  -0.62  -3.5e-01  -1.8e-01   3976    14151     1.00
#> log_lik[77]     -8.8e-01  2.2e-03    0.14   -1.1  -8.8e-01  -6.6e-01   4034    14357      1.0
#> log_lik[78]     -4.8e-01  2.7e-03    0.17  -0.79  -4.6e-01  -2.4e-01   4097    14581     1.00
#> log_lik[79]     -7.6e-01  2.9e-03    0.18   -1.1  -7.4e-01  -4.8e-01   3889    13840      1.0
#> log_lik[80]     -5.5e-01  3.0e-03    0.20  -0.91  -5.2e-01  -2.7e-01   4407    15683     1.00
#> log_lik[81]     -1.7e-01  1.7e-03    0.10  -0.37  -1.4e-01  -4.8e-02   3780    13450      1.0
#> log_lik[82]     -2.3e-01  2.3e-03    0.14  -0.51  -2.0e-01  -6.3e-02   3943    14031     1.00
#> log_lik[83]     -3.4e-01  1.3e-03   0.081  -0.48  -3.4e-01  -2.2e-01   3766    13404     1.00
#> log_lik[84]     -2.7e-01  1.5e-03   0.090  -0.44  -2.6e-01  -1.4e-01   3767    13405     1.00
#> log_lik[85]     -1.3e-01  1.2e-03   0.075  -0.28  -1.1e-01  -4.1e-02   3780    13453     1.00
#> log_lik[86]     -1.1e+00  5.0e-03    0.32   -1.7  -1.1e+00  -6.4e-01   4173    14850     1.00
#> log_lik[87]     -8.3e-01  2.0e-03    0.13   -1.1  -8.2e-01  -6.2e-01   4183    14887     1.00
#> log_lik[88]     -7.8e-01  3.8e-03    0.24   -1.2  -7.5e-01  -4.2e-01   4157    14794      1.0
#> log_lik[89]     -1.3e+00  5.1e-03    0.32   -1.9  -1.2e+00  -7.9e-01   3879    13805     1.00
#> log_lik[90]     -2.6e-01  2.2e-03    0.14  -0.52  -2.4e-01  -9.4e-02   4008    14262      1.0
#> log_lik[91]     -3.9e-01  2.1e-03    0.13  -0.63  -3.8e-01  -2.1e-01   4144    14746      1.0
#> log_lik[92]     -1.5e+00  5.6e-03    0.34   -2.1  -1.5e+00  -9.7e-01   3829    13625     1.00
#> log_lik[93]     -7.5e-01  3.4e-03    0.21   -1.1  -7.2e-01  -4.4e-01   3888    13835     1.00
#> log_lik[94]     -3.2e-01  1.4e-03   0.087  -0.47  -3.1e-01  -1.9e-01   3803    13533     1.00
#> log_lik[95]     -3.9e-01  1.9e-03    0.12  -0.60  -3.8e-01  -2.2e-01   3889    13841      1.0
#> log_lik[96]     -1.6e+00  4.8e-03    0.29   -2.1  -1.6e+00  -1.1e+00   3601    12815     1.00
#> log_lik[97]     -4.3e-01  1.6e-03    0.10  -0.61  -4.2e-01  -2.7e-01   4067    14472      1.0
#> log_lik[98]     -1.0e+00  5.6e-03    0.37   -1.7  -1.0e+00  -5.2e-01   4342    15452      1.0
#> log_lik[99]     -7.0e-01  2.2e-03    0.15  -0.96  -6.9e-01  -4.7e-01   4351    15484     1.00
#> log_lik[100]    -3.9e-01  1.5e-03   0.097  -0.55  -3.8e-01  -2.4e-01   4019    14301     1.00
#> 
#> Samples were drawn using hmc with nuts.
#> For each parameter, N_Eff is a crude measure of effective sample size,
#> and R_hat is the potential scale reduction factor on split chains (at 
#> convergence, R_hat=1).
# }