Run CmdStan's stansummary and diagnose utilities

Run CmdStan's stansummary and diagnose utilities. These are documented in the CmdStan Guide:

https://mc-stan.org/docs/cmdstan-guide/stansummary.html

https://mc-stan.org/docs/cmdstan-guide/diagnose.html

Although these methods can be used for models fit using the $variational() method, much of the output is currently only relevant for models fit using the $sample() method.

See the $summary() for computing similar summaries in R rather than calling CmdStan's utilites.

cmdstan_summary(flags = NULL)

cmdstan_diagnose()

Arguments

flags

An optional character vector of flags (e.g. flags = c("--sig_figs=1")).

See also

Examples

# \dontrun{
fit <- cmdstanr_example("logistic")
fit$cmdstan_diagnose()
#> Processing csv files: /var/folders/s0/zfzm55px2nd2v__zlw5xfj2h0000gn/T/RtmpGCiVFY/logistic-202311081026-1-1f791d.csv, /var/folders/s0/zfzm55px2nd2v__zlw5xfj2h0000gn/T/RtmpGCiVFY/logistic-202311081026-2-1f791d.csv, /var/folders/s0/zfzm55px2nd2v__zlw5xfj2h0000gn/T/RtmpGCiVFY/logistic-202311081026-3-1f791d.csv, /var/folders/s0/zfzm55px2nd2v__zlw5xfj2h0000gn/T/RtmpGCiVFY/logistic-202311081026-4-1f791d.csv
#> 
#> Checking sampler transitions treedepth.
#> Treedepth satisfactory for all transitions.
#> 
#> Checking sampler transitions for divergences.
#> No divergent transitions found.
#> 
#> Checking E-BFMI - sampler transitions HMC potential energy.
#> E-BFMI satisfactory.
#> 
#> Effective sample size satisfactory.
#> 
#> Split R-hat values satisfactory all parameters.
#> 
#> Processing complete, no problems detected.
fit$cmdstan_summary()
#> Inference for Stan model: logistic_model
#> 4 chains: each with iter=(1000,1000,1000,1000); warmup=(0,0,0,0); thin=(1,1,1,1); 4000 iterations saved.
#> 
#> Warmup took (0.023, 0.023, 0.023, 0.023) seconds, 0.092 seconds total
#> Sampling took (0.076, 0.070, 0.072, 0.073) seconds, 0.29 seconds total
#> 
#>                     Mean     MCSE  StdDev     5%       50%       95%  N_Eff  N_Eff/s    R_hat
#> 
#> lp__            -6.6e+01  3.0e-02     1.4    -69  -6.6e+01  -6.4e+01   2277     7826      1.0
#> accept_stat__       0.89  1.8e-03    0.12   0.63      0.94       1.0   4996    17168  1.0e+00
#> stepsize__          0.80  4.6e-02   0.066   0.73      0.85      0.88    2.0      6.9  2.6e+13
#> treedepth__          2.3  8.8e-02    0.50    2.0       2.0       3.0     32      109  1.0e+00
#> n_leapfrog__         4.8  2.8e-01     2.0    3.0       3.0       7.0     53      182  1.0e+00
#> divergent__         0.00      nan    0.00   0.00      0.00      0.00    nan      nan      nan
#> energy__              68  4.5e-02     2.0     65        68        72   1941     6671  1.0e+00
#> 
#> alpha            3.8e-01  3.5e-03    0.22  0.014   3.8e-01   7.4e-01   4054    13931      1.0
#> beta[1]         -6.7e-01  3.7e-03    0.25   -1.1  -6.7e-01  -2.8e-01   4566    15692      1.0
#> beta[2]         -2.8e-01  3.9e-03    0.23  -0.66  -2.8e-01   9.5e-02   3521    12100      1.0
#> beta[3]          6.8e-01  4.0e-03    0.26   0.25   6.8e-01   1.1e+00   4334    14894      1.0
#> log_lik[1]      -5.2e-01  1.6e-03    0.10  -0.69  -5.1e-01  -3.6e-01   4009    13775      1.0
#> log_lik[2]      -4.0e-01  2.2e-03    0.15  -0.69  -3.8e-01  -1.9e-01   4671    16052      1.0
#> log_lik[3]      -5.0e-01  3.7e-03    0.22  -0.91  -4.7e-01  -2.1e-01   3570    12267      1.0
#> log_lik[4]      -4.5e-01  2.4e-03    0.15  -0.72  -4.3e-01  -2.3e-01   3825    13143      1.0
#> log_lik[5]      -1.2e+00  4.1e-03    0.28   -1.7  -1.2e+00  -7.6e-01   4632    15916      1.0
#> log_lik[6]      -5.9e-01  2.8e-03    0.19  -0.93  -5.7e-01  -3.2e-01   4602    15813      1.0
#> log_lik[7]      -6.4e-01  1.9e-03    0.13  -0.86  -6.3e-01  -4.5e-01   4425    15206      1.0
#> log_lik[8]      -2.7e-01  2.3e-03    0.14  -0.54  -2.5e-01  -1.0e-01   3551    12204      1.0
#> log_lik[9]      -6.9e-01  2.6e-03    0.17  -0.99  -6.8e-01  -4.4e-01   4396    15105      1.0
#> log_lik[10]     -7.5e-01  4.0e-03    0.24   -1.2  -7.2e-01  -4.1e-01   3524    12111      1.0
#> log_lik[11]     -2.8e-01  1.9e-03    0.12  -0.51  -2.6e-01  -1.2e-01   4067    13978     1.00
#> log_lik[12]     -5.0e-01  3.5e-03    0.24  -0.94  -4.6e-01  -1.9e-01   4672    16056     1.00
#> log_lik[13]     -6.6e-01  3.0e-03    0.21   -1.0  -6.4e-01  -3.6e-01   4681    16086      1.0
#> log_lik[14]     -3.6e-01  2.9e-03    0.17  -0.69  -3.3e-01  -1.4e-01   3667    12602      1.0
#> log_lik[15]     -2.8e-01  1.6e-03    0.11  -0.47  -2.6e-01  -1.4e-01   4253    14617      1.0
#> log_lik[16]     -2.7e-01  1.5e-03   0.087  -0.43  -2.6e-01  -1.5e-01   3572    12276      1.0
#> log_lik[17]     -1.6e+00  4.6e-03    0.29   -2.1  -1.6e+00  -1.2e+00   3940    13539     1.00
#> log_lik[18]     -4.8e-01  1.7e-03    0.11  -0.68  -4.7e-01  -3.1e-01   4208    14461      1.0
#> log_lik[19]     -2.3e-01  1.3e-03   0.075  -0.37  -2.2e-01  -1.2e-01   3488    11985     1.00
#> log_lik[20]     -1.1e-01  1.3e-03   0.078  -0.26  -9.3e-02  -2.8e-02   3400    11685      1.0
#> log_lik[21]     -2.1e-01  1.4e-03   0.086  -0.37  -2.0e-01  -9.3e-02   3747    12875      1.0
#> log_lik[22]     -5.7e-01  2.3e-03    0.15  -0.84  -5.5e-01  -3.5e-01   4159    14294      1.0
#> log_lik[23]     -3.3e-01  2.3e-03    0.14  -0.60  -3.0e-01  -1.4e-01   3876    13319      1.0
#> log_lik[24]     -1.3e-01  1.1e-03   0.066  -0.26  -1.2e-01  -5.1e-02   3311    11379      1.0
#> log_lik[25]     -4.5e-01  1.9e-03    0.12  -0.67  -4.4e-01  -2.7e-01   4263    14650      1.0
#> log_lik[26]     -1.5e+00  4.9e-03    0.34   -2.1  -1.5e+00  -1.0e+00   4863    16712     1.00
#> log_lik[27]     -3.0e-01  1.9e-03    0.12  -0.53  -2.9e-01  -1.4e-01   3942    13545      1.0
#> log_lik[28]     -4.4e-01  1.3e-03   0.085  -0.59  -4.4e-01  -3.1e-01   4011    13785      1.0
#> log_lik[29]     -7.2e-01  3.4e-03    0.23   -1.1  -6.9e-01  -3.9e-01   4401    15125      1.0
#> log_lik[30]     -6.9e-01  2.9e-03    0.20   -1.0  -6.7e-01  -4.0e-01   4483    15406      1.0
#> log_lik[31]     -4.9e-01  2.5e-03    0.16  -0.78  -4.7e-01  -2.6e-01   4140    14226      1.0
#> log_lik[32]     -4.2e-01  1.6e-03    0.11  -0.61  -4.1e-01  -2.7e-01   4397    15110      1.0
#> log_lik[33]     -4.1e-01  1.9e-03    0.13  -0.64  -4.0e-01  -2.2e-01   4495    15446      1.0
#> log_lik[34]     -6.2e-02  8.6e-04   0.050  -0.16  -4.8e-02  -1.2e-02   3377    11605      1.0
#> log_lik[35]     -5.8e-01  2.5e-03    0.18  -0.90  -5.6e-01  -3.2e-01   5084    17471     1.00
#> log_lik[36]     -3.2e-01  2.0e-03    0.13  -0.56  -3.0e-01  -1.4e-01   4311    14814      1.0
#> log_lik[37]     -7.1e-01  3.7e-03    0.23   -1.1  -6.8e-01  -3.8e-01   3857    13253      1.0
#> log_lik[38]     -3.1e-01  2.3e-03    0.15  -0.58  -2.8e-01  -1.2e-01   3910    13435      1.0
#> log_lik[39]     -1.8e-01  1.8e-03    0.11  -0.39  -1.5e-01  -5.3e-02   3559    12229      1.0
#> log_lik[40]     -6.8e-01  2.0e-03    0.13  -0.92  -6.7e-01  -4.8e-01   4390    15085      1.0
#> log_lik[41]     -1.1e+00  4.2e-03    0.26   -1.6  -1.1e+00  -7.5e-01   3790    13024      1.0
#> log_lik[42]     -9.3e-01  3.1e-03    0.20   -1.3  -9.1e-01  -6.3e-01   4446    15278      1.0
#> log_lik[43]     -4.0e-01  3.5e-03    0.25  -0.89  -3.5e-01  -1.0e-01   5239    18003     1.00
#> log_lik[44]     -1.2e+00  3.0e-03    0.19   -1.5  -1.2e+00  -8.9e-01   3904    13415      1.0
#> log_lik[45]     -3.6e-01  1.8e-03    0.12  -0.57  -3.4e-01  -1.9e-01   4092    14061      1.0
#> log_lik[46]     -5.8e-01  2.0e-03    0.13  -0.82  -5.7e-01  -3.8e-01   4539    15598      1.0
#> log_lik[47]     -3.0e-01  2.0e-03    0.13  -0.55  -2.8e-01  -1.3e-01   4158    14288      1.0
#> log_lik[48]     -3.2e-01  1.3e-03   0.083  -0.47  -3.1e-01  -2.0e-01   4285    14726     1.00
#> log_lik[49]     -3.2e-01  1.3e-03   0.080  -0.46  -3.1e-01  -2.0e-01   3835    13178      1.0
#> log_lik[50]     -1.3e+00  4.6e-03    0.33   -1.9  -1.3e+00  -8.0e-01   5110    17559     1.00
#> log_lik[51]     -2.9e-01  1.4e-03   0.094  -0.45  -2.7e-01  -1.5e-01   4316    14833     1.00
#> log_lik[52]     -8.3e-01  2.3e-03    0.14   -1.1  -8.3e-01  -6.1e-01   4123    14167      1.0
#> log_lik[53]     -4.0e-01  2.0e-03    0.13  -0.64  -3.8e-01  -2.2e-01   3978    13669      1.0
#> log_lik[54]     -3.7e-01  2.2e-03    0.14  -0.63  -3.5e-01  -1.8e-01   4204    14447      1.0
#> log_lik[55]     -3.9e-01  2.0e-03    0.14  -0.64  -3.7e-01  -1.9e-01   4560    15671      1.0
#> log_lik[56]     -3.1e-01  2.9e-03    0.19  -0.68  -2.7e-01  -8.9e-02   4172    14335      1.0
#> log_lik[57]     -6.6e-01  1.9e-03    0.12  -0.87  -6.5e-01  -4.8e-01   4000    13747      1.0
#> log_lik[58]     -9.6e-01  5.3e-03    0.35   -1.6  -9.1e-01  -4.6e-01   4421    15193      1.0
#> log_lik[59]     -1.4e+00  5.3e-03    0.35   -2.0  -1.3e+00  -8.4e-01   4356    14970      1.0
#> log_lik[60]     -9.8e-01  2.5e-03    0.16   -1.3  -9.7e-01  -7.3e-01   4051    13921      1.0
#> log_lik[61]     -5.4e-01  1.5e-03    0.10  -0.71  -5.3e-01  -3.9e-01   4197    14422      1.0
#> log_lik[62]     -8.8e-01  4.6e-03    0.31   -1.4  -8.4e-01  -4.4e-01   4561    15674     1.00
#> log_lik[63]     -1.1e-01  1.3e-03   0.072  -0.26  -9.8e-02  -3.2e-02   3248    11162      1.0
#> log_lik[64]     -9.1e-01  3.9e-03    0.26   -1.4  -8.9e-01  -5.3e-01   4383    15063      1.0
#> log_lik[65]     -2.0e+00  9.5e-03    0.59   -3.0  -2.0e+00  -1.1e+00   3828    13154     1.00
#> log_lik[66]     -5.1e-01  2.0e-03    0.13  -0.75  -5.0e-01  -3.1e-01   4591    15777      1.0
#> log_lik[67]     -2.7e-01  1.3e-03   0.082  -0.42  -2.7e-01  -1.6e-01   3737    12841     1.00
#> log_lik[68]     -1.1e+00  3.9e-03    0.24   -1.5  -1.1e+00  -7.0e-01   3952    13582      1.0
#> log_lik[69]     -4.4e-01  1.3e-03   0.085  -0.59  -4.3e-01  -3.1e-01   3944    13552      1.0
#> log_lik[70]     -6.3e-01  3.4e-03    0.24   -1.1  -6.0e-01  -3.0e-01   5018    17243      1.0
#> log_lik[71]     -6.1e-01  2.8e-03    0.21  -0.98  -5.9e-01  -3.2e-01   5227    17963     1.00
#> log_lik[72]     -4.6e-01  2.6e-03    0.17  -0.78  -4.4e-01  -2.3e-01   4552    15644      1.0
#> log_lik[73]     -1.5e+00  5.9e-03    0.37   -2.2  -1.5e+00  -9.4e-01   3999    13742      1.0
#> log_lik[74]     -9.4e-01  3.0e-03    0.19   -1.3  -9.3e-01  -6.5e-01   4092    14063      1.0
#> log_lik[75]     -1.1e+00  5.6e-03    0.38   -1.8  -1.1e+00  -5.9e-01   4639    15941     1.00
#> log_lik[76]     -3.7e-01  2.3e-03    0.14  -0.63  -3.5e-01  -1.8e-01   3728    12812      1.0
#> log_lik[77]     -8.8e-01  2.2e-03    0.14   -1.1  -8.7e-01  -6.6e-01   4154    14274      1.0
#> log_lik[78]     -4.9e-01  2.4e-03    0.17  -0.80  -4.7e-01  -2.5e-01   4939    16971     1.00
#> log_lik[79]     -7.7e-01  3.2e-03    0.19   -1.1  -7.5e-01  -4.9e-01   3734    12831      1.0
#> log_lik[80]     -5.5e-01  2.8e-03    0.20  -0.91  -5.2e-01  -2.7e-01   5052    17361      1.0
#> log_lik[81]     -1.6e-01  1.6e-03   0.100  -0.35  -1.4e-01  -4.6e-02   4042    13890      1.0
#> log_lik[82]     -2.2e-01  2.1e-03    0.14  -0.50  -1.9e-01  -6.1e-02   4473    15371      1.0
#> log_lik[83]     -3.4e-01  1.3e-03   0.081  -0.49  -3.4e-01  -2.2e-01   3946    13559      1.0
#> log_lik[84]     -2.8e-01  1.4e-03   0.091  -0.44  -2.6e-01  -1.5e-01   4018    13807      1.0
#> log_lik[85]     -1.3e-01  1.3e-03   0.074  -0.27  -1.1e-01  -4.2e-02   3489    11990      1.0
#> log_lik[86]     -1.1e+00  4.7e-03    0.32   -1.7  -1.1e+00  -6.4e-01   4742    16295      1.0
#> log_lik[87]     -8.3e-01  2.0e-03    0.13   -1.0  -8.2e-01  -6.2e-01   4289    14739      1.0
#> log_lik[88]     -7.8e-01  3.5e-03    0.24   -1.2  -7.5e-01  -4.3e-01   4606    15828     1.00
#> log_lik[89]     -1.3e+00  4.9e-03    0.32   -1.9  -1.3e+00  -7.9e-01   4346    14936      1.0
#> log_lik[90]     -2.6e-01  2.1e-03    0.13  -0.51  -2.3e-01  -9.0e-02   4192    14407     1.00
#> log_lik[91]     -3.9e-01  1.9e-03    0.13  -0.62  -3.7e-01  -2.0e-01   4638    15940      1.0
#> log_lik[92]     -1.5e+00  5.2e-03    0.34   -2.1  -1.5e+00  -9.9e-01   4259    14636      1.0
#> log_lik[93]     -7.3e-01  3.6e-03    0.22   -1.1  -7.1e-01  -4.2e-01   3742    12858      1.0
#> log_lik[94]     -3.2e-01  1.4e-03   0.088  -0.48  -3.1e-01  -1.9e-01   3956    13596      1.0
#> log_lik[95]     -3.9e-01  1.8e-03    0.11  -0.59  -3.8e-01  -2.3e-01   3846    13217      1.0
#> log_lik[96]     -1.6e+00  4.5e-03    0.28   -2.1  -1.6e+00  -1.1e+00   3868    13293      1.0
#> log_lik[97]     -4.3e-01  1.5e-03    0.10  -0.61  -4.2e-01  -2.7e-01   4724    16235      1.0
#> log_lik[98]     -1.1e+00  5.5e-03    0.37   -1.7  -1.0e+00  -5.2e-01   4560    15668      1.0
#> log_lik[99]     -7.0e-01  2.2e-03    0.15  -0.95  -6.8e-01  -4.7e-01   4389    15084      1.0
#> log_lik[100]    -3.9e-01  1.5e-03   0.097  -0.56  -3.8e-01  -2.5e-01   4436    15244     1.00
#> 
#> Samples were drawn using hmc with nuts.
#> For each parameter, N_Eff is a crude measure of effective sample size,
#> and R_hat is the potential scale reduction factor on split chains (at 
#> convergence, R_hat=1).
# }