Run CmdStan's stansummary and diagnose

Run CmdStan's stansummary and diagnose utilities. These are documented in the CmdStan Guide:

https://mc-stan.org/docs/cmdstan-guide/stansummary.html

https://mc-stan.org/docs/cmdstan-guide/diagnose.html

Note

Although these methods also work for models fit using the $variational() method, much of the output is only relevant for models fit using the $sample() method.

Usage

$cmdstan_summary()
$cmdstan_diagnose()

See also

Examples

# \dontrun{
fit <- cmdstanr_example("logistic")

#> Compiling Stan program...

fit$cmdstan_diagnose()

#> Running bin/diagnose \
#>   /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpAcsuBs/logistic-202008031245-1-d8196b.csv \
#>   /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpAcsuBs/logistic-202008031245-2-d8196b.csv \
#>   /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpAcsuBs/logistic-202008031245-3-d8196b.csv \
#>   /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpAcsuBs/logistic-202008031245-4-d8196b.csv
#> Processing csv files: /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpAcsuBs/logistic-202008031245-1-d8196b.csv, /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpAcsuBs/logistic-202008031245-2-d8196b.csv, /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpAcsuBs/logistic-202008031245-3-d8196b.csv, /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpAcsuBs/logistic-202008031245-4-d8196b.csv
#> 
#> Checking sampler transitions treedepth.
#> Treedepth satisfactory for all transitions.
#> 
#> Checking sampler transitions for divergences.
#> No divergent transitions found.
#> 
#> Checking E-BFMI - sampler transitions HMC potential energy.
#> E-BFMI satisfactory for all transitions.
#> 
#> Effective sample size satisfactory.
#> 
#> Split R-hat values satisfactory all parameters.
#> 
#> Processing complete, no problems detected.

fit$cmdstan_summary()

#> Running bin/stansummary \
#>   /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpAcsuBs/logistic-202008031245-1-d8196b.csv \
#>   /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpAcsuBs/logistic-202008031245-2-d8196b.csv \
#>   /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpAcsuBs/logistic-202008031245-3-d8196b.csv \
#>   /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpAcsuBs/logistic-202008031245-4-d8196b.csv
#> Input files: /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpAcsuBs/logistic-202008031245-1-d8196b.csv, /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpAcsuBs/logistic-202008031245-2-d8196b.csv, /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpAcsuBs/logistic-202008031245-3-d8196b.csv, /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpAcsuBs/logistic-202008031245-4-d8196b.csv
#> Inference for Stan model: logistic_model
#> 4 chains: each with iter=(1000,1000,1000,1000); warmup=(0,0,0,0); thin=(1,1,1,1); 4000 iterations saved.
#> 
#> Warmup took (0.035, 0.030, 0.031, 0.030) seconds, 0.13 seconds total
#> Sampling took (0.043, 0.042, 0.040, 0.044) seconds, 0.17 seconds total
#> 
#>                  Mean     MCSE  StdDev     5%    50%    95%    N_Eff  N_Eff/s    R_hat
#> 
#> lp__              -66  3.1e-02     1.4    -69    -66    -64     2038    12058      1.0
#> accept_stat__    0.92  1.3e-03   0.092   0.73   0.95    1.0  5.3e+03  3.1e+04  1.0e+00
#> stepsize__       0.69  1.4e-02   0.019   0.66   0.70   0.71  2.0e+00  1.2e+01  1.1e+13
#> treedepth__       2.5  1.0e-02    0.52    2.0    3.0    3.0  2.7e+03  1.6e+04  1.0e+00
#> n_leapfrog__      5.7  4.8e-02     1.9    3.0    7.0    7.0  1.6e+03  9.7e+03  1.0e+00
#> divergent__      0.00      nan    0.00   0.00   0.00   0.00      nan      nan      nan
#> energy__           68  4.8e-02     2.0     65     68     72  1.7e+03  1.0e+04  1.0e+00
#> 
#> alpha            0.38  3.5e-03    0.22  0.023   0.38   0.75     3944    23336     1.00
#> beta[1]         -0.66  4.1e-03    0.25   -1.1  -0.66  -0.26     3786    22403      1.0
#> beta[2]         -0.27  3.5e-03    0.23  -0.65  -0.26   0.10     4331    25630     1.00
#> beta[3]          0.67  4.1e-03    0.26   0.25   0.67    1.1     3982    23564     1.00
#> 
#> Samples were drawn using hmc with nuts.
#> For each parameter, N_Eff is a crude measure of effective sample size,
#> and R_hat is the potential scale reduction factor on split chains (at 
#> convergence, R_hat=1).

# }