Run CmdStan's bin/stansummary and bin/diagnose

Run CmdStan's bin/stansummary and bin/diagnose

Note

Although these methods also work for models fit using the $variational() method, much of the output is only relevant for models fit using the $sample() method.

Usage

$cmdstan_summary()
$cmdstan_diagnose()

See also

Examples

# \dontrun{
fit <- cmdstanr_example("logistic")

#> Model executable is up to date!

fit$cmdstan_diagnose()

#> Running bin/diagnose \
#>   /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpvSr5hc/logistic-202007281357-1-e77e36.csv \
#>   /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpvSr5hc/logistic-202007281357-2-e77e36.csv \
#>   /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpvSr5hc/logistic-202007281357-3-e77e36.csv \
#>   /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpvSr5hc/logistic-202007281357-4-e77e36.csv
#> Processing csv files: /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpvSr5hc/logistic-202007281357-1-e77e36.csv, /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpvSr5hc/logistic-202007281357-2-e77e36.csv, /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpvSr5hc/logistic-202007281357-3-e77e36.csv, /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpvSr5hc/logistic-202007281357-4-e77e36.csv
#> 
#> Checking sampler transitions treedepth.
#> Treedepth satisfactory for all transitions.
#> 
#> Checking sampler transitions for divergences.
#> No divergent transitions found.
#> 
#> Checking E-BFMI - sampler transitions HMC potential energy.
#> E-BFMI satisfactory for all transitions.
#> 
#> Effective sample size satisfactory.
#> 
#> Split R-hat values satisfactory all parameters.
#> 
#> Processing complete, no problems detected.

fit$cmdstan_summary()

#> Running bin/stansummary \
#>   /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpvSr5hc/logistic-202007281357-1-e77e36.csv \
#>   /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpvSr5hc/logistic-202007281357-2-e77e36.csv \
#>   /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpvSr5hc/logistic-202007281357-3-e77e36.csv \
#>   /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpvSr5hc/logistic-202007281357-4-e77e36.csv
#> Input files: /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpvSr5hc/logistic-202007281357-1-e77e36.csv, /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpvSr5hc/logistic-202007281357-2-e77e36.csv, /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpvSr5hc/logistic-202007281357-3-e77e36.csv, /var/folders/h6/14xy_35x4wd2tz542dn0qhtc0000gn/T/RtmpvSr5hc/logistic-202007281357-4-e77e36.csv
#> Inference for Stan model: logistic_model
#> 4 chains: each with iter=(1000,1000,1000,1000); warmup=(0,0,0,0); thin=(1,1,1,1); 4000 iterations saved.
#> 
#> Warmup took (0.032, 0.032, 0.033, 0.030) seconds, 0.13 seconds total
#> Sampling took (0.037, 0.053, 0.038, 0.039) seconds, 0.17 seconds total
#> 
#>                  Mean     MCSE  StdDev     5%    50%    95%    N_Eff  N_Eff/s    R_hat
#> 
#> lp__              -66  3.2e-02     1.5    -69    -66    -64     2091    12523      1.0
#> accept_stat__    0.90  1.6e-03    0.12   0.66   0.94    1.0  5.5e+03  3.3e+04  1.0e+00
#> stepsize__       0.78  1.3e-02   0.018   0.75   0.79   0.79  2.0e+00  1.2e+01  5.7e+12
#> treedepth__       2.4  8.8e-03    0.52    2.0    2.0    3.0  3.5e+03  2.1e+04  1.0e+00
#> n_leapfrog__      5.0  3.3e-02     2.0    3.0    7.0    7.0  3.7e+03  2.2e+04  1.0e+00
#> divergent__      0.00      nan    0.00   0.00   0.00   0.00      nan      nan      nan
#> energy__           68  4.8e-02     2.0     65     68     72  1.8e+03  1.1e+04  1.0e+00
#> 
#> alpha            0.38  3.3e-03    0.22  0.024   0.39   0.74     4269    25564     1.00
#> beta[1]         -0.67  3.9e-03    0.25   -1.1  -0.66  -0.27     4232    25342     1.00
#> beta[2]         -0.27  3.5e-03    0.22  -0.64  -0.27   0.10     4198    25140      1.0
#> beta[3]          0.68  4.3e-03    0.27   0.23   0.67    1.1     3983    23851      1.0
#> 
#> Samples were drawn using hmc with nuts.
#> For each parameter, N_Eff is a crude measure of effective sample size,
#> and R_hat is the potential scale reduction factor on split chains (at 
#> convergence, R_hat=1).

# }