Run CmdStan's stansummary and diagnose utilities

Run CmdStan's stansummary and diagnose utilities. These are documented in the CmdStan Guide:

https://mc-stan.org/docs/cmdstan-guide/stansummary.html

https://mc-stan.org/docs/cmdstan-guide/diagnose.html

Although these methods can be used for models fit using the $variational() method, much of the output is currently only relevant for models fit using the $sample() method.

See the $summary() for computing similar summaries in R rather than calling CmdStan's utilites.

cmdstan_summary(flags = NULL)

cmdstan_diagnose()

Arguments

flags
An optional character vector of flags (e.g. flags = c("--sig_figs=1")).

See also

Examples

# \dontrun{
fit <- cmdstanr_example("logistic")
fit$cmdstan_diagnose()

#> Processing csv files: /var/folders/s0/zfzm55px2nd2v__zlw5xfj2h0000gn/T/RtmpmzUYEz/logistic-202203181225-1-61aacb.csv, /var/folders/s0/zfzm55px2nd2v__zlw5xfj2h0000gn/T/RtmpmzUYEz/logistic-202203181225-2-61aacb.csv, /var/folders/s0/zfzm55px2nd2v__zlw5xfj2h0000gn/T/RtmpmzUYEz/logistic-202203181225-3-61aacb.csv, /var/folders/s0/zfzm55px2nd2v__zlw5xfj2h0000gn/T/RtmpmzUYEz/logistic-202203181225-4-61aacb.csv
#> 
#> Checking sampler transitions treedepth.
#> Treedepth satisfactory for all transitions.
#> 
#> Checking sampler transitions for divergences.
#> No divergent transitions found.
#> 
#> Checking E-BFMI - sampler transitions HMC potential energy.
#> E-BFMI satisfactory.
#> 
#> Effective sample size satisfactory.
#> 
#> Split R-hat values satisfactory all parameters.
#> 
#> Processing complete, no problems detected.

fit$cmdstan_summary()

#> Inference for Stan model: logistic_model
#> 4 chains: each with iter=(1000,1000,1000,1000); warmup=(0,0,0,0); thin=(1,1,1,1); 4000 iterations saved.
#> 
#> Warmup took (0.034, 0.026, 0.028, 0.048) seconds, 0.14 seconds total
#> Sampling took (0.095, 0.084, 0.16, 0.12) seconds, 0.46 seconds total
#> 
#>                     Mean     MCSE  StdDev     5%       50%       95%    N_Eff  N_Eff/s    R_hat
#> 
#> lp__            -6.6e+01  3.6e-02     1.5    -69  -6.6e+01  -6.4e+01     1736     3791      1.0
#> accept_stat__       0.90  3.8e-03    0.11   0.67      0.94       1.0  8.5e+02  1.9e+03  1.0e+00
#> stepsize__          0.75  3.0e-02   0.043   0.68      0.78      0.79  2.0e+00  4.4e+00  1.6e+13
#> treedepth__          2.4  6.7e-02    0.53    2.0       2.0       3.0  6.2e+01  1.4e+02  1.0e+00
#> n_leapfrog__         5.2  2.6e-01     2.0    3.0       7.0       7.0  6.0e+01  1.3e+02  1.0e+00
#> divergent__         0.00      nan    0.00   0.00      0.00      0.00      nan      nan      nan
#> energy__              68  5.5e-02     2.1     65        68        72  1.4e+03  3.1e+03  1.0e+00
#> 
#> alpha            3.8e-01  3.3e-03    0.22  0.026   3.8e-01   7.5e-01     4388     9581     1.00
#> beta[1]         -6.7e-01  3.9e-03    0.25   -1.1  -6.6e-01  -2.6e-01     4345     9487     1.00
#> beta[2]         -2.7e-01  3.6e-03    0.23  -0.66  -2.7e-01   1.1e-01     4144     9048     1.00
#> beta[3]          6.8e-01  4.2e-03    0.27   0.26   6.7e-01   1.1e+00     3965     8656      1.0
#> log_lik[1]      -5.1e-01  1.5e-03    0.10  -0.69  -5.1e-01  -3.6e-01     4462     9742     1.00
#> log_lik[2]      -4.0e-01  2.2e-03    0.15  -0.68  -3.9e-01  -1.9e-01     4523     9876     1.00
#> log_lik[3]      -4.9e-01  3.3e-03    0.22  -0.91  -4.6e-01  -2.0e-01     4388     9582     1.00
#> log_lik[4]      -4.5e-01  2.4e-03    0.15  -0.72  -4.3e-01  -2.4e-01     3964     8655     1.00
#> log_lik[5]      -1.2e+00  4.2e-03    0.28   -1.7  -1.2e+00  -7.6e-01     4456     9730     1.00
#> log_lik[6]      -5.9e-01  2.9e-03    0.19  -0.94  -5.7e-01  -3.2e-01     4286     9359      1.0
#> log_lik[7]      -6.4e-01  1.8e-03    0.12  -0.85  -6.4e-01  -4.6e-01     4658    10170     1.00
#> log_lik[8]      -2.8e-01  2.3e-03    0.14  -0.54  -2.5e-01  -1.0e-01     3596     7851     1.00
#> log_lik[9]      -6.9e-01  2.5e-03    0.17  -0.99  -6.8e-01  -4.5e-01     4349     9495      1.0
#> log_lik[10]     -7.3e-01  3.7e-03    0.24   -1.2  -7.1e-01  -3.9e-01     4199     9168     1.00
#> log_lik[11]     -2.8e-01  2.0e-03    0.12  -0.51  -2.6e-01  -1.1e-01     3817     8334      1.0
#> log_lik[12]     -5.0e-01  3.6e-03    0.24  -0.96  -4.6e-01  -1.8e-01     4375     9553      1.0
#> log_lik[13]     -6.6e-01  3.0e-03    0.20   -1.0  -6.4e-01  -3.6e-01     4601    10046      1.0
#> log_lik[14]     -3.6e-01  2.7e-03    0.18  -0.69  -3.3e-01  -1.3e-01     4218     9210     1.00
#> log_lik[15]     -2.8e-01  1.6e-03    0.11  -0.47  -2.6e-01  -1.3e-01     4392     9589     1.00
#> log_lik[16]     -2.8e-01  1.4e-03   0.088  -0.43  -2.7e-01  -1.5e-01     3712     8105     1.00
#> log_lik[17]     -1.6e+00  5.0e-03    0.30   -2.1  -1.6e+00  -1.1e+00     3489     7618     1.00
#> log_lik[18]     -4.8e-01  1.7e-03    0.11  -0.67  -4.7e-01  -3.2e-01     3930     8582     1.00
#> log_lik[19]     -2.3e-01  1.2e-03   0.077  -0.37  -2.2e-01  -1.2e-01     4012     8761     1.00
#> log_lik[20]     -1.1e-01  1.2e-03   0.079  -0.27  -9.4e-02  -2.7e-02     4074     8896     1.00
#> log_lik[21]     -2.1e-01  1.4e-03   0.088  -0.37  -2.0e-01  -9.2e-02     3715     8110      1.0
#> log_lik[22]     -5.7e-01  2.3e-03    0.15  -0.83  -5.5e-01  -3.5e-01     4001     8736      1.0
#> log_lik[23]     -3.3e-01  2.3e-03    0.14  -0.59  -3.1e-01  -1.4e-01     3755     8198     1.00
#> log_lik[24]     -1.4e-01  1.1e-03   0.068  -0.26  -1.2e-01  -5.2e-02     3965     8657     1.00
#> log_lik[25]     -4.6e-01  1.8e-03    0.12  -0.67  -4.5e-01  -2.8e-01     4284     9354     1.00
#> log_lik[26]     -1.5e+00  5.6e-03    0.35   -2.1  -1.5e+00  -9.9e-01     4011     8759     1.00
#> log_lik[27]     -3.1e-01  2.0e-03    0.12  -0.53  -2.9e-01  -1.4e-01     3766     8224      1.0
#> log_lik[28]     -4.4e-01  1.3e-03   0.085  -0.59  -4.4e-01  -3.1e-01     4028     8795     1.00
#> log_lik[29]     -7.3e-01  3.4e-03    0.23   -1.1  -7.1e-01  -4.0e-01     4634    10119     1.00
#> log_lik[30]     -7.0e-01  3.1e-03    0.20   -1.0  -6.8e-01  -4.1e-01     3998     8729     1.00
#> log_lik[31]     -4.9e-01  2.6e-03    0.16  -0.78  -4.7e-01  -2.6e-01     4003     8741      1.0
#> log_lik[32]     -4.2e-01  1.6e-03    0.11  -0.61  -4.1e-01  -2.6e-01     4299     9385      1.0
#> log_lik[33]     -4.1e-01  1.9e-03    0.13  -0.63  -4.0e-01  -2.3e-01     4377     9558     1.00
#> log_lik[34]     -6.4e-02  8.5e-04   0.052  -0.16  -5.0e-02  -1.3e-02     3746     8179     1.00
#> log_lik[35]     -5.9e-01  2.6e-03    0.18  -0.91  -5.7e-01  -3.3e-01     4964    10838     1.00
#> log_lik[36]     -3.3e-01  2.1e-03    0.13  -0.58  -3.1e-01  -1.4e-01     4081     8911     1.00
#> log_lik[37]     -6.9e-01  3.3e-03    0.23   -1.1  -6.7e-01  -3.7e-01     4588    10017     1.00
#> log_lik[38]     -3.2e-01  2.3e-03    0.15  -0.59  -2.9e-01  -1.2e-01     4002     8738     1.00
#> log_lik[39]     -1.8e-01  1.7e-03    0.11  -0.39  -1.6e-01  -4.9e-02     4239     9255     1.00
#> log_lik[40]     -6.8e-01  1.9e-03    0.13  -0.90  -6.7e-01  -4.9e-01     4339     9474     1.00
#> log_lik[41]     -1.1e+00  4.2e-03    0.25   -1.6  -1.1e+00  -7.6e-01     3742     8171      1.0
#> log_lik[42]     -9.3e-01  3.0e-03    0.20   -1.3  -9.1e-01  -6.4e-01     4374     9549     1.00
#> log_lik[43]     -4.1e-01  3.8e-03    0.27  -0.92  -3.5e-01  -1.0e-01     4940    10786     1.00
#> log_lik[44]     -1.2e+00  3.0e-03    0.19   -1.5  -1.2e+00  -8.9e-01     4045     8831     1.00
#> log_lik[45]     -3.6e-01  1.9e-03    0.12  -0.58  -3.5e-01  -1.8e-01     4315     9421     1.00
#> log_lik[46]     -5.8e-01  1.9e-03    0.13  -0.81  -5.8e-01  -3.9e-01     4621    10089     1.00
#> log_lik[47]     -3.1e-01  2.0e-03    0.13  -0.54  -2.9e-01  -1.4e-01     4005     8745     1.00
#> log_lik[48]     -3.2e-01  1.3e-03   0.085  -0.48  -3.2e-01  -2.0e-01     4261     9303     1.00
#> log_lik[49]     -3.2e-01  1.3e-03   0.080  -0.46  -3.1e-01  -2.0e-01     3930     8581     1.00
#> log_lik[50]     -1.3e+00  5.2e-03    0.34   -1.9  -1.3e+00  -7.9e-01     4358     9516     1.00
#> log_lik[51]     -2.9e-01  1.4e-03   0.097  -0.47  -2.8e-01  -1.5e-01     4624    10096     1.00
#> log_lik[52]     -8.4e-01  2.2e-03    0.15   -1.1  -8.3e-01  -6.1e-01     4377     9558     1.00
#> log_lik[53]     -4.0e-01  2.1e-03    0.13  -0.63  -3.9e-01  -2.1e-01     3841     8387      1.0
#> log_lik[54]     -3.7e-01  2.2e-03    0.15  -0.63  -3.5e-01  -1.7e-01     4317     9426     1.00
#> log_lik[55]     -3.9e-01  2.0e-03    0.13  -0.63  -3.7e-01  -1.9e-01     4429     9671     1.00
#> log_lik[56]     -3.2e-01  3.0e-03    0.19  -0.70  -2.8e-01  -9.1e-02     4073     8894     1.00
#> log_lik[57]     -6.5e-01  1.7e-03    0.12  -0.86  -6.5e-01  -4.7e-01     4496     9816     1.00
#> log_lik[58]     -9.4e-01  4.9e-03    0.35   -1.6  -9.0e-01  -4.5e-01     4920    10743     1.00
#> log_lik[59]     -1.4e+00  5.3e-03    0.34   -2.0  -1.3e+00  -8.6e-01     4125     9007     1.00
#> log_lik[60]     -9.8e-01  2.5e-03    0.16   -1.3  -9.7e-01  -7.3e-01     4281     9348     1.00
#> log_lik[61]     -5.4e-01  1.5e-03    0.10  -0.72  -5.3e-01  -3.9e-01     4572     9982     1.00
#> log_lik[62]     -8.8e-01  4.9e-03    0.32   -1.5  -8.5e-01  -4.2e-01     4308     9407      1.0
#> log_lik[63]     -1.2e-01  1.2e-03   0.074  -0.26  -1.0e-01  -3.2e-02     3584     7826     1.00
#> log_lik[64]     -9.0e-01  4.0e-03    0.25   -1.4  -8.7e-01  -5.4e-01     4099     8950     1.00
#> log_lik[65]     -2.0e+00  9.8e-03    0.59   -3.1  -2.0e+00  -1.1e+00     3642     7952      1.0
#> log_lik[66]     -5.0e-01  1.9e-03    0.13  -0.74  -4.9e-01  -3.1e-01     4752    10376     1.00
#> log_lik[67]     -2.8e-01  1.3e-03   0.083  -0.43  -2.7e-01  -1.6e-01     4034     8808     1.00
#> log_lik[68]     -1.1e+00  3.9e-03    0.24   -1.5  -1.0e+00  -7.1e-01     3761     8211     1.00
#> log_lik[69]     -4.3e-01  1.3e-03   0.084  -0.58  -4.3e-01  -3.1e-01     4219     9213     1.00
#> log_lik[70]     -6.4e-01  3.7e-03    0.24   -1.1  -6.1e-01  -3.0e-01     4278     9341     1.00
#> log_lik[71]     -6.1e-01  2.9e-03    0.20  -0.98  -5.9e-01  -3.2e-01     4865    10623      1.0
#> log_lik[72]     -4.6e-01  2.5e-03    0.17  -0.76  -4.4e-01  -2.2e-01     4453     9724     1.00
#> log_lik[73]     -1.5e+00  5.7e-03    0.36   -2.1  -1.4e+00  -9.3e-01     3935     8591     1.00
#> log_lik[74]     -9.5e-01  3.0e-03    0.20   -1.3  -9.4e-01  -6.6e-01     4389     9583     1.00
#> log_lik[75]     -1.1e+00  6.1e-03    0.40   -1.9  -1.1e+00  -5.7e-01     4353     9505      1.0
#> log_lik[76]     -3.7e-01  2.2e-03    0.14  -0.62  -3.5e-01  -1.7e-01     4299     9387     1.00
#> log_lik[77]     -8.8e-01  2.1e-03    0.14   -1.1  -8.7e-01  -6.7e-01     4303     9396      1.0
#> log_lik[78]     -4.9e-01  2.6e-03    0.18  -0.82  -4.6e-01  -2.4e-01     4716    10297     1.00
#> log_lik[79]     -7.6e-01  2.9e-03    0.19   -1.1  -7.5e-01  -4.8e-01     4325     9444     1.00
#> log_lik[80]     -5.4e-01  2.9e-03    0.20  -0.91  -5.2e-01  -2.6e-01     4653    10160     1.00
#> log_lik[81]     -1.6e-01  1.6e-03   0.099  -0.35  -1.4e-01  -4.9e-02     3852     8410     1.00
#> log_lik[82]     -2.2e-01  2.1e-03    0.14  -0.49  -1.9e-01  -6.1e-02     4220     9215     1.00
#> log_lik[83]     -3.4e-01  1.3e-03   0.083  -0.49  -3.4e-01  -2.2e-01     4202     9174     1.00
#> log_lik[84]     -2.7e-01  1.4e-03   0.092  -0.44  -2.6e-01  -1.4e-01     4250     9280     1.00
#> log_lik[85]     -1.3e-01  1.2e-03   0.075  -0.27  -1.1e-01  -4.0e-02     4072     8891     1.00
#> log_lik[86]     -1.1e+00  5.0e-03    0.32   -1.7  -1.1e+00  -6.5e-01     4140     9040     1.00
#> log_lik[87]     -8.3e-01  2.0e-03    0.13   -1.0  -8.2e-01  -6.3e-01     4293     9373     1.00
#> log_lik[88]     -7.8e-01  3.5e-03    0.24   -1.2  -7.5e-01  -4.3e-01     4503     9832      1.0
#> log_lik[89]     -1.3e+00  4.9e-03    0.32   -1.9  -1.2e+00  -8.1e-01     4104     8960     1.00
#> log_lik[90]     -2.6e-01  2.1e-03    0.13  -0.52  -2.4e-01  -9.4e-02     4118     8991     1.00
#> log_lik[91]     -3.9e-01  1.9e-03    0.13  -0.61  -3.7e-01  -2.1e-01     4368     9537     1.00
#> log_lik[92]     -1.5e+00  5.4e-03    0.34   -2.1  -1.5e+00  -9.8e-01     4061     8866     1.00
#> log_lik[93]     -7.5e-01  3.6e-03    0.23   -1.2  -7.2e-01  -4.2e-01     4046     8835     1.00
#> log_lik[94]     -3.2e-01  1.4e-03   0.089  -0.47  -3.1e-01  -1.9e-01     4234     9244     1.00
#> log_lik[95]     -3.9e-01  1.8e-03    0.11  -0.58  -3.8e-01  -2.2e-01     3790     8276      1.0
#> log_lik[96]     -1.6e+00  4.8e-03    0.29   -2.1  -1.6e+00  -1.1e+00     3777     8248     1.00
#> log_lik[97]     -4.3e-01  1.6e-03    0.10  -0.62  -4.2e-01  -2.7e-01     4262     9306     1.00
#> log_lik[98]     -1.0e+00  5.2e-03    0.36   -1.7  -1.0e+00  -5.2e-01     4933    10771     1.00
#> log_lik[99]     -6.9e-01  2.1e-03    0.14  -0.94  -6.9e-01  -4.8e-01     4361     9523     1.00
#> log_lik[100]    -3.9e-01  1.5e-03    0.10  -0.57  -3.8e-01  -2.4e-01     4557     9949     1.00
#> 
#> Samples were drawn using hmc with nuts.
#> For each parameter, N_Eff is a crude measure of effective sample size,
#> and R_hat is the potential scale reduction factor on split chains (at 
#> convergence, R_hat=1).

# }