Run CmdStan's stansummary and diagnose utilities

Run CmdStan's stansummary and diagnose utilities. These are documented in the CmdStan Guide:

https://mc-stan.org/docs/cmdstan-guide/stansummary.html

https://mc-stan.org/docs/cmdstan-guide/diagnose.html

Although these methods can be used for models fit using the $variational() method, much of the output is currently only relevant for models fit using the $sample() method.

See the $summary() for computing similar summaries in R rather than calling CmdStan's utilites.

cmdstan_summary(flags = NULL)

cmdstan_diagnose()

Arguments

flags

An optional character vector of flags (e.g. flags = c("--sig_figs=1")).

See also

Examples

# \dontrun{
fit <- cmdstanr_example("logistic")
fit$cmdstan_diagnose()
#> Processing csv files: /var/folders/s0/zfzm55px2nd2v__zlw5xfj2h0000gn/T/RtmpFBtN6X/logistic-202307251435-1-7b6c9f.csv, /var/folders/s0/zfzm55px2nd2v__zlw5xfj2h0000gn/T/RtmpFBtN6X/logistic-202307251435-2-7b6c9f.csv, /var/folders/s0/zfzm55px2nd2v__zlw5xfj2h0000gn/T/RtmpFBtN6X/logistic-202307251435-3-7b6c9f.csv, /var/folders/s0/zfzm55px2nd2v__zlw5xfj2h0000gn/T/RtmpFBtN6X/logistic-202307251435-4-7b6c9f.csv
#> 
#> Checking sampler transitions treedepth.
#> Treedepth satisfactory for all transitions.
#> 
#> Checking sampler transitions for divergences.
#> No divergent transitions found.
#> 
#> Checking E-BFMI - sampler transitions HMC potential energy.
#> E-BFMI satisfactory.
#> 
#> Effective sample size satisfactory.
#> 
#> Split R-hat values satisfactory all parameters.
#> 
#> Processing complete, no problems detected.
fit$cmdstan_summary()
#> Inference for Stan model: logistic_model
#> 4 chains: each with iter=(1000,1000,1000,1000); warmup=(0,0,0,0); thin=(1,1,1,1); 4000 iterations saved.
#> 
#> Warmup took (0.022, 0.022, 0.022, 0.021) seconds, 0.087 seconds total
#> Sampling took (0.069, 0.072, 0.068, 0.066) seconds, 0.28 seconds total
#> 
#>                     Mean     MCSE  StdDev     5%       50%       95%  N_Eff  N_Eff/s    R_hat
#> 
#> lp__            -6.6e+01  3.3e-02     1.4    -69  -6.6e+01  -6.4e+01   1875     6818      1.0
#> accept_stat__       0.91  1.3e-03   0.099   0.71      0.95       1.0   5487    19952  1.0e+00
#> stepsize__          0.75  2.8e-02   0.040   0.69      0.76      0.80    2.0      7.3  1.2e+13
#> treedepth__          2.4  9.4e-03    0.53    2.0       2.0       3.0   3168    11520  1.0e+00
#> n_leapfrog__         5.3  3.3e-02     2.0    3.0       7.0       7.0   3698    13446  1.0e+00
#> divergent__         0.00      nan    0.00   0.00      0.00      0.00    nan      nan      nan
#> energy__              68  5.2e-02     2.0     65        68        72   1501     5457  1.0e+00
#> 
#> alpha            3.8e-01  3.3e-03    0.22  0.031   3.8e-01   7.3e-01   4281    15567     1.00
#> beta[1]         -6.6e-01  3.9e-03    0.25   -1.1  -6.6e-01  -2.5e-01   4206    15294     1.00
#> beta[2]         -2.8e-01  3.7e-03    0.23  -0.65  -2.8e-01   1.0e-01   3812    13860      1.0
#> beta[3]          6.8e-01  4.4e-03    0.27   0.25   6.8e-01   1.1e+00   3764    13688     1.00
#> log_lik[1]      -5.2e-01  1.5e-03   0.097  -0.69  -5.1e-01  -3.7e-01   4137    15045      1.0
#> log_lik[2]      -4.0e-01  2.4e-03    0.15  -0.69  -3.8e-01  -2.0e-01   4070    14798     1.00
#> log_lik[3]      -5.0e-01  3.5e-03    0.22  -0.90  -4.7e-01  -2.1e-01   3974    14450      1.0
#> log_lik[4]      -4.5e-01  2.4e-03    0.15  -0.72  -4.3e-01  -2.4e-01   3841    13967     1.00
#> log_lik[5]      -1.2e+00  4.3e-03    0.28   -1.7  -1.2e+00  -7.6e-01   4262    15497     1.00
#> log_lik[6]      -5.9e-01  2.8e-03    0.19  -0.94  -5.7e-01  -3.1e-01   4639    16869     1.00
#> log_lik[7]      -6.4e-01  1.8e-03    0.13  -0.86  -6.3e-01  -4.5e-01   4605    16746     1.00
#> log_lik[8]      -2.8e-01  2.3e-03    0.14  -0.54  -2.5e-01  -1.1e-01   3527    12826      1.0
#> log_lik[9]      -6.9e-01  2.6e-03    0.17  -0.99  -6.8e-01  -4.3e-01   4522    16442     1.00
#> log_lik[10]     -7.4e-01  3.7e-03    0.23   -1.2  -7.2e-01  -4.1e-01   3966    14423      1.0
#> log_lik[11]     -2.8e-01  2.1e-03    0.13  -0.52  -2.6e-01  -1.1e-01   3682    13390      1.0
#> log_lik[12]     -5.0e-01  3.9e-03    0.24  -0.93  -4.5e-01  -1.9e-01   3842    13973     1.00
#> log_lik[13]     -6.6e-01  3.1e-03    0.21   -1.0  -6.4e-01  -3.6e-01   4585    16673     1.00
#> log_lik[14]     -3.7e-01  2.8e-03    0.17  -0.70  -3.4e-01  -1.4e-01   3858    14029      1.0
#> log_lik[15]     -2.8e-01  1.7e-03    0.11  -0.48  -2.7e-01  -1.4e-01   3977    14463      1.0
#> log_lik[16]     -2.8e-01  1.5e-03   0.087  -0.44  -2.6e-01  -1.5e-01   3495    12708      1.0
#> log_lik[17]     -1.6e+00  5.1e-03    0.29   -2.1  -1.6e+00  -1.1e+00   3184    11579      1.0
#> log_lik[18]     -4.8e-01  1.7e-03    0.11  -0.68  -4.7e-01  -3.1e-01   4019    14615      1.0
#> log_lik[19]     -2.3e-01  1.3e-03   0.075  -0.37  -2.2e-01  -1.3e-01   3432    12479      1.0
#> log_lik[20]     -1.1e-01  1.3e-03   0.080  -0.27  -9.4e-02  -2.9e-02   3626    13186      1.0
#> log_lik[21]     -2.1e-01  1.5e-03   0.088  -0.38  -2.0e-01  -9.4e-02   3506    12749      1.0
#> log_lik[22]     -5.7e-01  2.3e-03    0.15  -0.84  -5.5e-01  -3.4e-01   4293    15611      1.0
#> log_lik[23]     -3.3e-01  2.4e-03    0.14  -0.60  -3.0e-01  -1.4e-01   3628    13194      1.0
#> log_lik[24]     -1.4e-01  1.1e-03   0.066  -0.26  -1.2e-01  -5.2e-02   3504    12742      1.0
#> log_lik[25]     -4.5e-01  1.8e-03    0.12  -0.67  -4.4e-01  -2.8e-01   4234    15396     1.00
#> log_lik[26]     -1.5e+00  5.5e-03    0.34   -2.1  -1.5e+00  -1.0e+00   3760    13672      1.0
#> log_lik[27]     -3.1e-01  2.0e-03    0.12  -0.54  -2.9e-01  -1.4e-01   3688    13412      1.0
#> log_lik[28]     -4.5e-01  1.3e-03   0.083  -0.59  -4.4e-01  -3.2e-01   4007    14572      1.0
#> log_lik[29]     -7.2e-01  3.4e-03    0.22   -1.1  -7.0e-01  -4.0e-01   4367    15880     1.00
#> log_lik[30]     -6.9e-01  3.1e-03    0.20   -1.0  -6.7e-01  -4.1e-01   4146    15077      1.0
#> log_lik[31]     -4.9e-01  2.5e-03    0.17  -0.80  -4.6e-01  -2.5e-01   4260    15491      1.0
#> log_lik[32]     -4.2e-01  1.6e-03    0.11  -0.61  -4.2e-01  -2.7e-01   4253    15465     1.00
#> log_lik[33]     -4.1e-01  2.0e-03    0.13  -0.65  -4.0e-01  -2.3e-01   4207    15297     1.00
#> log_lik[34]     -6.3e-02  8.6e-04   0.051  -0.16  -4.9e-02  -1.3e-02   3530    12835     1.00
#> log_lik[35]     -5.9e-01  2.6e-03    0.18  -0.91  -5.7e-01  -3.2e-01   4846    17624     1.00
#> log_lik[36]     -3.2e-01  2.2e-03    0.13  -0.57  -3.0e-01  -1.5e-01   3788    13775      1.0
#> log_lik[37]     -7.0e-01  3.4e-03    0.22   -1.1  -6.8e-01  -3.8e-01   4249    15451      1.0
#> log_lik[38]     -3.1e-01  2.4e-03    0.15  -0.60  -2.8e-01  -1.2e-01   3822    13898     1.00
#> log_lik[39]     -1.8e-01  1.8e-03    0.11  -0.40  -1.6e-01  -5.3e-02   3734    13579      1.0
#> log_lik[40]     -6.8e-01  2.0e-03    0.13  -0.91  -6.7e-01  -4.8e-01   4407    16026     1.00
#> log_lik[41]     -1.1e+00  4.2e-03    0.26   -1.6  -1.1e+00  -7.4e-01   3931    14294      1.0
#> log_lik[42]     -9.3e-01  3.1e-03    0.20   -1.3  -9.1e-01  -6.2e-01   4217    15335     1.00
#> log_lik[43]     -4.1e-01  4.0e-03    0.27  -0.92  -3.5e-01  -1.0e-01   4495    16345     1.00
#> log_lik[44]     -1.2e+00  2.9e-03    0.18   -1.5  -1.2e+00  -8.9e-01   3801    13820      1.0
#> log_lik[45]     -3.6e-01  1.9e-03    0.12  -0.58  -3.5e-01  -1.9e-01   3946    14348      1.0
#> log_lik[46]     -5.8e-01  2.0e-03    0.13  -0.81  -5.7e-01  -3.8e-01   4290    15599     1.00
#> log_lik[47]     -3.0e-01  2.1e-03    0.13  -0.55  -2.8e-01  -1.3e-01   3827    13916     1.00
#> log_lik[48]     -3.2e-01  1.3e-03   0.082  -0.47  -3.2e-01  -2.0e-01   3733    13576      1.0
#> log_lik[49]     -3.2e-01  1.3e-03   0.079  -0.46  -3.1e-01  -2.0e-01   3527    12825      1.0
#> log_lik[50]     -1.3e+00  5.1e-03    0.33   -1.9  -1.3e+00  -8.0e-01   4204    15288     1.00
#> log_lik[51]     -2.9e-01  1.5e-03   0.094  -0.46  -2.8e-01  -1.6e-01   4007    14571     1.00
#> log_lik[52]     -8.3e-01  2.2e-03    0.14   -1.1  -8.2e-01  -6.2e-01   4159    15122     1.00
#> log_lik[53]     -4.0e-01  2.1e-03    0.13  -0.65  -3.8e-01  -2.1e-01   4104    14922      1.0
#> log_lik[54]     -3.7e-01  2.2e-03    0.14  -0.63  -3.6e-01  -1.8e-01   4025    14635      1.0
#> log_lik[55]     -3.9e-01  2.2e-03    0.14  -0.64  -3.7e-01  -2.0e-01   4074    14813     1.00
#> log_lik[56]     -3.2e-01  3.2e-03    0.19  -0.70  -2.7e-01  -9.4e-02   3722    13536      1.0
#> log_lik[57]     -6.6e-01  1.7e-03    0.12  -0.86  -6.5e-01  -4.8e-01   4340    15782     1.00
#> log_lik[58]     -9.5e-01  5.2e-03    0.35   -1.6  -9.1e-01  -4.6e-01   4377    15915     1.00
#> log_lik[59]     -1.4e+00  5.6e-03    0.35   -2.0  -1.3e+00  -8.4e-01   3916    14239     1.00
#> log_lik[60]     -9.8e-01  2.4e-03    0.16   -1.2  -9.7e-01  -7.3e-01   4048    14720     1.00
#> log_lik[61]     -5.4e-01  1.5e-03   0.097  -0.71  -5.3e-01  -3.9e-01   4267    15517     1.00
#> log_lik[62]     -8.7e-01  4.9e-03    0.32   -1.5  -8.3e-01  -4.3e-01   4186    15223     1.00
#> log_lik[63]     -1.2e-01  1.3e-03   0.074  -0.26  -9.7e-02  -3.2e-02   3420    12438      1.0
#> log_lik[64]     -9.1e-01  4.0e-03    0.26   -1.4  -8.8e-01  -5.3e-01   4216    15333      1.0
#> log_lik[65]     -2.0e+00  1.0e-02    0.60   -3.1  -2.0e+00  -1.1e+00   3486    12678     1.00
#> log_lik[66]     -5.1e-01  2.0e-03    0.13  -0.75  -5.0e-01  -3.1e-01   4559    16578     1.00
#> log_lik[67]     -2.8e-01  1.4e-03   0.081  -0.42  -2.7e-01  -1.6e-01   3503    12740      1.0
#> log_lik[68]     -1.1e+00  3.9e-03    0.24   -1.5  -1.1e+00  -7.0e-01   3902    14189      1.0
#> log_lik[69]     -4.4e-01  1.3e-03   0.082  -0.58  -4.3e-01  -3.1e-01   4099    14906      1.0
#> log_lik[70]     -6.4e-01  3.8e-03    0.24   -1.1  -6.0e-01  -3.0e-01   4023    14629      1.0
#> log_lik[71]     -6.1e-01  3.0e-03    0.21  -0.99  -5.9e-01  -3.2e-01   4800    17455     1.00
#> log_lik[72]     -4.7e-01  2.7e-03    0.18  -0.80  -4.4e-01  -2.2e-01   4327    15734     1.00
#> log_lik[73]     -1.5e+00  6.0e-03    0.37   -2.1  -1.5e+00  -9.4e-01   3679    13379     1.00
#> log_lik[74]     -9.5e-01  3.0e-03    0.19   -1.3  -9.4e-01  -6.6e-01   4126    15004      1.0
#> log_lik[75]     -1.1e+00  6.0e-03    0.39   -1.8  -1.1e+00  -5.8e-01   4255    15474     1.00
#> log_lik[76]     -3.8e-01  2.2e-03    0.14  -0.63  -3.6e-01  -1.8e-01   3886    14131      1.0
#> log_lik[77]     -8.8e-01  2.1e-03    0.14   -1.1  -8.7e-01  -6.6e-01   4420    16075     1.00
#> log_lik[78]     -4.9e-01  2.6e-03    0.17  -0.80  -4.7e-01  -2.5e-01   4490    16326     1.00
#> log_lik[79]     -7.7e-01  2.9e-03    0.19   -1.1  -7.5e-01  -4.9e-01   4123    14992     1.00
#> log_lik[80]     -5.4e-01  3.0e-03    0.20  -0.92  -5.2e-01  -2.6e-01   4287    15591     1.00
#> log_lik[81]     -1.6e-01  1.6e-03   0.099  -0.36  -1.4e-01  -4.9e-02   3918    14248     1.00
#> log_lik[82]     -2.3e-01  2.3e-03    0.14  -0.51  -1.9e-01  -6.4e-02   3867    14061     1.00
#> log_lik[83]     -3.4e-01  1.3e-03   0.080  -0.49  -3.4e-01  -2.3e-01   3809    13853      1.0
#> log_lik[84]     -2.8e-01  1.5e-03   0.090  -0.45  -2.7e-01  -1.5e-01   3815    13872      1.0
#> log_lik[85]     -1.3e-01  1.3e-03   0.075  -0.28  -1.2e-01  -4.4e-02   3631    13202      1.0
#> log_lik[86]     -1.1e+00  5.2e-03    0.33   -1.7  -1.1e+00  -6.4e-01   4016    14603     1.00
#> log_lik[87]     -8.2e-01  1.9e-03    0.13   -1.0  -8.2e-01  -6.3e-01   4341    15786     1.00
#> log_lik[88]     -7.8e-01  3.7e-03    0.25   -1.2  -7.5e-01  -4.3e-01   4546    16530     1.00
#> log_lik[89]     -1.3e+00  5.2e-03    0.32   -1.8  -1.3e+00  -8.0e-01   3869    14070     1.00
#> log_lik[90]     -2.6e-01  2.1e-03    0.13  -0.51  -2.3e-01  -9.4e-02   4086    14857     1.00
#> log_lik[91]     -3.9e-01  2.0e-03    0.13  -0.63  -3.8e-01  -2.0e-01   4332    15753     1.00
#> log_lik[92]     -1.5e+00  5.5e-03    0.34   -2.1  -1.5e+00  -9.7e-01   3795    13801      1.0
#> log_lik[93]     -7.3e-01  3.5e-03    0.22   -1.1  -7.1e-01  -4.2e-01   3875    14091     1.00
#> log_lik[94]     -3.2e-01  1.4e-03   0.086  -0.48  -3.1e-01  -1.9e-01   3797    13806      1.0
#> log_lik[95]     -3.9e-01  1.9e-03    0.11  -0.60  -3.7e-01  -2.2e-01   3692    13426      1.0
#> log_lik[96]     -1.6e+00  4.9e-03    0.28   -2.1  -1.6e+00  -1.1e+00   3247    11808      1.0
#> log_lik[97]     -4.3e-01  1.6e-03    0.10  -0.61  -4.2e-01  -2.8e-01   4122    14989      1.0
#> log_lik[98]     -1.1e+00  5.5e-03    0.37   -1.7  -1.0e+00  -5.3e-01   4388    15957     1.00
#> log_lik[99]     -7.0e-01  2.2e-03    0.15  -0.95  -6.9e-01  -4.8e-01   4345    15799     1.00
#> log_lik[100]    -3.9e-01  1.5e-03   0.096  -0.56  -3.8e-01  -2.5e-01   4119    14977      1.0
#> 
#> Samples were drawn using hmc with nuts.
#> For each parameter, N_Eff is a crude measure of effective sample size,
#> and R_hat is the potential scale reduction factor on split chains (at 
#> convergence, R_hat=1).
# }