Run CmdStan's stansummary and diagnose utilities

Run CmdStan's stansummary and diagnose utilities. These are documented in the CmdStan Guide:

stansummary

diagnose

These methods can be used for models fit using $sample(), $variational(), or $pathfinder(). Much of the output is currently only relevant for models fit using $sample(). These methods are not available for optimization, Laplace approximation, or standalone generated quantities.

See the $summary() for computing similar summaries in R rather than calling CmdStan's utilities.

Usage

cmdstan_summary(flags = NULL)

cmdstan_diagnose()

Arguments

flags

An optional character vector of flags (e.g. flags = c("--sig_figs=1")).

Value

Both methods invisibly return a list containing the command's exit status, standard output, and standard error.

See also

Examples

# \dontrun{
fit <- cmdstanr_example("logistic")
fit$cmdstan_diagnose()
#> Checking sampler transitions treedepth.
#> Treedepth satisfactory for all transitions.
#> 
#> Checking sampler transitions for divergences.
#> No divergent transitions found.
#> 
#> Checking E-BFMI - sampler transitions HMC potential energy.
#> E-BFMI satisfactory.
#> 
#> Rank-normalized split effective sample size satisfactory for all parameters.
#> 
#> Rank-normalized split R-hat values satisfactory for all parameters.
#> 
#> Processing complete, no problems detected.
fit$cmdstan_summary()
#> Inference for Stan model: logistic_model
#> 4 chains: each with iter=1000; warmup=1000; thin=1; 1000 iterations saved.
#> 
#> Warmup took (0.014, 0.014, 0.013, 0.013) seconds, 0.054 seconds total
#> Sampling took (0.040, 0.039, 0.041, 0.038) seconds, 0.16 seconds total
#> 
#>                     Mean     MCSE  StdDev    MAD     5%       50%       95%  ESS_bulk  ESS_tail  ESS_bulk/s  R_hat
#> 
#> lp__            -6.6e+01  3.2e-02     1.4    1.3    -69  -6.6e+01  -6.4e+01      1995      2391       12626    1.0
#> accept_stat__       0.90  1.5e-03    0.11  0.084   0.67      0.94       1.0      5799      3565       36701    1.0
#> stepsize__          0.76      nan   0.036  0.028   0.73      0.75      0.82       nan       nan         nan    nan
#> treedepth__          2.4  5.5e-02    0.52   0.00    2.0       2.0       3.0       239       452        1510    1.0
#> n_leapfrog__         5.1  1.6e-01     2.0   0.00    3.0       7.0       7.0       452       489        2858    1.0
#> divergent__         0.00      nan    0.00   0.00   0.00      0.00      0.00       nan       nan         nan    nan
#> energy__              68  4.8e-02     2.0    1.9     65        68        72      1807      2294       11439   1.00
#> 
#> alpha            3.8e-01  3.3e-03    0.22   0.22  0.021   3.8e-01   7.3e-01      4492      3390       28431    1.0
#> beta[1]         -6.7e-01  3.9e-03    0.25   0.25   -1.1  -6.6e-01  -2.7e-01      4137      2927       26186    1.0
#> beta[2]         -2.7e-01  3.5e-03    0.22   0.22  -0.63  -2.7e-01   9.9e-02      4039      3365       25561    1.0
#> beta[3]          6.8e-01  4.2e-03    0.27   0.27   0.25   6.8e-01   1.1e+00      4157      2896       26308    1.0
#> log_lik[1]      -5.2e-01  1.5e-03   0.099  0.100  -0.69  -5.1e-01  -3.7e-01      4313      3072       27298    1.0
#> log_lik[2]      -4.0e-01  2.3e-03    0.15   0.14  -0.67  -3.8e-01  -1.9e-01      4442      3222       28115    1.0
#> log_lik[3]      -4.9e-01  3.3e-03    0.21   0.20  -0.89  -4.6e-01  -2.1e-01      4210      3156       26644    1.0
#> log_lik[4]      -4.5e-01  2.5e-03    0.16   0.15  -0.73  -4.3e-01  -2.3e-01      3872      3079       24508    1.0
#> log_lik[5]      -1.2e+00  4.1e-03    0.28   0.27   -1.7  -1.2e+00  -7.6e-01      4670      3399       29558    1.0
#> log_lik[6]      -5.9e-01  2.9e-03    0.19   0.18  -0.94  -5.7e-01  -3.2e-01      4464      3088       28256    1.0
#> log_lik[7]      -6.4e-01  1.9e-03    0.13   0.13  -0.86  -6.3e-01  -4.5e-01      4497      3243       28461    1.0
#> log_lik[8]      -2.8e-01  2.2e-03    0.13   0.12  -0.54  -2.6e-01  -1.1e-01      4007      3056       25358    1.0
#> log_lik[9]      -6.9e-01  2.5e-03    0.17   0.16  -0.98  -6.8e-01  -4.4e-01      4379      3312       27712    1.0
#> log_lik[10]     -7.3e-01  3.7e-03    0.23   0.22   -1.2  -7.1e-01  -3.9e-01      3988      3398       25238    1.0
#> log_lik[11]     -2.8e-01  2.0e-03    0.13   0.11  -0.52  -2.6e-01  -1.2e-01      4022      3061       25453   1.00
#> log_lik[12]     -5.0e-01  3.7e-03    0.24   0.22  -0.96  -4.5e-01  -1.8e-01      4506      2995       28516    1.0
#> log_lik[13]     -6.6e-01  3.1e-03    0.21   0.20   -1.0  -6.3e-01  -3.6e-01      4474      3169       28315    1.0
#> log_lik[14]     -3.6e-01  2.7e-03    0.17   0.16  -0.67  -3.3e-01  -1.4e-01      3989      2971       25248    1.0
#> log_lik[15]     -2.8e-01  1.6e-03    0.11   0.10  -0.47  -2.6e-01  -1.4e-01      4358      3083       27585    1.0
#> log_lik[16]     -2.8e-01  1.4e-03   0.087  0.084  -0.44  -2.7e-01  -1.5e-01      3753      3113       23755    1.0
#> log_lik[17]     -1.6e+00  4.6e-03    0.29   0.28   -2.1  -1.6e+00  -1.1e+00      3884      3183       24584    1.0
#> log_lik[18]     -4.8e-01  1.6e-03    0.11   0.10  -0.67  -4.7e-01  -3.2e-01      4525      3061       28640   1.00
#> log_lik[19]     -2.3e-01  1.2e-03   0.075  0.071  -0.37  -2.2e-01  -1.3e-01      3809      3106       24110   1.00
#> log_lik[20]     -1.1e-01  1.2e-03   0.075  0.060  -0.26  -9.1e-02  -2.9e-02      3922      2993       24820    1.0
#> log_lik[21]     -2.1e-01  1.4e-03   0.088  0.081  -0.37  -2.0e-01  -9.3e-02      3746      3162       23707   1.00
#> log_lik[22]     -5.7e-01  2.3e-03    0.15   0.14  -0.83  -5.6e-01  -3.5e-01      4240      3132       26833    1.0
#> log_lik[23]     -3.3e-01  2.1e-03    0.14   0.13  -0.60  -3.1e-01  -1.5e-01      4325      3011       27375    1.0
#> log_lik[24]     -1.4e-01  1.1e-03   0.066  0.060  -0.26  -1.2e-01  -5.4e-02      3734      2957       23632    1.0
#> log_lik[25]     -4.6e-01  1.9e-03    0.12   0.13  -0.67  -4.5e-01  -2.7e-01      4141      2917       26206    1.0
#> log_lik[26]     -1.5e+00  5.0e-03    0.34   0.34   -2.1  -1.5e+00  -9.9e-01      4633      3302       29326    1.0
#> log_lik[27]     -3.1e-01  1.9e-03    0.12   0.11  -0.53  -2.9e-01  -1.4e-01      3928      3051       24860   1.00
#> log_lik[28]     -4.5e-01  1.3e-03   0.083  0.082  -0.59  -4.4e-01  -3.2e-01      4339      3397       27460   1.00
#> log_lik[29]     -7.3e-01  3.4e-03    0.23   0.22   -1.1  -7.1e-01  -3.9e-01      4692      3460       29697    1.0
#> log_lik[30]     -7.0e-01  2.8e-03    0.19   0.19   -1.0  -6.8e-01  -4.2e-01      4737      3182       29983    1.0
#> log_lik[31]     -4.9e-01  2.5e-03    0.16   0.16  -0.79  -4.7e-01  -2.6e-01      4135      3083       26170   1.00
#> log_lik[32]     -4.2e-01  1.6e-03    0.11   0.10  -0.62  -4.1e-01  -2.6e-01      4483      3299       28373    1.0
#> log_lik[33]     -4.1e-01  2.0e-03    0.13   0.13  -0.65  -4.0e-01  -2.2e-01      4193      2876       26538    1.0
#> log_lik[34]     -6.4e-02  8.6e-04   0.051  0.036  -0.17  -5.0e-02  -1.3e-02      3715      2854       23511    1.0
#> log_lik[35]     -5.9e-01  2.7e-03    0.19   0.18  -0.92  -5.7e-01  -3.2e-01      4837      3162       30612    1.0
#> log_lik[36]     -3.3e-01  2.0e-03    0.13   0.12  -0.59  -3.1e-01  -1.5e-01      4773      3196       30208    1.0
#> log_lik[37]     -6.9e-01  3.4e-03    0.23   0.22   -1.1  -6.6e-01  -3.8e-01      4445      3395       28131    1.0
#> log_lik[38]     -3.2e-01  2.5e-03    0.15   0.14  -0.60  -2.9e-01  -1.2e-01      4014      3373       25407    1.0
#> log_lik[39]     -1.8e-01  1.7e-03    0.10  0.090  -0.38  -1.5e-01  -5.5e-02      4039      3077       25561    1.0
#> log_lik[40]     -6.8e-01  1.9e-03    0.13   0.13  -0.91  -6.8e-01  -4.8e-01      4643      3303       29389    1.0
#> log_lik[41]     -1.1e+00  4.0e-03    0.25   0.25   -1.6  -1.1e+00  -7.5e-01      3982      3214       25203   1.00
#> log_lik[42]     -9.3e-01  3.0e-03    0.20   0.20   -1.3  -9.2e-01  -6.3e-01      4422      3008       27987    1.0
#> log_lik[43]     -4.1e-01  3.8e-03    0.27   0.23  -0.94  -3.5e-01  -1.1e-01      5187      3473       32828    1.0
#> log_lik[44]     -1.2e+00  2.9e-03    0.18   0.18   -1.5  -1.2e+00  -8.9e-01      4074      3247       25785    1.0
#> log_lik[45]     -3.6e-01  1.8e-03    0.12   0.11  -0.56  -3.4e-01  -1.9e-01      4017      3005       25425    1.0
#> log_lik[46]     -5.8e-01  1.9e-03    0.13   0.13  -0.81  -5.7e-01  -3.8e-01      4638      3010       29357    1.0
#> log_lik[47]     -3.1e-01  2.1e-03    0.13   0.12  -0.54  -2.9e-01  -1.3e-01      4016      2877       25417    1.0
#> log_lik[48]     -3.2e-01  1.3e-03   0.082  0.079  -0.47  -3.2e-01  -2.1e-01      4213      3166       26662   1.00
#> log_lik[49]     -3.2e-01  1.3e-03   0.079  0.077  -0.46  -3.1e-01  -2.0e-01      3892      3167       24635   1.00
#> log_lik[50]     -1.3e+00  4.7e-03    0.33   0.33   -1.9  -1.3e+00  -7.9e-01      4859      2856       30752    1.0
#> log_lik[51]     -2.9e-01  1.4e-03   0.094  0.090  -0.46  -2.8e-01  -1.6e-01      4474      2891       28317    1.0
#> log_lik[52]     -8.4e-01  2.2e-03    0.14   0.14   -1.1  -8.3e-01  -6.2e-01      4398      3275       27838    1.0
#> log_lik[53]     -4.0e-01  2.0e-03    0.13   0.12  -0.64  -3.9e-01  -2.2e-01      4017      3214       25425   1.00
#> log_lik[54]     -3.7e-01  2.2e-03    0.14   0.13  -0.63  -3.5e-01  -1.7e-01      4068      2941       25746    1.0
#> log_lik[55]     -3.8e-01  2.1e-03    0.14   0.13  -0.64  -3.7e-01  -1.9e-01      4297      3044       27195   1.00
#> log_lik[56]     -3.2e-01  2.9e-03    0.19   0.16  -0.70  -2.8e-01  -9.5e-02      5134      3161       32493    1.0
#> log_lik[57]     -6.6e-01  1.8e-03    0.12   0.12  -0.87  -6.5e-01  -4.7e-01      4387      3322       27769    1.0
#> log_lik[58]     -9.4e-01  5.2e-03    0.35   0.34   -1.6  -8.9e-01  -4.5e-01      4854      3468       30723    1.0
#> log_lik[59]     -1.4e+00  5.5e-03    0.35   0.35   -2.0  -1.3e+00  -8.4e-01      4122      2786       26089    1.0
#> log_lik[60]     -9.8e-01  2.4e-03    0.16   0.16   -1.3  -9.7e-01  -7.4e-01      4282      3001       27103    1.0
#> log_lik[61]     -5.4e-01  1.5e-03   0.099   0.10  -0.71  -5.3e-01  -3.9e-01      4565      3218       28895    1.0
#> log_lik[62]     -8.8e-01  4.8e-03    0.31   0.30   -1.5  -8.4e-01  -4.3e-01      4481      3165       28360    1.0
#> log_lik[63]     -1.2e-01  1.2e-03   0.074  0.059  -0.26  -9.9e-02  -3.3e-02      3613      3087       22868    1.0
#> log_lik[64]     -8.9e-01  3.5e-03    0.24   0.23   -1.3  -8.7e-01  -5.3e-01      4956      3141       31367    1.0
#> log_lik[65]     -2.0e+00  9.7e-03    0.60   0.60   -3.1  -2.0e+00  -1.1e+00      3887      2987       24604   1.00
#> log_lik[66]     -5.0e-01  2.0e-03    0.14   0.13  -0.76  -4.9e-01  -3.1e-01      4820      3343       30505    1.0
#> log_lik[67]     -2.8e-01  1.3e-03   0.081  0.077  -0.42  -2.7e-01  -1.6e-01      4020      3096       25442    1.0
#> log_lik[68]     -1.1e+00  3.5e-03    0.23   0.23   -1.5  -1.0e+00  -7.0e-01      4331      3031       27413    1.0
#> log_lik[69]     -4.4e-01  1.3e-03   0.083  0.082  -0.58  -4.3e-01  -3.1e-01      4237      3203       26819    1.0
#> log_lik[70]     -6.4e-01  3.3e-03    0.23   0.22   -1.1  -6.2e-01  -3.1e-01      5256      3362       33268    1.0
#> log_lik[71]     -6.1e-01  3.0e-03    0.21   0.21  -1.00  -5.9e-01  -3.1e-01      4752      3119       30076    1.0
#> log_lik[72]     -4.6e-01  2.7e-03    0.17   0.17  -0.78  -4.4e-01  -2.3e-01      4263      3143       26984    1.0
#> log_lik[73]     -1.5e+00  6.1e-03    0.38   0.38   -2.2  -1.5e+00  -9.3e-01      3956      2940       25036    1.0
#> log_lik[74]     -9.5e-01  2.9e-03    0.19   0.19   -1.3  -9.4e-01  -6.5e-01      4373      3360       27675    1.0
#> log_lik[75]     -1.1e+00  5.9e-03    0.39   0.38   -1.8  -1.1e+00  -5.7e-01      4424      3212       28000    1.0
#> log_lik[76]     -3.7e-01  2.1e-03    0.14   0.13  -0.62  -3.5e-01  -1.8e-01      4031      3005       25514    1.0
#> log_lik[77]     -8.8e-01  2.1e-03    0.14   0.13   -1.1  -8.7e-01  -6.6e-01      4505      3282       28514    1.0
#> log_lik[78]     -4.9e-01  2.5e-03    0.17   0.17  -0.81  -4.7e-01  -2.5e-01      4762      3173       30141    1.0
#> log_lik[79]     -7.6e-01  3.0e-03    0.19   0.19   -1.1  -7.4e-01  -4.7e-01      4108      3258       26002    1.0
#> log_lik[80]     -5.4e-01  2.9e-03    0.20   0.19  -0.90  -5.1e-01  -2.6e-01      4823      3206       30528    1.0
#> log_lik[81]     -1.6e-01  1.7e-03    0.10  0.087  -0.36  -1.4e-01  -4.6e-02      3910      2814       24747    1.0
#> log_lik[82]     -2.2e-01  2.2e-03    0.14   0.11  -0.49  -1.9e-01  -6.2e-02      4229      3107       26766   1.00
#> log_lik[83]     -3.4e-01  1.3e-03   0.080  0.079  -0.48  -3.4e-01  -2.2e-01      3995      3290       25282    1.0
#> log_lik[84]     -2.7e-01  1.4e-03   0.090  0.087  -0.44  -2.6e-01  -1.5e-01      4104      3191       25973    1.0
#> log_lik[85]     -1.3e-01  1.2e-03   0.072  0.063  -0.27  -1.1e-01  -4.2e-02      3856      3052       24408    1.0
#> log_lik[86]     -1.1e+00  4.8e-03    0.32   0.32   -1.7  -1.1e+00  -6.6e-01      4560      3163       28859    1.0
#> log_lik[87]     -8.2e-01  1.8e-03    0.13   0.13   -1.0  -8.1e-01  -6.2e-01      4794      3393       30342   1.00
#> log_lik[88]     -7.8e-01  3.6e-03    0.24   0.23   -1.2  -7.5e-01  -4.3e-01      4483      3091       28374    1.0
#> log_lik[89]     -1.3e+00  5.1e-03    0.32   0.33   -1.8  -1.3e+00  -8.0e-01      4222      2938       26720    1.0
#> log_lik[90]     -2.6e-01  2.2e-03    0.14   0.12  -0.53  -2.4e-01  -8.9e-02      4141      3138       26212    1.0
#> log_lik[91]     -3.9e-01  2.0e-03    0.13   0.13  -0.62  -3.7e-01  -2.0e-01      4173      3114       26411    1.0
#> log_lik[92]     -1.5e+00  5.2e-03    0.34   0.35   -2.1  -1.5e+00  -9.8e-01      4343      3186       27487   1.00
#> log_lik[93]     -7.4e-01  3.5e-03    0.22   0.21   -1.2  -7.2e-01  -4.2e-01      3960      3172       25066    1.0
#> log_lik[94]     -3.2e-01  1.4e-03   0.086  0.085  -0.47  -3.1e-01  -1.9e-01      3970      3142       25126    1.0
#> log_lik[95]     -3.9e-01  1.8e-03    0.11   0.11  -0.59  -3.8e-01  -2.3e-01      3961      3342       25070   1.00
#> log_lik[96]     -1.6e+00  4.5e-03    0.28   0.28   -2.1  -1.6e+00  -1.1e+00      3817      3116       24160    1.0
#> log_lik[97]     -4.3e-01  1.5e-03    0.10  0.097  -0.61  -4.2e-01  -2.8e-01      4946      3123       31303    1.0
#> log_lik[98]     -1.0e+00  5.5e-03    0.38   0.36   -1.7  -1.0e+00  -5.2e-01      4832      3167       30584    1.0
#> log_lik[99]     -6.9e-01  2.1e-03    0.14   0.14  -0.94  -6.8e-01  -4.8e-01      4657      3135       29472    1.0
#> log_lik[100]    -3.9e-01  1.5e-03   0.096  0.095  -0.56  -3.8e-01  -2.5e-01      4346      3423       27504    1.0
#> 
#> Samples were drawn using hmc with nuts.
#> For each parameter, ESS_bulk and ESS_tail measure the effective sample size for the entire sample (bulk) and for the .05 and .95 tails (tail), 
#> and R_hat measures the potential scale reduction on split chains. At convergence R_hat will be very close to 1.00.
# }