Run CmdStan's stansummary and diagnose utilities

Run CmdStan's stansummary and diagnose utilities. These are documented in the CmdStan Guide:

stansummary

diagnose

These methods can be used for models fit using $sample(), $variational(), or $pathfinder(). Much of the output is currently only relevant for models fit using $sample(). These methods are not available for optimization, Laplace approximation, or standalone generated quantities.

See the $summary() for computing similar summaries in R rather than calling CmdStan's utilities.

Usage

cmdstan_summary(flags = NULL)

cmdstan_diagnose()

Arguments

flags

An optional character vector of flags (e.g. flags = c("--sig_figs=1")).

Value

Both methods invisibly return a list containing the command's exit status, standard output, and standard error.

See also

Examples

# \dontrun{
fit <- cmdstanr_example("logistic")
fit$cmdstan_diagnose()
#> Checking sampler transitions treedepth.
#> Treedepth satisfactory for all transitions.
#> 
#> Checking sampler transitions for divergences.
#> No divergent transitions found.
#> 
#> Checking E-BFMI - sampler transitions HMC potential energy.
#> E-BFMI satisfactory.
#> 
#> Rank-normalized split effective sample size satisfactory for all parameters.
#> 
#> Rank-normalized split R-hat values satisfactory for all parameters.
#> 
#> Processing complete, no problems detected.
fit$cmdstan_summary()
#> Inference for Stan model: logistic_model
#> 4 chains: each with iter=1000; warmup=1000; thin=1; 1000 iterations saved.
#> 
#> Warmup took (0.020, 0.019, 0.019, 0.019) seconds, 0.077 seconds total
#> Sampling took (0.054, 0.054, 0.056, 0.058) seconds, 0.22 seconds total
#> 
#>                     Mean     MCSE  StdDev    MAD        5%       50%       95%  ESS_bulk  ESS_tail  ESS_bulk/s  R_hat
#> 
#> lp__            -6.6e+01  3.0e-02     1.4    1.3  -6.9e+01  -6.6e+01  -6.4e+01      2257      3114       10165    1.0
#> accept_stat__       0.90  1.5e-03    0.11  0.084      0.67      0.94       1.0      5881      4001       26490    1.0
#> stepsize__          0.77      nan   0.023  0.031      0.74      0.76      0.79       nan       nan         nan    nan
#> treedepth__          2.4  9.1e-03    0.52   0.00       2.0       2.0       3.0      3390      3315       15270    1.0
#> n_leapfrog__         5.1  3.5e-02     2.0   0.00       3.0       7.0       7.0      3590      3443       16170    1.0
#> divergent__         0.00      nan    0.00   0.00      0.00      0.00      0.00       nan       nan         nan    nan
#> energy__              68  4.9e-02     2.0    1.9        65        68        72      1729      2652        7790    1.0
#> 
#> alpha            3.8e-01  3.4e-03    0.22   0.22   7.9e-03   3.8e-01   7.6e-01      4346      3040       19578    1.0
#> beta[1]         -6.6e-01  3.9e-03    0.25   0.24  -1.1e+00  -6.5e-01  -2.6e-01      3956      3018       17819    1.0
#> beta[2]         -2.7e-01  3.4e-03    0.23   0.23  -6.4e-01  -2.7e-01   1.0e-01      4375      3217       19707    1.0
#> beta[3]          6.7e-01  4.0e-03    0.27   0.27   2.4e-01   6.7e-01   1.1e+00      4720      3373       21262   1.00
#> log_lik[1]      -5.2e-01  1.5e-03    0.10  0.100  -6.9e-01  -5.1e-01  -3.6e-01      4220      3303       19011    1.0
#> log_lik[2]      -4.1e-01  2.3e-03    0.15   0.15  -6.9e-01  -3.9e-01  -2.0e-01      4465      3139       20113    1.0
#> log_lik[3]      -5.0e-01  3.4e-03    0.22   0.21  -9.1e-01  -4.6e-01  -2.1e-01      4326      2959       19486    1.0
#> log_lik[4]      -4.5e-01  2.4e-03    0.15   0.15  -7.3e-01  -4.3e-01  -2.4e-01      4303      3327       19383   1.00
#> log_lik[5]      -1.2e+00  4.1e-03    0.28   0.28  -1.7e+00  -1.2e+00  -7.5e-01      5152      3141       23208    1.0
#> log_lik[6]      -5.9e-01  2.9e-03    0.19   0.19  -9.4e-01  -5.7e-01  -3.1e-01      4701      2864       21177    1.0
#> log_lik[7]      -6.4e-01  1.9e-03    0.13   0.12  -8.7e-01  -6.4e-01  -4.5e-01      4575      2934       20609    1.0
#> log_lik[8]      -2.8e-01  2.2e-03    0.14   0.12  -5.4e-01  -2.5e-01  -1.1e-01      4147      3106       18678    1.0
#> log_lik[9]      -6.9e-01  2.6e-03    0.17   0.17  -9.9e-01  -6.7e-01  -4.4e-01      4386      2688       19756    1.0
#> log_lik[10]     -7.3e-01  3.6e-03    0.23   0.23  -1.2e+00  -7.1e-01  -4.0e-01      4328      3346       19496    1.0
#> log_lik[11]     -2.8e-01  2.0e-03    0.13   0.12  -5.3e-01  -2.6e-01  -1.2e-01      4432      2653       19963    1.0
#> log_lik[12]     -5.0e-01  3.7e-03    0.25   0.23  -9.8e-01  -4.6e-01  -1.8e-01      5126      3240       23088    1.0
#> log_lik[13]     -6.6e-01  3.2e-03    0.21   0.21  -1.0e+00  -6.4e-01  -3.7e-01      4608      2755       20757    1.0
#> log_lik[14]     -3.6e-01  2.8e-03    0.17   0.16  -6.9e-01  -3.3e-01  -1.4e-01      3962      3103       17845   1.00
#> log_lik[15]     -2.8e-01  1.6e-03    0.11   0.10  -4.8e-01  -2.7e-01  -1.3e-01      4477      2992       20168   1.00
#> log_lik[16]     -2.8e-01  1.4e-03   0.090  0.087  -4.4e-01  -2.7e-01  -1.5e-01      3908      2830       17604    1.0
#> log_lik[17]     -1.6e+00  4.9e-03    0.29   0.29  -2.1e+00  -1.6e+00  -1.1e+00      3669      2780       16525   1.00
#> log_lik[18]     -4.8e-01  1.7e-03    0.11   0.11  -6.8e-01  -4.7e-01  -3.1e-01      4340      2734       19548    1.0
#> log_lik[19]     -2.3e-01  1.3e-03   0.077  0.073  -3.8e-01  -2.3e-01  -1.3e-01      3691      2822       16624   1.00
#> log_lik[20]     -1.1e-01  1.3e-03   0.078  0.060  -2.6e-01  -9.6e-02  -2.9e-02      3607      3057       16248    1.0
#> log_lik[21]     -2.1e-01  1.4e-03   0.090  0.086  -3.8e-01  -2.0e-01  -9.5e-02      4024      2928       18126   1.00
#> log_lik[22]     -5.7e-01  2.3e-03    0.15   0.15  -8.4e-01  -5.5e-01  -3.5e-01      4278      2637       19271    1.0
#> log_lik[23]     -3.3e-01  2.2e-03    0.14   0.14  -6.1e-01  -3.1e-01  -1.4e-01      4326      3090       19487    1.0
#> log_lik[24]     -1.4e-01  1.1e-03   0.067  0.059  -2.6e-01  -1.3e-01  -5.4e-02      3544      2996       15964    1.0
#> log_lik[25]     -4.6e-01  1.9e-03    0.12   0.12  -6.8e-01  -4.5e-01  -2.8e-01      4378      3273       19720   1.00
#> log_lik[26]     -1.5e+00  5.4e-03    0.35   0.34  -2.1e+00  -1.5e+00  -9.8e-01      4159      3381       18734    1.0
#> log_lik[27]     -3.1e-01  1.9e-03    0.13   0.12  -5.4e-01  -2.9e-01  -1.4e-01      4287      2769       19311    1.0
#> log_lik[28]     -4.5e-01  1.3e-03   0.086  0.085  -6.0e-01  -4.4e-01  -3.1e-01      4148      2903       18686    1.0
#> log_lik[29]     -7.3e-01  3.2e-03    0.22   0.22  -1.1e+00  -7.0e-01  -4.0e-01      5069      3495       22833    1.0
#> log_lik[30]     -6.9e-01  2.9e-03    0.20   0.19  -1.0e+00  -6.7e-01  -4.1e-01      4479      2997       20175    1.0
#> log_lik[31]     -4.9e-01  2.5e-03    0.17   0.16  -8.0e-01  -4.7e-01  -2.5e-01      4342      2890       19560    1.0
#> log_lik[32]     -4.2e-01  1.6e-03    0.11   0.11  -6.2e-01  -4.1e-01  -2.6e-01      4743      2965       21367    1.0
#> log_lik[33]     -4.2e-01  2.0e-03    0.13   0.12  -6.5e-01  -4.0e-01  -2.3e-01      4068      3133       18325    1.0
#> log_lik[34]     -6.5e-02  8.8e-04   0.052  0.035  -1.6e-01  -5.0e-02  -1.4e-02      3631      3119       16356    1.0
#> log_lik[35]     -5.9e-01  2.6e-03    0.18   0.18  -9.3e-01  -5.7e-01  -3.3e-01      5199      2985       23420   1.00
#> log_lik[36]     -3.3e-01  2.0e-03    0.13   0.13  -5.7e-01  -3.1e-01  -1.5e-01      4519      3357       20356    1.0
#> log_lik[37]     -6.9e-01  3.3e-03    0.22   0.22  -1.1e+00  -6.7e-01  -3.8e-01      4753      3358       21409    1.0
#> log_lik[38]     -3.2e-01  2.3e-03    0.15   0.14  -6.0e-01  -2.9e-01  -1.2e-01      4269      3054       19232    1.0
#> log_lik[39]     -1.8e-01  1.8e-03    0.11  0.091  -3.9e-01  -1.6e-01  -5.4e-02      3796      2972       17098   1.00
#> log_lik[40]     -6.8e-01  2.0e-03    0.13   0.13  -9.1e-01  -6.7e-01  -4.8e-01      4506      2756       20298    1.0
#> log_lik[41]     -1.1e+00  4.1e-03    0.26   0.26  -1.6e+00  -1.1e+00  -7.3e-01      4186      2859       18856    1.0
#> log_lik[42]     -9.2e-01  3.0e-03    0.20   0.19  -1.3e+00  -9.0e-01  -6.2e-01      4284      3055       19295    1.0
#> log_lik[43]     -4.1e-01  3.8e-03    0.27   0.23  -9.4e-01  -3.5e-01  -9.9e-02      5267      3190       23726    1.0
#> log_lik[44]     -1.2e+00  3.1e-03    0.19   0.19  -1.5e+00  -1.2e+00  -8.8e-01      3863      2830       17400    1.0
#> log_lik[45]     -3.6e-01  1.9e-03    0.12   0.12  -5.7e-01  -3.5e-01  -1.9e-01      3902      3066       17578   1.00
#> log_lik[46]     -5.9e-01  2.0e-03    0.13   0.13  -8.2e-01  -5.8e-01  -3.9e-01      4541      2996       20454    1.0
#> log_lik[47]     -3.1e-01  2.0e-03    0.13   0.12  -5.4e-01  -2.9e-01  -1.3e-01      4357      3468       19625   1.00
#> log_lik[48]     -3.3e-01  1.4e-03   0.085  0.080  -4.8e-01  -3.2e-01  -2.0e-01      3666      3089       16512    1.0
#> log_lik[49]     -3.2e-01  1.3e-03   0.081  0.082  -4.7e-01  -3.1e-01  -2.0e-01      3903      2529       17581    1.0
#> log_lik[50]     -1.3e+00  5.2e-03    0.34   0.33  -1.9e+00  -1.3e+00  -7.8e-01      4525      3387       20382    1.0
#> log_lik[51]     -2.9e-01  1.5e-03   0.096  0.090  -4.7e-01  -2.8e-01  -1.5e-01      4007      3124       18050    1.0
#> log_lik[52]     -8.4e-01  2.2e-03    0.15   0.14  -1.1e+00  -8.3e-01  -6.1e-01      4401      3245       19825    1.0
#> log_lik[53]     -4.0e-01  2.1e-03    0.13   0.13  -6.5e-01  -3.9e-01  -2.2e-01      4276      2848       19262    1.0
#> log_lik[54]     -3.7e-01  2.3e-03    0.14   0.14  -6.4e-01  -3.5e-01  -1.8e-01      3992      3149       17981   1.00
#> log_lik[55]     -3.9e-01  2.1e-03    0.14   0.13  -6.4e-01  -3.8e-01  -2.0e-01      4247      3196       19129    1.0
#> log_lik[56]     -3.2e-01  2.9e-03    0.19   0.17  -6.9e-01  -2.8e-01  -9.2e-02      4798      3066       21611    1.0
#> log_lik[57]     -6.5e-01  1.7e-03    0.12   0.12  -8.6e-01  -6.5e-01  -4.7e-01      4555      3006       20516    1.0
#> log_lik[58]     -9.4e-01  5.0e-03    0.35   0.33  -1.6e+00  -9.0e-01  -4.5e-01      5245      3286       23625    1.0
#> log_lik[59]     -1.3e+00  5.4e-03    0.34   0.33  -1.9e+00  -1.3e+00  -8.3e-01      3994      2987       17992    1.0
#> log_lik[60]     -9.8e-01  2.5e-03    0.16   0.16  -1.3e+00  -9.7e-01  -7.3e-01      4120      3191       18559    1.0
#> log_lik[61]     -5.4e-01  1.5e-03    0.10  0.099  -7.2e-01  -5.3e-01  -3.8e-01      4281      3159       19284    1.0
#> log_lik[62]     -8.8e-01  4.9e-03    0.32   0.31  -1.5e+00  -8.4e-01  -4.2e-01      4631      3053       20861   1.00
#> log_lik[63]     -1.2e-01  1.2e-03   0.076  0.061  -2.6e-01  -1.0e-01  -3.3e-02      3849      3001       17337   1.00
#> log_lik[64]     -9.0e-01  3.8e-03    0.25   0.25  -1.4e+00  -8.8e-01  -5.3e-01      4634      2745       20875    1.0
#> log_lik[65]     -2.0e+00  9.3e-03    0.61   0.63  -3.1e+00  -2.0e+00  -1.1e+00      4332      2748       19512   1.00
#> log_lik[66]     -5.1e-01  1.9e-03    0.14   0.13  -7.5e-01  -4.9e-01  -3.1e-01      5207      3091       23453    1.0
#> log_lik[67]     -2.8e-01  1.4e-03   0.084  0.080  -4.3e-01  -2.7e-01  -1.6e-01      3812      3219       17170   1.00
#> log_lik[68]     -1.1e+00  3.7e-03    0.24   0.24  -1.5e+00  -1.0e+00  -6.9e-01      4287      2821       19311    1.0
#> log_lik[69]     -4.4e-01  1.3e-03   0.085  0.085  -5.9e-01  -4.3e-01  -3.1e-01      4179      2655       18826    1.0
#> log_lik[70]     -6.4e-01  3.5e-03    0.24   0.23  -1.1e+00  -6.1e-01  -3.0e-01      4898      3168       22062    1.0
#> log_lik[71]     -6.2e-01  3.0e-03    0.21   0.21  -1.0e+00  -5.9e-01  -3.2e-01      5015      2963       22591    1.0
#> log_lik[72]     -4.7e-01  2.8e-03    0.17   0.17  -7.9e-01  -4.5e-01  -2.3e-01      4247      3067       19129    1.0
#> log_lik[73]     -1.5e+00  5.6e-03    0.36   0.36  -2.1e+00  -1.4e+00  -9.2e-01      4119      2735       18553   1.00
#> log_lik[74]     -9.5e-01  2.9e-03    0.19   0.19  -1.3e+00  -9.4e-01  -6.5e-01      4582      3126       20637    1.0
#> log_lik[75]     -1.1e+00  6.1e-03    0.40   0.39  -1.9e+00  -1.1e+00  -5.7e-01      4433      3481       19967   1.00
#> log_lik[76]     -3.7e-01  2.2e-03    0.14   0.13  -6.3e-01  -3.6e-01  -1.8e-01      3979      3042       17922   1.00
#> log_lik[77]     -8.8e-01  2.1e-03    0.14   0.14  -1.1e+00  -8.7e-01  -6.6e-01      4522      2642       20370    1.0
#> log_lik[78]     -4.9e-01  2.6e-03    0.18   0.17  -8.2e-01  -4.7e-01  -2.4e-01      4715      3107       21238    1.0
#> log_lik[79]     -7.6e-01  2.8e-03    0.19   0.18  -1.1e+00  -7.4e-01  -4.8e-01      4515      3244       20337    1.0
#> log_lik[80]     -5.5e-01  3.0e-03    0.20   0.19  -9.3e-01  -5.2e-01  -2.6e-01      5039      3150       22699    1.0
#> log_lik[81]     -1.7e-01  1.6e-03   0.098  0.084  -3.6e-01  -1.4e-01  -5.2e-02      3937      3050       17735    1.0
#> log_lik[82]     -2.3e-01  2.3e-03    0.14   0.12  -5.0e-01  -2.0e-01  -6.3e-02      4228      3041       19043    1.0
#> log_lik[83]     -3.4e-01  1.3e-03   0.082  0.081  -4.9e-01  -3.4e-01  -2.2e-01      3768      2812       16973    1.0
#> log_lik[84]     -2.8e-01  1.4e-03   0.091  0.088  -4.4e-01  -2.7e-01  -1.5e-01      3984      2896       17948   1.00
#> log_lik[85]     -1.3e-01  1.2e-03   0.074  0.062  -2.7e-01  -1.2e-01  -4.3e-02      3547      3003       15977    1.0
#> log_lik[86]     -1.1e+00  4.8e-03    0.32   0.31  -1.7e+00  -1.1e+00  -6.4e-01      4574      3143       20605    1.0
#> log_lik[87]     -8.3e-01  2.1e-03    0.13   0.13  -1.1e+00  -8.2e-01  -6.2e-01      4134      3065       18620    1.0
#> log_lik[88]     -7.8e-01  3.7e-03    0.25   0.25  -1.2e+00  -7.6e-01  -4.3e-01      4691      2908       21132    1.0
#> log_lik[89]     -1.3e+00  4.9e-03    0.32   0.31  -1.8e+00  -1.2e+00  -7.9e-01      4285      3252       19302   1.00
#> log_lik[90]     -2.7e-01  2.1e-03    0.13   0.12  -5.2e-01  -2.4e-01  -9.5e-02      4302      3056       19377    1.0
#> log_lik[91]     -3.9e-01  2.1e-03    0.13   0.12  -6.3e-01  -3.8e-01  -2.1e-01      4172      3141       18792    1.0
#> log_lik[92]     -1.5e+00  5.2e-03    0.34   0.35  -2.1e+00  -1.5e+00  -9.5e-01      4668      2994       21027    1.0
#> log_lik[93]     -7.4e-01  3.4e-03    0.22   0.21  -1.1e+00  -7.2e-01  -4.2e-01      4255      3352       19169   1.00
#> log_lik[94]     -3.2e-01  1.4e-03   0.088  0.083  -4.8e-01  -3.1e-01  -1.9e-01      3745      2687       16868   1.00
#> log_lik[95]     -3.9e-01  1.8e-03    0.12   0.11  -6.0e-01  -3.8e-01  -2.2e-01      4117      2977       18545    1.0
#> log_lik[96]     -1.6e+00  4.7e-03    0.28   0.27  -2.1e+00  -1.5e+00  -1.1e+00      3622      2834       16315   1.00
#> log_lik[97]     -4.3e-01  1.6e-03    0.10   0.10  -6.2e-01  -4.2e-01  -2.8e-01      4186      3327       18856    1.0
#> log_lik[98]     -1.0e+00  5.3e-03    0.37   0.35  -1.7e+00  -1.0e+00  -5.1e-01      5196      3207       23403    1.0
#> log_lik[99]     -7.0e-01  2.2e-03    0.15   0.14  -9.6e-01  -6.9e-01  -4.8e-01      4456      2878       20074    1.0
#> log_lik[100]    -3.9e-01  1.6e-03   0.099  0.096  -5.7e-01  -3.8e-01  -2.4e-01      4011      3033       18065    1.0
#> 
#> Samples were drawn using hmc with nuts.
#> For each parameter, ESS_bulk and ESS_tail measure the effective sample size for the entire sample (bulk) and for the .05 and .95 tails (tail), 
#> and R_hat measures the potential scale reduction on split chains. At convergence R_hat will be very close to 1.00.
# }