Run CmdStan's stansummary and diagnose utilities

Run CmdStan's stansummary and diagnose utilities. These are documented in the CmdStan Guide:

https://mc-stan.org/docs/cmdstan-guide/stansummary.html

https://mc-stan.org/docs/cmdstan-guide/diagnose.html

Although these methods can be used for models fit using the $variational() method, much of the output is currently only relevant for models fit using the $sample() method.

See the $summary() for computing similar summaries in R rather than calling CmdStan's utilites.

Usage

cmdstan_summary(flags = NULL)

cmdstan_diagnose()

Arguments

flags

An optional character vector of flags (e.g. flags = c("--sig_figs=1")).

See also

Examples

# \dontrun{
fit <- cmdstanr_example("logistic")
fit$cmdstan_diagnose()
#> Checking sampler transitions treedepth.
#> Treedepth satisfactory for all transitions.
#> 
#> Checking sampler transitions for divergences.
#> No divergent transitions found.
#> 
#> Checking E-BFMI - sampler transitions HMC potential energy.
#> E-BFMI satisfactory.
#> 
#> Rank-normalized split effective sample size satisfactory for all parameters.
#> 
#> Rank-normalized split R-hat values satisfactory for all parameters.
#> 
#> Processing complete, no problems detected.
fit$cmdstan_summary()
#> Inference for Stan model: logistic_model
#> 4 chains: each with iter=1000; warmup=1000; thin=1; 1000 iterations saved.
#> 
#> Warmup took (0.019, 0.019, 0.019, 0.019) seconds, 0.076 seconds total
#> Sampling took (0.056, 0.055, 0.055, 0.056) seconds, 0.22 seconds total
#> 
#>                     Mean     MCSE  StdDev    MAD     5%       50%       95%  ESS_bulk  ESS_tail  ESS_bulk/s  R_hat
#> 
#> lp__            -6.6e+01  3.0e-02     1.4    1.2    -69  -6.6e+01  -6.4e+01      2171      3108        9781    1.0
#> accept_stat__       0.90  1.5e-03    0.11  0.085   0.67      0.94       1.0      6024      4230       27135    1.0
#> stepsize__          0.78      nan   0.032  0.023   0.73      0.80      0.81       nan       nan         nan    nan
#> treedepth__          2.4      nan    0.52   0.00    2.0       2.0       3.0      3725      3348       16778    1.0
#> n_leapfrog__         5.0  4.0e-02     2.0   0.00    3.0       7.0       7.0      2816      2853       12684    1.0
#> divergent__         0.00      nan    0.00   0.00   0.00      0.00      0.00       nan       nan         nan    nan
#> energy__              68  4.7e-02     2.0    1.8     65        68        72      1799      2771        8104    1.0
#> 
#> alpha            3.8e-01  3.5e-03    0.22   0.22  0.017   3.8e-01   7.4e-01      3949      2897       17786    1.0
#> beta[1]         -6.7e-01  3.9e-03    0.25   0.25   -1.1  -6.7e-01  -2.7e-01      4194      2966       18891    1.0
#> beta[2]         -2.7e-01  3.7e-03    0.22   0.22  -0.64  -2.7e-01   8.6e-02      3695      3372       16642    1.0
#> beta[3]          6.8e-01  4.4e-03    0.26   0.25   0.27   6.8e-01   1.1e+00      3636      2912       16380    1.0
#> log_lik[1]      -5.1e-01  1.6e-03    0.10  0.099  -0.70  -5.1e-01  -3.6e-01      3892      3119       17533    1.0
#> log_lik[2]      -4.0e-01  2.2e-03    0.14   0.14  -0.67  -3.8e-01  -2.0e-01      4257      2816       19176    1.0
#> log_lik[3]      -4.9e-01  3.6e-03    0.22   0.21  -0.91  -4.6e-01  -2.1e-01      3929      3081       17699   1.00
#> log_lik[4]      -4.5e-01  2.5e-03    0.15   0.15  -0.72  -4.3e-01  -2.4e-01      3720      3147       16758    1.0
#> log_lik[5]      -1.2e+00  4.4e-03    0.28   0.28   -1.7  -1.2e+00  -7.6e-01      4042      3402       18209    1.0
#> log_lik[6]      -5.9e-01  2.9e-03    0.19   0.19  -0.92  -5.7e-01  -3.3e-01      4280      3222       19280    1.0
#> log_lik[7]      -6.4e-01  1.9e-03    0.12   0.12  -0.86  -6.3e-01  -4.5e-01      4252      3215       19154    1.0
#> log_lik[8]      -2.8e-01  2.2e-03    0.13   0.12  -0.52  -2.5e-01  -1.1e-01      3663      3200       16501    1.0
#> log_lik[9]      -6.9e-01  2.5e-03    0.16   0.16  -0.98  -6.8e-01  -4.5e-01      4299      3356       19364    1.0
#> log_lik[10]     -7.4e-01  3.9e-03    0.24   0.23   -1.2  -7.1e-01  -4.0e-01      3876      2900       17459    1.0
#> log_lik[11]     -2.8e-01  2.0e-03    0.12   0.11  -0.50  -2.6e-01  -1.2e-01      3671      2974       16534    1.0
#> log_lik[12]     -4.9e-01  3.6e-03    0.23   0.21  -0.95  -4.6e-01  -1.9e-01      4155      2791       18717    1.0
#> log_lik[13]     -6.6e-01  3.1e-03    0.20   0.21   -1.0  -6.3e-01  -3.6e-01      4454      3330       20063    1.0
#> log_lik[14]     -3.6e-01  2.8e-03    0.17   0.16  -0.70  -3.3e-01  -1.4e-01      3899      3185       17564    1.0
#> log_lik[15]     -2.8e-01  1.7e-03    0.10  0.099  -0.47  -2.6e-01  -1.3e-01      3799      3134       17112    1.0
#> log_lik[16]     -2.7e-01  1.4e-03   0.085  0.081  -0.43  -2.6e-01  -1.5e-01      3507      3080       15795    1.0
#> log_lik[17]     -1.6e+00  4.9e-03    0.29   0.28   -2.1  -1.6e+00  -1.2e+00      3663      3056       16501    1.0
#> log_lik[18]     -4.8e-01  1.6e-03    0.10   0.10  -0.66  -4.7e-01  -3.2e-01      4189      3215       18868    1.0
#> log_lik[19]     -2.3e-01  1.2e-03   0.075  0.073  -0.36  -2.2e-01  -1.2e-01      3590      2916       16170    1.0
#> log_lik[20]     -1.1e-01  1.3e-03   0.077  0.059  -0.26  -9.2e-02  -2.7e-02      3696      2966       16648   1.00
#> log_lik[21]     -2.1e-01  1.4e-03   0.085  0.078  -0.36  -2.0e-01  -9.4e-02      3427      2622       15437    1.0
#> log_lik[22]     -5.7e-01  2.2e-03    0.14   0.14  -0.83  -5.5e-01  -3.5e-01      4116      3272       18538    1.0
#> log_lik[23]     -3.3e-01  2.1e-03    0.14   0.13  -0.58  -3.1e-01  -1.5e-01      3962      3151       17846    1.0
#> log_lik[24]     -1.4e-01  1.1e-03   0.066  0.058  -0.26  -1.2e-01  -5.1e-02      3574      2827       16101    1.0
#> log_lik[25]     -4.6e-01  1.9e-03    0.12   0.12  -0.67  -4.4e-01  -2.8e-01      3985      3178       17951    1.0
#> log_lik[26]     -1.5e+00  5.3e-03    0.35   0.34   -2.1  -1.5e+00  -1.0e+00      4339      2973       19547    1.0
#> log_lik[27]     -3.0e-01  2.0e-03    0.12   0.11  -0.52  -2.9e-01  -1.4e-01      3645      2969       16420    1.0
#> log_lik[28]     -4.4e-01  1.3e-03   0.083  0.083  -0.59  -4.4e-01  -3.2e-01      3869      3037       17428    1.0
#> log_lik[29]     -7.3e-01  3.5e-03    0.23   0.22   -1.1  -7.1e-01  -3.9e-01      4217      3039       18998    1.0
#> log_lik[30]     -6.9e-01  2.8e-03    0.18   0.19   -1.0  -6.8e-01  -4.3e-01      4492      3385       20235    1.0
#> log_lik[31]     -4.9e-01  2.5e-03    0.16   0.15  -0.77  -4.7e-01  -2.6e-01      3988      3295       17963    1.0
#> log_lik[32]     -4.2e-01  1.7e-03    0.11   0.10  -0.61  -4.1e-01  -2.7e-01      3897      3308       17555    1.0
#> log_lik[33]     -4.1e-01  2.0e-03    0.13   0.12  -0.64  -3.9e-01  -2.3e-01      4134      2815       18624    1.0
#> log_lik[34]     -6.3e-02  8.5e-04   0.050  0.034  -0.16  -4.9e-02  -1.3e-02      3598      2957       16207    1.0
#> log_lik[35]     -5.9e-01  2.7e-03    0.18   0.18  -0.93  -5.7e-01  -3.2e-01      4662      3156       20998   1.00
#> log_lik[36]     -3.2e-01  2.0e-03    0.13   0.12  -0.56  -3.1e-01  -1.5e-01      4238      3048       19088    1.0
#> log_lik[37]     -7.0e-01  3.7e-03    0.23   0.22   -1.1  -6.7e-01  -3.7e-01      4046      3015       18225    1.0
#> log_lik[38]     -3.1e-01  2.4e-03    0.15   0.14  -0.61  -2.9e-01  -1.2e-01      3856      3258       17369    1.0
#> log_lik[39]     -1.8e-01  1.8e-03    0.11  0.088  -0.39  -1.5e-01  -5.1e-02      3779      3160       17020    1.0
#> log_lik[40]     -6.8e-01  1.9e-03    0.13   0.12  -0.90  -6.7e-01  -4.9e-01      4328      3085       19497    1.0
#> log_lik[41]     -1.1e+00  4.0e-03    0.25   0.25   -1.6  -1.1e+00  -7.5e-01      3809      3357       17160    1.0
#> log_lik[42]     -9.3e-01  3.0e-03    0.19   0.19   -1.3  -9.2e-01  -6.4e-01      4274      3061       19251    1.0
#> log_lik[43]     -4.1e-01  4.0e-03    0.26   0.22  -0.91  -3.5e-01  -1.0e-01      4724      3132       21280    1.0
#> log_lik[44]     -1.2e+00  3.1e-03    0.19   0.19   -1.5  -1.2e+00  -8.9e-01      3691      3332       16626    1.0
#> log_lik[45]     -3.6e-01  1.9e-03    0.12   0.11  -0.58  -3.4e-01  -1.9e-01      3887      3351       17510    1.0
#> log_lik[46]     -5.8e-01  1.9e-03    0.13   0.12  -0.80  -5.7e-01  -3.9e-01      4322      3290       19467    1.0
#> log_lik[47]     -3.0e-01  2.0e-03    0.13   0.12  -0.54  -2.8e-01  -1.3e-01      3767      2734       16968    1.0
#> log_lik[48]     -3.2e-01  1.3e-03   0.083  0.081  -0.47  -3.2e-01  -2.0e-01      3899      3125       17565    1.0
#> log_lik[49]     -3.2e-01  1.3e-03   0.079  0.077  -0.45  -3.1e-01  -2.0e-01      3523      3053       15869    1.0
#> log_lik[50]     -1.3e+00  5.0e-03    0.33   0.33   -1.9  -1.3e+00  -7.9e-01      4545      3361       20473    1.0
#> log_lik[51]     -2.9e-01  1.5e-03   0.095  0.090  -0.45  -2.8e-01  -1.5e-01      4189      3004       18870   1.00
#> log_lik[52]     -8.4e-01  2.3e-03    0.15   0.14   -1.1  -8.3e-01  -6.1e-01      3989      2943       17967    1.0
#> log_lik[53]     -4.0e-01  2.0e-03    0.13   0.12  -0.63  -3.8e-01  -2.2e-01      3797      3260       17101    1.0
#> log_lik[54]     -3.7e-01  2.3e-03    0.14   0.14  -0.64  -3.5e-01  -1.7e-01      3995      3290       17996    1.0
#> log_lik[55]     -3.8e-01  2.1e-03    0.13   0.12  -0.62  -3.7e-01  -2.0e-01      4146      2822       18677    1.0
#> log_lik[56]     -3.2e-01  2.9e-03    0.19   0.16  -0.67  -2.8e-01  -9.7e-02      4370      2973       19685    1.0
#> log_lik[57]     -6.6e-01  1.9e-03    0.12   0.12  -0.86  -6.5e-01  -4.7e-01      3984      3102       17947    1.0
#> log_lik[58]     -9.5e-01  5.5e-03    0.35   0.34   -1.6  -9.0e-01  -4.5e-01      4407      3220       19852   1.00
#> log_lik[59]     -1.4e+00  5.3e-03    0.34   0.34   -2.0  -1.3e+00  -8.5e-01      4062      2814       18296    1.0
#> log_lik[60]     -9.8e-01  2.6e-03    0.16   0.16   -1.3  -9.7e-01  -7.3e-01      3893      3008       17537    1.0
#> log_lik[61]     -5.4e-01  1.6e-03   0.099  0.099  -0.71  -5.3e-01  -3.9e-01      4055      2768       18266    1.0
#> log_lik[62]     -8.8e-01  4.9e-03    0.31   0.30   -1.5  -8.4e-01  -4.4e-01      4268      3314       19224    1.0
#> log_lik[63]     -1.1e-01  1.2e-03   0.071  0.058  -0.25  -9.9e-02  -3.3e-02      3327      2903       14985    1.0
#> log_lik[64]     -9.0e-01  3.6e-03    0.24   0.24   -1.3  -8.8e-01  -5.5e-01      4483      3442       20193    1.0
#> log_lik[65]     -2.0e+00  1.0e-02    0.58   0.55   -3.1  -2.0e+00  -1.1e+00      3469      2875       15624    1.0
#> log_lik[66]     -5.1e-01  2.1e-03    0.14   0.13  -0.75  -4.9e-01  -3.1e-01      4241      3219       19102    1.0
#> log_lik[67]     -2.8e-01  1.3e-03   0.081  0.078  -0.42  -2.7e-01  -1.6e-01      3796      3184       17101    1.0
#> log_lik[68]     -1.1e+00  3.6e-03    0.23   0.23   -1.5  -1.0e+00  -7.2e-01      4100      3268       18468    1.0
#> log_lik[69]     -4.3e-01  1.4e-03   0.084  0.085  -0.58  -4.3e-01  -3.1e-01      3739      3111       16841    1.0
#> log_lik[70]     -6.4e-01  3.4e-03    0.23   0.22   -1.1  -6.1e-01  -3.2e-01      4643      3088       20913    1.0
#> log_lik[71]     -6.1e-01  3.1e-03    0.21   0.20  -0.99  -5.9e-01  -3.2e-01      4649      3191       20940   1.00
#> log_lik[72]     -4.6e-01  2.6e-03    0.17   0.16  -0.78  -4.4e-01  -2.3e-01      4343      3065       19565    1.0
#> log_lik[73]     -1.5e+00  6.0e-03    0.36   0.36   -2.1  -1.5e+00  -9.3e-01      3831      3013       17257    1.0
#> log_lik[74]     -9.5e-01  3.1e-03    0.20   0.20   -1.3  -9.4e-01  -6.5e-01      3981      2989       17931    1.0
#> log_lik[75]     -1.1e+00  6.1e-03    0.39   0.39   -1.9  -1.1e+00  -5.9e-01      4306      3363       19394    1.0
#> log_lik[76]     -3.7e-01  2.3e-03    0.14   0.13  -0.64  -3.5e-01  -1.8e-01      3850      3058       17344    1.0
#> log_lik[77]     -8.8e-01  2.2e-03    0.14   0.14   -1.1  -8.7e-01  -6.6e-01      4041      3172       18201    1.0
#> log_lik[78]     -4.9e-01  2.6e-03    0.17   0.17  -0.80  -4.6e-01  -2.4e-01      4455      3331       20068    1.0
#> log_lik[79]     -7.6e-01  3.1e-03    0.19   0.19   -1.1  -7.4e-01  -4.7e-01      3942      2923       17758    1.0
#> log_lik[80]     -5.4e-01  2.8e-03    0.19   0.18  -0.89  -5.2e-01  -2.7e-01      4606      3050       20747    1.0
#> log_lik[81]     -1.6e-01  1.6e-03    0.10  0.080  -0.35  -1.4e-01  -4.9e-02      3821      3031       17212    1.0
#> log_lik[82]     -2.2e-01  2.1e-03    0.13   0.11  -0.48  -1.9e-01  -6.6e-02      4076      2738       18360    1.0
#> log_lik[83]     -3.4e-01  1.4e-03   0.082  0.080  -0.48  -3.4e-01  -2.2e-01      3626      3189       16334    1.0
#> log_lik[84]     -2.7e-01  1.5e-03   0.091  0.086  -0.43  -2.6e-01  -1.4e-01      3606      3216       16245    1.0
#> log_lik[85]     -1.3e-01  1.2e-03   0.074  0.062  -0.27  -1.1e-01  -4.0e-02      3631      3139       16357   1.00
#> log_lik[86]     -1.1e+00  4.7e-03    0.31   0.30   -1.7  -1.1e+00  -6.6e-01      4377      3017       19718    1.0
#> log_lik[87]     -8.2e-01  2.0e-03    0.13   0.12   -1.0  -8.2e-01  -6.2e-01      4229      3001       19048    1.0
#> log_lik[88]     -7.8e-01  3.6e-03    0.24   0.24   -1.2  -7.5e-01  -4.3e-01      4368      3215       19675    1.0
#> log_lik[89]     -1.3e+00  4.9e-03    0.31   0.30   -1.8  -1.3e+00  -8.1e-01      4041      2764       18203    1.0
#> log_lik[90]     -2.6e-01  2.2e-03    0.14   0.12  -0.53  -2.4e-01  -9.2e-02      4059      3272       18282    1.0
#> log_lik[91]     -3.9e-01  2.0e-03    0.13   0.13  -0.63  -3.7e-01  -2.0e-01      4188      3216       18865    1.0
#> log_lik[92]     -1.5e+00  5.6e-03    0.34   0.33   -2.1  -1.5e+00  -9.9e-01      3760      3151       16938    1.0
#> log_lik[93]     -7.4e-01  3.6e-03    0.22   0.22   -1.1  -7.2e-01  -4.2e-01      3836      2983       17279    1.0
#> log_lik[94]     -3.2e-01  1.5e-03   0.088  0.086  -0.47  -3.1e-01  -1.9e-01      3654      3163       16460    1.0
#> log_lik[95]     -3.9e-01  1.8e-03    0.11   0.11  -0.59  -3.8e-01  -2.3e-01      3750      3346       16891    1.0
#> log_lik[96]     -1.6e+00  4.9e-03    0.29   0.28   -2.1  -1.6e+00  -1.2e+00      3526      2979       15884    1.0
#> log_lik[97]     -4.3e-01  1.5e-03   0.100  0.097  -0.61  -4.2e-01  -2.8e-01      4416      3135       19893    1.0
#> log_lik[98]     -1.0e+00  5.7e-03    0.37   0.36   -1.7  -1.0e+00  -5.2e-01      4490      3148       20227   1.00
#> log_lik[99]     -6.9e-01  2.1e-03    0.14   0.14  -0.93  -6.8e-01  -4.8e-01      4395      3215       19798    1.0
#> log_lik[100]    -3.9e-01  1.5e-03   0.098  0.095  -0.57  -3.8e-01  -2.4e-01      4062      2972       18298    1.0
#> 
#> Samples were drawn using hmc with nuts.
#> For each parameter, ESS_bulk and ESS_tail measure the effective sample size for the entire sample (bulk) and for the .05 and .95 tails (tail), 
#> and R_hat measures the potential scale reduction on split chains. At convergence R_hat will be very close to 1.00.
# }