Extract posterior draws

Extract posterior draws after MCMC or approximate posterior draws after variational approximation using formats provided by the posterior package.

The variables include the parameters, transformed parameters, and generated quantities from the Stan program as well as lp__, the total log probability (target) accumulated in the model block.

Usage

$draws(variables = NULL, inc_warmup = FALSE, ...)

Arguments

variables: (character vector) The variables (parameters and generated quantities) to read in. If NULL (the default) then the draws of all variables are included.

inc_warmup: (logical) For MCMC only, should warmup draws be included? Defaults to FALSE.

...: Arguments passed on to posterior::as_draws_array().

Value

For MCMC, a 3-D draws_array object (iteration x chain x variable).

For variational inference, a 2-D draws_matrix object (draw x variable). An additional variable lp_approx__ is also included, which is the log density of the variational approximation to the posterior evaluated at each of the draws.

For optimization, a 1-row draws_matrix with one column per variable. These are not actually draws, just point estimates stored in the draws_matrix format.

Examples

# \dontrun{
# logistic regression with intercept alpha and coefficients beta
fit <- cmdstanr_example("logistic")

#> Model executable is up to date!

# returned as 3-D array (see ?posterior::draws_array)
draws <- fit$draws()
dim(draws)

#> [1] 1000    4    5

str(draws)

#>  'draws_array' num [1:1000, 1:4, 1:5] -66.9 -64.7 -66.7 -66.1 -65.3 ...
#>  - attr(*, "dimnames")=List of 3
#>   ..$ iteration: chr [1:1000] "1" "2" "3" "4" ...
#>   ..$ chain    : chr [1:4] "1" "2" "3" "4"
#>   ..$ variable : chr [1:5] "lp__" "alpha" "beta[1]" "beta[2]" ...

# can easily convert to other formats (data frame, matrix, list)
library("posterior")

#> This is posterior version 0.1.0

#> 
#> Attaching package: ‘posterior’

#> The following objects are masked from ‘package:rstan’:
#> 
#>     ess_bulk, ess_tail

as_draws_df(draws)  # see also as_draws_matrix, as_draws_list

#> # A draws_df: 1000 iterations, 4 chains, and 5 variables
#>    lp__  alpha beta[1] beta[2] beta[3]
#> 1   -67  0.387   -1.07   -0.41    0.32
#> 2   -65  0.463   -0.54   -0.32    0.93
#> 3   -67  0.408   -0.63   -0.79    0.66
#> 4   -66  0.616   -0.95   -0.41    0.46
#> 5   -65  0.449   -0.31   -0.15    0.45
#> 6   -67  0.393   -1.23   -0.42    0.82
#> 7   -67  0.806   -0.79   -0.42    1.05
#> 8   -68  0.860   -0.57   -0.41    1.13
#> 9   -70 -0.077   -0.99   -0.20    0.12
#> 10  -65  0.283   -0.52   -0.38    0.28
#> # ... with 3990 more draws
#> # ... hidden meta-columns {'.chain', '.iteration', '.draw'}

# can select specific parameters
fit$draws("alpha")

#> # A draws_array: 1000 iterations, 4 chains, and 1 variables
#> , , variable = alpha
#> 
#>          chain
#> iteration    1    2    3    4
#>         1 0.39 0.30 0.22 0.48
#>         2 0.46 0.35 0.28 0.56
#>         3 0.41 0.26 0.30 0.46
#>         4 0.62 0.28 0.33 0.60
#>         5 0.45 0.41 0.62 0.14
#> 
#> # ... with 995 more iterations

fit$draws("beta")  # selects entire vector beta

#> # A draws_array: 1000 iterations, 4 chains, and 3 variables
#> , , variable = beta[1]
#> 
#>          chain
#> iteration     1     2     3     4
#>         1 -1.07 -0.82 -0.69 -0.52
#>         2 -0.54 -0.74 -0.50 -0.80
#>         3 -0.63 -0.56 -0.68 -0.84
#>         4 -0.95 -0.73 -0.47 -0.88
#>         5 -0.31 -0.55 -1.11 -0.57
#> 
#> , , variable = beta[2]
#> 
#>          chain
#> iteration     1      2     3       4
#>         1 -0.41  0.205 -0.79  0.0799
#>         2 -0.32 -0.562 -0.76 -0.2266
#>         3 -0.79 -0.158 -0.65 -0.3127
#>         4 -0.41 -0.363 -0.77 -0.6677
#>         5 -0.15 -0.011 -0.17 -0.0065
#> 
#> , , variable = beta[3]
#> 
#>          chain
#> iteration    1    2    3    4
#>         1 0.32 0.26 1.19 0.65
#>         2 0.93 1.05 1.04 1.16
#>         3 0.66 0.37 1.12 0.20
#>         4 0.46 0.77 0.42 0.90
#>         5 0.45 1.01 0.84 0.79
#> 
#> # ... with 995 more iterations

fit$draws(c("alpha", "beta[2]"))

#> # A draws_array: 1000 iterations, 4 chains, and 2 variables
#> , , variable = alpha
#> 
#>          chain
#> iteration    1    2    3    4
#>         1 0.39 0.30 0.22 0.48
#>         2 0.46 0.35 0.28 0.56
#>         3 0.41 0.26 0.30 0.46
#>         4 0.62 0.28 0.33 0.60
#>         5 0.45 0.41 0.62 0.14
#> 
#> , , variable = beta[2]
#> 
#>          chain
#> iteration     1      2     3       4
#>         1 -0.41  0.205 -0.79  0.0799
#>         2 -0.32 -0.562 -0.76 -0.2266
#>         3 -0.79 -0.158 -0.65 -0.3127
#>         4 -0.41 -0.363 -0.77 -0.6677
#>         5 -0.15 -0.011 -0.17 -0.0065
#> 
#> # ... with 995 more iterations

# }