Extract inverse metric (mass matrix) after MCMC

Extract the inverse metric (mass matrix) for each MCMC chain.

inv_metric(matrix = TRUE)

Arguments

matrix

(logical) If a diagonal metric was used, setting matrix = FALSE returns a list containing just the diagonals of the matrices instead of the full matrices. Setting matrix = FALSE has no effect for dense metrics.

Value

A list of length equal to the number of MCMC chains. See the matrix

argument for details.

See also

Examples

# \dontrun{
fit <- cmdstanr_example("logistic")
fit$inv_metric()
#> $`1`
#>           [,1]      [,2]      [,3]      [,4]
#> [1,] 0.0461164 0.0000000 0.0000000 0.0000000
#> [2,] 0.0000000 0.0586249 0.0000000 0.0000000
#> [3,] 0.0000000 0.0000000 0.0569083 0.0000000
#> [4,] 0.0000000 0.0000000 0.0000000 0.0695349
#> 
#> $`2`
#>           [,1]      [,2]      [,3]      [,4]
#> [1,] 0.0423827 0.0000000 0.0000000 0.0000000
#> [2,] 0.0000000 0.0638639 0.0000000 0.0000000
#> [3,] 0.0000000 0.0000000 0.0463094 0.0000000
#> [4,] 0.0000000 0.0000000 0.0000000 0.0646605
#> 
#> $`3`
#>           [,1]      [,2]      [,3]      [,4]
#> [1,] 0.0555959 0.0000000 0.0000000 0.0000000
#> [2,] 0.0000000 0.0595803 0.0000000 0.0000000
#> [3,] 0.0000000 0.0000000 0.0515264 0.0000000
#> [4,] 0.0000000 0.0000000 0.0000000 0.0759908
#> 
#> $`4`
#>         [,1]     [,2]      [,3]    [,4]
#> [1,] 0.05104 0.000000 0.0000000 0.00000
#> [2,] 0.00000 0.051252 0.0000000 0.00000
#> [3,] 0.00000 0.000000 0.0567318 0.00000
#> [4,] 0.00000 0.000000 0.0000000 0.06699
#> 
fit$inv_metric(matrix=FALSE)
#> $`1`
#> [1] 0.0461164 0.0586249 0.0569083 0.0695349
#> 
#> $`2`
#> [1] 0.0423827 0.0638639 0.0463094 0.0646605
#> 
#> $`3`
#> [1] 0.0555959 0.0595803 0.0515264 0.0759908
#> 
#> $`4`
#> [1] 0.0510400 0.0512520 0.0567318 0.0669900
#> 

fit <- cmdstanr_example("logistic", metric = "dense_e")
fit$inv_metric()
#> $`1`
#>              [,1]        [,2]         [,3]        [,4]
#> [1,]  0.048223400 -0.00737038  0.000297676  0.00680596
#> [2,] -0.007370380  0.05440320 -0.006681250 -0.00162053
#> [3,]  0.000297676 -0.00668125  0.043576600 -0.00859282
#> [4,]  0.006805960 -0.00162053 -0.008592820  0.06856220
#> 
#> $`2`
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.04746980 -0.00175084 -0.00118626  0.00664285
#> [2,] -0.00175084  0.06320920 -0.00172180 -0.01259270
#> [3,] -0.00118626 -0.00172180  0.04810590 -0.00872089
#> [4,]  0.00664285 -0.01259270 -0.00872089  0.06689500
#> 
#> $`3`
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.04482880 -0.00377658 -0.00343049  0.00304829
#> [2,] -0.00377658  0.04803130 -0.00957526 -0.00566723
#> [3,] -0.00343049 -0.00957526  0.05099840 -0.00574551
#> [4,]  0.00304829 -0.00566723 -0.00574551  0.07092750
#> 
#> $`4`
#>             [,1]         [,2]         [,3]        [,4]
#> [1,]  0.04787540 -0.003918010 -0.004273680  0.00253868
#> [2,] -0.00391801  0.049792000  0.000712128 -0.00453177
#> [3,] -0.00427368  0.000712128  0.045850800 -0.00639672
#> [4,]  0.00253868 -0.004531770 -0.006396720  0.07511110
#> 
# }