Extract inverse metric (mass matrix) after MCMC

Extract the inverse metric (mass matrix) for each MCMC chain.

inv_metric(matrix = TRUE)

Arguments

matrix
(logical) If a diagonal metric was used, setting matrix = FALSE returns a list containing just the diagonals of the matrices instead of the full matrices. Setting matrix = FALSE has no effect for dense metrics.

Value

A list of length equal to the number of MCMC chains. See the matrix argument for details.

See also

Examples

# \dontrun{
fit <- cmdstanr_example("logistic")

#> Model executable is up to date!

fit$inv_metric()

#> $`1`
#>           [,1]      [,2]      [,3]      [,4]
#> [1,] 0.0409791 0.0000000 0.0000000 0.0000000
#> [2,] 0.0000000 0.0699631 0.0000000 0.0000000
#> [3,] 0.0000000 0.0000000 0.0478488 0.0000000
#> [4,] 0.0000000 0.0000000 0.0000000 0.0720792
#> 
#> $`2`
#>           [,1]      [,2]      [,3]      [,4]
#> [1,] 0.0521783 0.0000000 0.0000000 0.0000000
#> [2,] 0.0000000 0.0651446 0.0000000 0.0000000
#> [3,] 0.0000000 0.0000000 0.0503775 0.0000000
#> [4,] 0.0000000 0.0000000 0.0000000 0.0796248
#> 
#> $`3`
#>           [,1]     [,2]      [,3]      [,4]
#> [1,] 0.0476085 0.000000 0.0000000 0.0000000
#> [2,] 0.0000000 0.067774 0.0000000 0.0000000
#> [3,] 0.0000000 0.000000 0.0545375 0.0000000
#> [4,] 0.0000000 0.000000 0.0000000 0.0657108
#> 
#> $`4`
#>           [,1]      [,2]     [,3]      [,4]
#> [1,] 0.0473458 0.0000000 0.000000 0.0000000
#> [2,] 0.0000000 0.0639189 0.000000 0.0000000
#> [3,] 0.0000000 0.0000000 0.062392 0.0000000
#> [4,] 0.0000000 0.0000000 0.000000 0.0728948
#> 

fit$inv_metric(matrix=FALSE)

#> $`1`
#> [1] 0.0409791 0.0699631 0.0478488 0.0720792
#> 
#> $`2`
#> [1] 0.0521783 0.0651446 0.0503775 0.0796248
#> 
#> $`3`
#> [1] 0.0476085 0.0677740 0.0545375 0.0657108
#> 
#> $`4`
#> [1] 0.0473458 0.0639189 0.0623920 0.0728948
#> 

fit <- cmdstanr_example("logistic", metric = "dense_e")

#> Model executable is up to date!

fit$inv_metric()

#> $`1`
#>              [,1]        [,2]         [,3]        [,4]
#> [1,]  0.051420700  0.00120054 -0.000585942  0.00104648
#> [2,]  0.001200540  0.06446300 -0.012509400 -0.00647790
#> [3,] -0.000585942 -0.01250940  0.043045100 -0.00969323
#> [4,]  0.001046480 -0.00647790 -0.009693230  0.06779780
#> 
#> $`2`
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.04722150 -0.00314819  0.00108351  0.00300970
#> [2,] -0.00314819  0.05286400  0.00134121 -0.00685724
#> [3,]  0.00108351  0.00134121  0.04952160 -0.01142720
#> [4,]  0.00300970 -0.00685724 -0.01142720  0.06245950
#> 
#> $`3`
#>             [,1]         [,2]         [,3]        [,4]
#> [1,]  0.04664440 -0.002356350  0.000594790  0.00696777
#> [2,] -0.00235635  0.065111800  0.000294205 -0.00373292
#> [3,]  0.00059479  0.000294205  0.047578700 -0.01097830
#> [4,]  0.00696777 -0.003732920 -0.010978300  0.06668740
#> 
#> $`4`
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.04962380 -0.00332921  0.00168712  0.00224150
#> [2,] -0.00332921  0.05970130 -0.00491432 -0.00774457
#> [3,]  0.00168712 -0.00491432  0.04630120 -0.00663267
#> [4,]  0.00224150 -0.00774457 -0.00663267  0.07457350
#> 

# }