Extract inverse metric (mass matrix)

Return a list containing the inverse metric (mass matrix) for each chain.

Usage

$inv_metric(matrix = TRUE)

Arguments

matrix: (logical) If a diagonal metric was used, setting matrix = FALSE returns a list containing just the diagonals of the matrices instead of the full matrices. Setting matrix = FALSE has no effect for dense metrics.

Value

A list of length equal to the number of MCMC chains. See the matrix argument for details.

See also

Examples

# \dontrun{
fit <- cmdstanr_example("logistic")

#> Model executable is up to date!

fit$inv_metric()

#> [[1]]
#>           [,1]      [,2]      [,3]      [,4]
#> [1,] 0.0475557 0.0000000 0.0000000 0.0000000
#> [2,] 0.0000000 0.0569999 0.0000000 0.0000000
#> [3,] 0.0000000 0.0000000 0.0578795 0.0000000
#> [4,] 0.0000000 0.0000000 0.0000000 0.0712875
#> 
#> [[2]]
#>           [,1]      [,2]      [,3]      [,4]
#> [1,] 0.0456437 0.0000000 0.0000000 0.0000000
#> [2,] 0.0000000 0.0576204 0.0000000 0.0000000
#> [3,] 0.0000000 0.0000000 0.0559887 0.0000000
#> [4,] 0.0000000 0.0000000 0.0000000 0.0685771
#> 
#> [[3]]
#>           [,1]      [,2]      [,3]      [,4]
#> [1,] 0.0461312 0.0000000 0.0000000 0.0000000
#> [2,] 0.0000000 0.0628157 0.0000000 0.0000000
#> [3,] 0.0000000 0.0000000 0.0466361 0.0000000
#> [4,] 0.0000000 0.0000000 0.0000000 0.0699689
#> 
#> [[4]]
#>           [,1]      [,2]      [,3]      [,4]
#> [1,] 0.0434073 0.0000000 0.0000000 0.0000000
#> [2,] 0.0000000 0.0685943 0.0000000 0.0000000
#> [3,] 0.0000000 0.0000000 0.0483899 0.0000000
#> [4,] 0.0000000 0.0000000 0.0000000 0.0702005
#> 

fit$inv_metric(matrix=FALSE)

#> [[1]]
#> [1] 0.0475557 0.0569999 0.0578795 0.0712875
#> 
#> [[2]]
#> [1] 0.0456437 0.0576204 0.0559887 0.0685771
#> 
#> [[3]]
#> [1] 0.0461312 0.0628157 0.0466361 0.0699689
#> 
#> [[4]]
#> [1] 0.0434073 0.0685943 0.0483899 0.0702005
#> 

fit <- cmdstanr_example("logistic", metric = "dense_e")

#> Model executable is up to date!

fit$inv_metric()

#> [[1]]
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.04656590 -0.00610981  0.00424381 -0.00100949
#> [2,] -0.00610981  0.06209830 -0.00587438 -0.00300882
#> [3,]  0.00424381 -0.00587438  0.04098100 -0.01262950
#> [4,] -0.00100949 -0.00300882 -0.01262950  0.06624320
#> 
#> [[2]]
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.04556050 -0.00515669  0.00690407 -0.00379630
#> [2,] -0.00515669  0.06277340 -0.00282879 -0.00752052
#> [3,]  0.00690407 -0.00282879  0.05169290 -0.01233880
#> [4,] -0.00379630 -0.00752052 -0.01233880  0.06552800
#> 
#> [[3]]
#>              [,1]         [,2]        [,3]        [,4]
#> [1,]  0.048585400 -0.000740865 -0.00208509  0.00145319
#> [2,] -0.000740865  0.067828500 -0.01015370 -0.00763077
#> [3,] -0.002085090 -0.010153700  0.05735690 -0.01572740
#> [4,]  0.001453190 -0.007630770 -0.01572740  0.07374680
#> 
#> [[4]]
#>              [,1]         [,2]         [,3]         [,4]
#> [1,]  0.043063200  0.000215595  0.000264007 -0.000423114
#> [2,]  0.000215595  0.056354100 -0.006857240 -0.008637950
#> [3,]  0.000264007 -0.006857240  0.051397800 -0.015648600
#> [4,] -0.000423114 -0.008637950 -0.015648600  0.078480500
#> 

# }