Extract inverse metric (mass matrix) after MCMC

Extract the inverse metric (mass matrix) for each MCMC chain.

Usage

inv_metric(matrix = TRUE)

Arguments

matrix

(logical) If a diagonal metric was used, setting matrix = FALSE returns a list containing just the diagonals of the matrices instead of the full matrices. Setting matrix = FALSE has no effect for dense metrics.

Value

A list of length equal to the number of MCMC chains. See the matrix argument for details.

See also

Examples

# \dontrun{
fit <- cmdstanr_example("logistic")
fit$inv_metric()
#> $`1`
#>           [,1]      [,2]     [,3]      [,4]
#> [1,] 0.0465167 0.0000000 0.000000 0.0000000
#> [2,] 0.0000000 0.0620914 0.000000 0.0000000
#> [3,] 0.0000000 0.0000000 0.050783 0.0000000
#> [4,] 0.0000000 0.0000000 0.000000 0.0729152
#> 
#> $`2`
#>           [,1]      [,2]      [,3]      [,4]
#> [1,] 0.0441307 0.0000000 0.0000000 0.0000000
#> [2,] 0.0000000 0.0636629 0.0000000 0.0000000
#> [3,] 0.0000000 0.0000000 0.0466777 0.0000000
#> [4,] 0.0000000 0.0000000 0.0000000 0.0671717
#> 
#> $`3`
#>           [,1]    [,2]      [,3]      [,4]
#> [1,] 0.0435155 0.00000 0.0000000 0.0000000
#> [2,] 0.0000000 0.06023 0.0000000 0.0000000
#> [3,] 0.0000000 0.00000 0.0461039 0.0000000
#> [4,] 0.0000000 0.00000 0.0000000 0.0683583
#> 
#> $`4`
#>           [,1]      [,2]      [,3]      [,4]
#> [1,] 0.0473502 0.0000000 0.0000000 0.0000000
#> [2,] 0.0000000 0.0626683 0.0000000 0.0000000
#> [3,] 0.0000000 0.0000000 0.0489999 0.0000000
#> [4,] 0.0000000 0.0000000 0.0000000 0.0843854
#> 
fit$inv_metric(matrix=FALSE)
#> $`1`
#> [1] 0.0465167 0.0620914 0.0507830 0.0729152
#> 
#> $`2`
#> [1] 0.0441307 0.0636629 0.0466777 0.0671717
#> 
#> $`3`
#> [1] 0.0435155 0.0602300 0.0461039 0.0683583
#> 
#> $`4`
#> [1] 0.0473502 0.0626683 0.0489999 0.0843854
#> 

fit <- cmdstanr_example("logistic", metric = "dense_e")
fit$inv_metric()
#> $`1`
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.04672720  0.00167042 -0.00114755  0.00811720
#> [2,]  0.00167042  0.06716350 -0.00613677  0.00100108
#> [3,] -0.00114755 -0.00613677  0.06399690 -0.00787439
#> [4,]  0.00811720  0.00100108 -0.00787439  0.07539580
#> 
#> $`2`
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.04772890 -0.00189944 -0.00147119  0.00308748
#> [2,] -0.00189944  0.06492360 -0.00283877 -0.00746201
#> [3,] -0.00147119 -0.00283877  0.05286320 -0.00843557
#> [4,]  0.00308748 -0.00746201 -0.00843557  0.06810290
#> 
#> $`3`
#>             [,1]         [,2]         [,3]        [,4]
#> [1,]  0.04272210 -0.004145890  0.000913590  0.00262301
#> [2,] -0.00414589  0.055612700  0.000284277 -0.00761463
#> [3,]  0.00091359  0.000284277  0.047423300 -0.01277940
#> [4,]  0.00262301 -0.007614630 -0.012779400  0.06443380
#> 
#> $`4`
#>              [,1]        [,2]         [,3]        [,4]
#> [1,]  0.046579900 -0.00284613  0.000977285  0.00202717
#> [2,] -0.002846130  0.05566120 -0.009374590 -0.00543387
#> [3,]  0.000977285 -0.00937459  0.051704000 -0.01207620
#> [4,]  0.002027170 -0.00543387 -0.012076200  0.07242500
#> 
# }