Extract inverse metric (mass matrix) after MCMC

Extract the inverse metric (mass matrix) for each MCMC chain.

Usage

inv_metric(matrix = TRUE)

Arguments

matrix

(logical) If a diagonal metric was used, setting matrix = FALSE returns a list containing just the diagonals of the matrices instead of the full matrices. Setting matrix = FALSE has no effect for dense metrics.

Value

A list of length equal to the number of MCMC chains. See the matrix argument for details.

See also

Examples

# \dontrun{
fit <- cmdstanr_example("logistic")
fit$inv_metric()
#> $`1`
#>           [,1]      [,2]      [,3]      [,4]
#> [1,] 0.0463354 0.0000000 0.0000000 0.0000000
#> [2,] 0.0000000 0.0601104 0.0000000 0.0000000
#> [3,] 0.0000000 0.0000000 0.0507043 0.0000000
#> [4,] 0.0000000 0.0000000 0.0000000 0.0636408
#> 
#> $`2`
#>           [,1]      [,2]      [,3]      [,4]
#> [1,] 0.0481084 0.0000000 0.0000000 0.0000000
#> [2,] 0.0000000 0.0561195 0.0000000 0.0000000
#> [3,] 0.0000000 0.0000000 0.0488351 0.0000000
#> [4,] 0.0000000 0.0000000 0.0000000 0.0732373
#> 
#> $`3`
#>          [,1]      [,2]      [,3]      [,4]
#> [1,] 0.042979 0.0000000 0.0000000 0.0000000
#> [2,] 0.000000 0.0677383 0.0000000 0.0000000
#> [3,] 0.000000 0.0000000 0.0434428 0.0000000
#> [4,] 0.000000 0.0000000 0.0000000 0.0843981
#> 
#> $`4`
#>           [,1]      [,2]     [,3]      [,4]
#> [1,] 0.0496743 0.0000000 0.000000 0.0000000
#> [2,] 0.0000000 0.0569424 0.000000 0.0000000
#> [3,] 0.0000000 0.0000000 0.043422 0.0000000
#> [4,] 0.0000000 0.0000000 0.000000 0.0829493
#> 
fit$inv_metric(matrix=FALSE)
#> $`1`
#> [1] 0.0463354 0.0601104 0.0507043 0.0636408
#> 
#> $`2`
#> [1] 0.0481084 0.0561195 0.0488351 0.0732373
#> 
#> $`3`
#> [1] 0.0429790 0.0677383 0.0434428 0.0843981
#> 
#> $`4`
#> [1] 0.0496743 0.0569424 0.0434220 0.0829493
#> 

fit <- cmdstanr_example("logistic", metric = "dense_e")
fit$inv_metric()
#> $`1`
#>             [,1]        [,2]        [,3]        [,4]
#> [1,]  0.04207890 -0.00411326  0.00140018  0.00316808
#> [2,] -0.00411326  0.07005260 -0.00358534 -0.00721287
#> [3,]  0.00140018 -0.00358534  0.04989460 -0.00990938
#> [4,]  0.00316808 -0.00721287 -0.00990938  0.07372540
#> 
#> $`2`
#>             [,1]         [,2]         [,3]        [,4]
#> [1,]  0.04375550 -0.005185460  0.003637290  0.00263089
#> [2,] -0.00518546  0.062107700 -0.000240788 -0.01515610
#> [3,]  0.00363729 -0.000240788  0.048456800 -0.00481198
#> [4,]  0.00263089 -0.015156100 -0.004811980  0.07412660
#> 
#> $`3`
#>              [,1]        [,2]         [,3]        [,4]
#> [1,]  4.03214e-02 -0.00510548 -1.79004e-05  0.00148148
#> [2,] -5.10548e-03  0.06109750 -7.34705e-03  0.00242718
#> [3,] -1.79004e-05 -0.00734705  4.52891e-02 -0.01039730
#> [4,]  1.48148e-03  0.00242718 -1.03973e-02  0.06407940
#> 
#> $`4`
#>              [,1]         [,2]        [,3]        [,4]
#> [1,]  5.18858e-02 -6.89848e-05  0.00192026  0.00576423
#> [2,] -6.89848e-05  6.60642e-02 -0.00448511 -0.00922828
#> [3,]  1.92026e-03 -4.48511e-03  0.04440530 -0.01030510
#> [4,]  5.76423e-03 -9.22828e-03 -0.01030510  0.07884090
#> 
# }