Save fitted model object to a file

This method is a wrapper around base::saveRDS() that ensures that all posterior draws and diagnostics are saved when saving a fitted model object. Because the contents of the CmdStan output CSV files are only read into R lazily (i.e., as needed), the $save_object() method is the safest way to guarantee that everything has been read in before saving.

Usage

$save_object(file, ...)

Arguments

file: (string) Path where the file should be saved.

...: Other arguments to pass to base::saveRDS() besides object and file.

See also

Examples

# \dontrun{
fit <- cmdstanr_example("logistic")

#> Model executable is up to date!

temp_rds_file <- tempfile(fileext = ".RDS")
fit$save_object(file = temp_rds_file)
rm(fit)

fit <- readRDS(temp_rds_file)
fit$summary()

#> # A tibble: 5 x 10
#>   variable    mean  median    sd   mad       q5      q95  rhat ess_bulk ess_tail
#>   <chr>      <dbl>   <dbl> <dbl> <dbl>    <dbl>    <dbl> <dbl>    <dbl>    <dbl>
#> 1 lp__     -66.0   -65.6   1.45  1.22  -68.8    -64.3    1.00     2005.    2605.
#> 2 alpha      0.374   0.378 0.218 0.222   0.0190   0.727  1.00     4845.    3167.
#> 3 beta[1]   -0.664  -0.656 0.251 0.249  -1.08    -0.269  0.999    4132.    2792.
#> 4 beta[2]   -0.270  -0.267 0.225 0.222  -0.643    0.0991 1.00     4210.    3019.
#> 5 beta[3]    0.674   0.673 0.266 0.255   0.238    1.12   1.00     4088.    2823.

# }