Save output and data files

All fitted model objects have methods for saving (moving to a specified location) the files created by CmdStanR to hold CmdStan output csv files and input data files. These methods move the files from their current location (possibly the temporary directory) to a user-specified location. The paths stored in the fitted model object will also be updated to point to the new file locations.

The versions without the save_ prefix (e.g., $output_files()) return the current file paths without moving any files.

Usage

$save_output_files(dir = ".", basename = NULL, timestamp = TRUE, random = TRUE)
$save_latent_dynamics_files(dir = ".", basename = NULL, timestamp = TRUE, random = TRUE)
$save_data_file(dir = ".", basename = NULL, timestamp = TRUE, random = TRUE)

$output_files()
$latent_dynamics_files()
$data_file()

Arguments

dir: (string) Path to directory where the files should be saved.

basename: (string) Base filename to use. See Details.

timestamp: (logical) Should a timestamp be added to the file name(s)? Defaults to TRUE. See Details.

random: (logical) Should random alphanumeric characters be added to the end of the file name(s)? Defaults to TRUE. See Details.

Details

For $save_output_files() the files moved to dir will have names of the form basename-timestamp-id-random, where

basename is the user's provided basename argument;

timestamp is of the form format(Sys.time(), "%Y%m%d%H%M");

id is the MCMC chain id (or 1 for non MCMC);

random contains six random alphanumeric characters;

For $save_latent_dynamics_files() everything is the same as for $save_output_files() except "-diagnostic-" is included in the new file name after basename.

For $save_data_file() no id is included in the file name because even with multiple MCMC chains the data file is the same.

Value

The $save_* methods print a message with the new file paths and (invisibly) return a character vector of the new paths (or NA for any that couldn't be copied). They also have the side effect of setting the internal paths in the fitted model object to the new paths.

The methods without the save_ prefix return character vectors of file paths without moving any files.

See also