Compute a summary table of MCMC estimates and diagnostics

The $summary() method runs summarise_draws() from the posterior package. For MCMC only post-warmup draws are included in the summary.

The $print() method prints the same summary but removes the extra formatting used for printing tibbles.

Usage

$summary(variables = NULL, ...)
$print(variables = NULL, ..., digits = 2, max_rows = 10)

Arguments

variables: (character vector) The variables to include.

...: Optional arguments to pass to posterior::summarise_draws().

digits: (integer) For print only, the number of digits to use for rounding.

max_rows: (integer) For print only, the maximum number of rows to print.

Value

The $summary() method returns the tibble created by posterior::summarise_draws().

The $print() method returns the fitted model object itself (invisibly), which is the standard behavior for print methods in R.

See also

Examples

# \dontrun{
fit <- cmdstanr_example("logistic")

#> Model executable is up to date!

fit$summary()

#> # A tibble: 5 x 10
#>   variable    mean  median    sd   mad       q5      q95  rhat ess_bulk ess_tail
#>   <chr>      <dbl>   <dbl> <dbl> <dbl>    <dbl>    <dbl> <dbl>    <dbl>    <dbl>
#> 1 lp__     -66.0   -65.7   1.41  1.26  -68.7    -64.3     1.00    2084.    2857.
#> 2 alpha      0.378   0.380 0.220 0.221   0.0169   0.739   1.00    4594.    3471.
#> 3 beta[1]   -0.664  -0.657 0.248 0.252  -1.08    -0.271   1.00    4286.    3269.
#> 4 beta[2]   -0.270  -0.271 0.227 0.222  -0.644    0.0887  1.00    3983.    3171.
#> 5 beta[3]    0.679   0.674 0.268 0.264   0.249    1.14    1.00    4082.    2901.

fit$print()

#>  variable   mean median   sd  mad     q5    q95 rhat ess_bulk ess_tail
#>   lp__    -65.98 -65.65 1.41 1.26 -68.71 -64.29 1.00     2083     2857
#>   alpha     0.38   0.38 0.22 0.22   0.02   0.74 1.00     4594     3470
#>   beta[1]  -0.66  -0.66 0.25 0.25  -1.08  -0.27 1.00     4285     3269
#>   beta[2]  -0.27  -0.27 0.23 0.22  -0.64   0.09 1.00     3982     3170
#>   beta[3]   0.68   0.67 0.27 0.26   0.25   1.14 1.00     4082     2900

fit$print(max_rows = 2) # same as print(fit, max_rows = 2)

#>  variable   mean median   sd  mad     q5    q95 rhat ess_bulk ess_tail
#>     lp__  -65.98 -65.65 1.41 1.26 -68.71 -64.29 1.00     2083     2857
#>     alpha   0.38   0.38 0.22 0.22   0.02   0.74 1.00     4594     3470
#> 
#>  # showing 2 of 5 rows (change via 'max_rows' argument)

# include only certain variables
fit$summary("beta")

#> # A tibble: 3 x 10
#>   variable   mean median    sd   mad     q5     q95  rhat ess_bulk ess_tail
#>   <chr>     <dbl>  <dbl> <dbl> <dbl>  <dbl>   <dbl> <dbl>    <dbl>    <dbl>
#> 1 beta[1]  -0.664 -0.657 0.248 0.252 -1.08  -0.271   1.00    4286.    3269.
#> 2 beta[2]  -0.270 -0.271 0.227 0.222 -0.644  0.0887  1.00    3983.    3171.
#> 3 beta[3]   0.679  0.674 0.268 0.264  0.249  1.14    1.00    4082.    2901.

fit$print(c("alpha", "beta[2]"))

#>  variable  mean median   sd  mad    q5  q95 rhat ess_bulk ess_tail
#>   alpha    0.38   0.38 0.22 0.22  0.02 0.74 1.00     4594     3470
#>   beta[2] -0.27  -0.27 0.23 0.22 -0.64 0.09 1.00     3982     3170

# include all variables but only certain summaries
fit$summary(NULL, c("mean", "sd"))

#> # A tibble: 5 x 3
#>   variable    mean    sd
#>   <chr>      <dbl> <dbl>
#> 1 lp__     -66.0   1.41 
#> 2 alpha      0.378 0.220
#> 3 beta[1]   -0.664 0.248
#> 4 beta[2]   -0.270 0.227
#> 5 beta[3]    0.679 0.268

# can use functions created from formulas
# for example, calculate Pr(beta > 0)
fit$summary("beta", prob_gt_0 = ~ mean(. > 0))

#> # A tibble: 3 x 2
#>   variable prob_gt_0
#>   <chr>        <dbl>
#> 1 beta[1]    0.00425
#> 2 beta[2]    0.111  
#> 3 beta[3]    0.999  

# }