Compute a summary table of MCMC estimates and diagnostics

The $summary() method runs summarise_draws() from the posterior package. For MCMC only post-warmup draws are included in the summary.

The $print() method prints the same summary but removes the extra formatting used for printing tibbles.

Usage

$summary(variables = NULL, ...)
$print(variables = NULL, ..., digits = 2, max_rows = 10)

Arguments

variables: (character vector) The variables to include.

...: Optional arguments to pass to posterior::summarise_draws().

digits: (integer) For print only, the number of digits to use for rounding.

max_rows: (integer) For print only, the maximum number of rows to print.

Value

The $summary() method returns the tibble created by posterior::summarise_draws().

The $print() method returns the fitted model object itself (invisibly), which is the standard behavior for print methods in R.

See also

Examples

# \dontrun{
fit <- cmdstanr_example("logistic")

#> Model executable is up to date!

fit$summary()

#> # A tibble: 5 x 10
#>   variable    mean  median    sd   mad       q5      q95  rhat ess_bulk ess_tail
#>   <chr>      <dbl>   <dbl> <dbl> <dbl>    <dbl>    <dbl> <dbl>    <dbl>    <dbl>
#> 1 lp__     -66.0   -65.6   1.43  1.25  -68.8    -64.3     1.00    1971.    2555.
#> 2 alpha      0.382   0.387 0.219 0.225   0.0173   0.737   1.00    4138.    2869.
#> 3 beta[1]   -0.676  -0.671 0.251 0.252  -1.09    -0.263   1.00    3948.    2758.
#> 4 beta[2]   -0.272  -0.267 0.223 0.217  -0.648    0.0906  1.00    4275.    2786.
#> 5 beta[3]    0.681   0.675 0.273 0.274   0.241    1.13    1.00    4037.    2735.

fit$print()

#>  variable   mean median   sd  mad     q5    q95 rhat ess_bulk ess_tail
#>   lp__    -65.98 -65.65 1.43 1.25 -68.76 -64.30 1.00     1971     2554
#>   alpha     0.38   0.39 0.22 0.22   0.02   0.74 1.00     4137     2868
#>   beta[1]  -0.68  -0.67 0.25 0.25  -1.09  -0.26 1.00     3947     2757
#>   beta[2]  -0.27  -0.27 0.22 0.22  -0.65   0.09 1.00     4275     2785
#>   beta[3]   0.68   0.68 0.27 0.27   0.24   1.13 1.00     4036     2735

fit$print(max_rows = 2) # same as print(fit, max_rows = 2)

#>  variable   mean median   sd  mad     q5    q95 rhat ess_bulk ess_tail
#>     lp__  -65.98 -65.65 1.43 1.25 -68.76 -64.30 1.00     1971     2554
#>     alpha   0.38   0.39 0.22 0.22   0.02   0.74 1.00     4137     2868
#> 
#>  # showing 2 of 5 rows (change via 'max_rows' argument)

# include only certain variables
fit$summary("beta")

#> # A tibble: 3 x 10
#>   variable   mean median    sd   mad     q5     q95  rhat ess_bulk ess_tail
#>   <chr>     <dbl>  <dbl> <dbl> <dbl>  <dbl>   <dbl> <dbl>    <dbl>    <dbl>
#> 1 beta[1]  -0.676 -0.671 0.251 0.252 -1.09  -0.263   1.00    3948.    2758.
#> 2 beta[2]  -0.272 -0.267 0.223 0.217 -0.648  0.0906  1.00    4275.    2786.
#> 3 beta[3]   0.681  0.675 0.273 0.274  0.241  1.13    1.00    4037.    2735.

fit$print(c("alpha", "beta[2]"))

#>  variable  mean median   sd  mad    q5  q95 rhat ess_bulk ess_tail
#>   alpha    0.38   0.39 0.22 0.22  0.02 0.74 1.00     4137     2868
#>   beta[2] -0.27  -0.27 0.22 0.22 -0.65 0.09 1.00     4275     2785

# include all variables but only certain summaries
fit$summary(NULL, c("mean", "sd"))

#> # A tibble: 5 x 3
#>   variable    mean    sd
#>   <chr>      <dbl> <dbl>
#> 1 lp__     -66.0   1.43 
#> 2 alpha      0.382 0.219
#> 3 beta[1]   -0.676 0.251
#> 4 beta[2]   -0.272 0.223
#> 5 beta[3]    0.681 0.273

# can use functions created from formulas
# for example, calculate Pr(beta > 0)
fit$summary("beta", prob_gt_0 = ~ mean(. > 0))

#> # A tibble: 3 x 2
#>   variable prob_gt_0
#>   <chr>        <dbl>
#> 1 beta[1]    0.00225
#> 2 beta[2]    0.103  
#> 3 beta[3]    0.997  

# }