Install CmdStan or clean and rebuild an existing installation

The install_cmdstan() function attempts to download and install the latest release of CmdStan or a development version from a repository. Currently the necessary C++ tool chain is assumed to be available (see Appendix B of the CmdStan guide), but in the future CmdStanR may help install the requirements.

The rebuild_cmdstan() function cleans and rebuilds the cmdstan installation. Use this function in case of any issues when compiling models.

The cmdstan_make_local() function is used to read/write makefile flags and variables from/to the make/local file of a CmdStan installation. Writing to the make/local file can be used to permanently add makefile flags/variables to an installation. For example adding specific compiler switches, changing the C++ compiler, etc. A change to the make/local file should typically be followed by calling rebuild_cmdstan().

install_cmdstan(
  dir = NULL,
  cores = getOption("mc.cores", 2),
  quiet = FALSE,
  overwrite = FALSE,
  timeout = 1200,
  release_url = NULL,
  flags = list()
)

rebuild_cmdstan(
  dir = cmdstan_path(),
  cores = getOption("mc.cores", 2),
  quiet = FALSE,
  timeout = 600
)

cmdstan_make_local(dir = cmdstan_path(), flags = NULL, append = TRUE)

Arguments

dir
Path to the directory in which to install CmdStan. The default is to install it in a directory called .cmdstanr within the user's home directory (i.e, file.path(Sys.getenv("HOME"), ".cmdstanr")).

cores
The number of CPU cores to use to parallelize building CmdStan and speed up installation. If cores is not specified then the default is to look for the option "mc.cores", which can be set for an entire R session by options(mc.cores=value). If the "mc.cores" option has not been set then the default is 2.

quiet
Should the verbose output from the system processes be suppressed when building the CmdStan binaries? The default is FALSE.

overwrite
When an existing installation is found in dir, should CmdStan still be downloaded and reinstalled? The default is FALSE, in which case an informative error is thrown instead of overwriting the user's installation.

timeout
Timeout (in seconds) for the CmdStan build stage of the installation process.

release_url
Specifies the URL to a specific Cmdstan release to be installed. By default set to NULL, which downloads the latest stable release from GitHub.

flags
A list specifying any makefile flags/variables to be written to the make/local file. For example, list("CXX" = "clang++") will force the use of clang for compilation.

append
For cmdstan_make_local(), should the listed makefile flags be appended to the end of the existing make/local file? The default is TRUE. If FALSE the file is overwritten.

Value

For cmdstan_make_local(), if flags=NULL then the existing contents of make/local are returned without writing anything, otherwise the updated contents are returned.

Examples

# install_cmdstan(cores = 4)

flags <- list(
  "CXX" = "clang++",
  "CXXFLAGS+= -march-native",
  PRECOMPILED_HEADERS = TRUE
)
# cmdstan_make_local(flags = flags)
# rebuild_cmdstan()