Read CmdStan CSV files into R

read_cmdstan_csv() is used internally by CmdStanR to read CmdStan's output CSV files into R. It can also be used by CmdStan users as a more flexible and efficient alternative to rstan::read_stan_csv().

read_cmdstan_csv(files, variables = NULL, sampler_diagnostics = NULL)

Arguments

files
A character vector of paths to the CSV files to read.

variables
Optionally, a character vector naming the variables (parameters, transformed parameters, and generated quantities) to read in.

If NULL (the default) then all variables are included.

If an empty string (variables="") then none are included.

For non-scalar variables all elements or specific elements can be selected:

variables = "theta" selects all elements of theta;

variables = c("theta[1]", "theta[3]") selects only the 1st and 3rd elements.

sampler_diagnostics
Works the same way as variables but for sampler diagnostic variables (e.g., "treedepth__", "accept_stat__", etc.). Ignored if the model was not fit using MCMC.

Value

A named list with the following components:

metadata: A list of the meta information from the run that produced the CSV file(s). See Examples below.

The other components in the returned list depend on the method that produced the CSV file(s).

For sampling the returned list also includes the following components:

inv_metric: A list (one element per chain) of inverse mass matrices or their diagonals, depending on the type of metric used.

step_size: A list (one element per chain) of the step sizes used.

warmup_draws: If save_warmup was TRUE when fitting the model then a draws_array of warmup draws.

post_warmup_draws: A draws_array of post-warmup draws.

warmup_sampler_diagnostics: If save_warmup was TRUE when fitting the model then a draws_array of warmup draws of the sampler diagnostic variables.

post_warmup_sampler_diagnostics: A draws_array of post-warmup draws of the sampler diagnostic variables.

For optimization the returned list also includes the following components:

point_estimates: Point estimates for the model parameters.

For variational inference the returned list also includes the following components:

draws: A draws_matrix of draws from the approximate posterior distribution.

For standalone generated quantities the returned list also includes the following components:

generated_quantities: A draws_array of the generated quantities.

Examples

# \dontrun{
stan_program <- write_stan_tempfile(
  "parameters {
    real alpha_scalar;
    vector[2] theta_vector;
    matrix[2,2] tau_matrix;
  }
  model {
    alpha_scalar ~ std_normal();
    theta_vector ~ std_normal();
    to_vector(tau_matrix) ~ std_normal();
  }"
)

# only using capture.output to avoid too much printed output in example
out <- utils::capture.output(
  mod <- cmdstan_model(stan_program),
  fit <- mod$sample(save_warmup=TRUE)
)

#> Compiling Stan program...

# Read in everything
x <- read_cmdstan_csv(fit$output_files())
str(x)

#> List of 7
#>  $ metadata                       :List of 30
#>   ..$ stan_version_major  : num 2
#>   ..$ stan_version_minor  : num 23
#>   ..$ stan_version_patch  : num 0
#>   ..$ method              : chr "sample"
#>   ..$ save_warmup         : num 1
#>   ..$ thin                : num 1
#>   ..$ gamma               : num 0.05
#>   ..$ kappa               : num 0.75
#>   ..$ t0                  : num 10
#>   ..$ init_buffer         : num 75
#>   ..$ term_buffer         : num 50
#>   ..$ window              : num 25
#>   ..$ algorithm           : chr "hmc"
#>   ..$ engine              : chr "nuts"
#>   ..$ metric              : chr "diag_e"
#>   ..$ stepsize_jitter     : num 0
#>   ..$ id                  : num [1:4] 1 2 3 4
#>   ..$ init                : num 2
#>   ..$ seed                : num [1:4] 3.00e+08 9.19e+08 1.79e+09 7.26e+08
#>   ..$ refresh             : num 100
#>   ..$ model_params        : chr [1:8] "lp__" "alpha_scalar" "theta_vector[1]" "theta_vector[2]" ...
#>   ..$ sampler_diagnostics : chr [1:6] "accept_stat__" "stepsize__" "treedepth__" "n_leapfrog__" ...
#>   ..$ model_name          : chr "file6aac4a66980_model"
#>   ..$ adapt_engaged       : num 1
#>   ..$ adapt_delta         : num 0.8
#>   ..$ max_treedepth       : num 10
#>   ..$ step_size           : num 1
#>   ..$ step_size_adaptation: num [1:4] 0.812 0.724 0.736 0.733
#>   ..$ iter_warmup         : num 1000
#>   ..$ iter_sampling       : num 1000
#>  $ inv_metric                     :List of 4
#>   ..$ : num [1:7] 0.985 1.03 0.966 0.8 0.999 ...
#>   ..$ : num [1:7] 0.96 1.033 0.993 0.8 0.896 ...
#>   ..$ : num [1:7] 0.963 0.949 1.024 0.961 1.004 ...
#>   ..$ : num [1:7] 0.934 0.881 0.947 1.026 0.941 ...
#>  $ step_size                      :List of 4
#>   ..$ : num 0.812
#>   ..$ : num 0.724
#>   ..$ : num 0.736
#>   ..$ : num 0.733
#>  $ warmup_draws                   : 'draws_array' num [1:1000, 1:4, 1:8] -5.62 -5.62 -2.48 -2.51 -2 ...
#>   ..- attr(*, "dimnames")=List of 3
#>   .. ..$ iteration: chr [1:1000] "1" "2" "3" "4" ...
#>   .. ..$ chain    : chr [1:4] "1" "2" "3" "4"
#>   .. ..$ variable : chr [1:8] "lp__" "alpha_scalar" "theta_vector[1]" "theta_vector[2]" ...
#>  $ post_warmup_draws              : 'draws_array' num [1:1000, 1:4, 1:8] -4.92 -3.24 -3.73 -3.38 -2.57 ...
#>   ..- attr(*, "dimnames")=List of 3
#>   .. ..$ iteration: chr [1:1000] "1" "2" "3" "4" ...
#>   .. ..$ chain    : chr [1:4] "1" "2" "3" "4"
#>   .. ..$ variable : chr [1:8] "lp__" "alpha_scalar" "theta_vector[1]" "theta_vector[2]" ...
#>  $ warmup_sampler_diagnostics     : 'draws_array' num [1:1000, 1:4, 1:6] 1.19e-08 6.53e-22 1.00 9.82e-01 9.85e-01 ...
#>   ..- attr(*, "dimnames")=List of 3
#>   .. ..$ iteration: chr [1:1000] "1" "2" "3" "4" ...
#>   .. ..$ chain    : chr [1:4] "1" "2" "3" "4"
#>   .. ..$ variable : chr [1:6] "accept_stat__" "stepsize__" "treedepth__" "n_leapfrog__" ...
#>  $ post_warmup_sampler_diagnostics: 'draws_array' num [1:1000, 1:4, 1:6] 0.816 0.996 0.825 1 0.801 ...
#>   ..- attr(*, "dimnames")=List of 3
#>   .. ..$ iteration: chr [1:1000] "1" "2" "3" "4" ...
#>   .. ..$ chain    : chr [1:4] "1" "2" "3" "4"
#>   .. ..$ variable : chr [1:6] "accept_stat__" "stepsize__" "treedepth__" "n_leapfrog__" ...

# Don't read in any of the sampler diagnostic variables
x <- read_cmdstan_csv(fit$output_files(), sampler_diagnostics = "")

# Don't read in any of the parameters or generated quantities
x <- read_cmdstan_csv(fit$output_files(), variables = "")

# Read in only specific parameters and sampler diagnostics
x <- read_cmdstan_csv(
  fit$output_files(),
  variables = c("alpha_scalar", "theta_vector[2]"),
  sampler_diagnostics = c("n_leapfrog__", "accept_stat__")
)

# For non-scalar parameters all elements can be selected or only some elements,
# e.g. all of "theta_vector" but only one element of "tau_matrix"
x <- read_cmdstan_csv(
  fit$output_files(),
  variables = c("theta_vector", "tau_matrix[2,1]")
)
# }